% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agent.r
\docType{class}
\name{Agent}
\alias{Agent}
\title{Agent Class}
\format{R6 class}
\usage{
# Agent$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for Agent objects.
}
\section{Fields}{

\describe{
\item{\code{agentText}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Agent object.

}
\item{\code{$fromList(AgentList)}}{

  Create Agent object from list.

}

\item{\code{$toList()}}{

  Get list representation of Agent.

}
\item{\code{fromJSONString(AgentJson)}}{

  Create Agent object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Agent.

}
}
}

\keyword{datasets}
