% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiClient.r
\name{ApiClient}
\alias{ApiClient}
\title{ApiClient class}
\description{
Generic class for API client calls, all nbaR client classes
derive from this class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{basePath}}{specifies the base URL of the API, defaults to
http://api.biodiversitydata.nl/v2}

\item{\code{userAgent}}{Set the user agent of the request, defaults to
nbaR/0.1.0}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApiClient-new}{\code{ApiClient$new()}}
\item \href{#method-ApiClient-callApi}{\code{ApiClient$callApi()}}
\item \href{#method-ApiClient-handleError}{\code{ApiClient$handleError()}}
\item \href{#method-ApiClient-ping}{\code{ApiClient$ping()}}
\item \href{#method-ApiClient-print}{\code{ApiClient$print()}}
\item \href{#method-ApiClient-clone}{\code{ApiClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApiClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiClient$new(basePath, userAgent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiClient-callApi"></a>}}
\if{latex}{\out{\hypertarget{method-ApiClient-callApi}{}}}
\subsection{Method \code{callApi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiClient$callApi(url, method, queryParams, headerParams, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiClient-handleError"></a>}}
\if{latex}{\out{\hypertarget{method-ApiClient-handleError}{}}}
\subsection{Method \code{handleError()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiClient$handleError(response)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiClient-ping"></a>}}
\if{latex}{\out{\hypertarget{method-ApiClient-ping}{}}}
\subsection{Method \code{ping()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiClient$ping()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiClient-print"></a>}}
\if{latex}{\out{\hypertarget{method-ApiClient-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiClient$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApiClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
