% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssociatedTaxon.r
\docType{data}
\name{AssociatedTaxon}
\alias{AssociatedTaxon}
\title{AssociatedTaxon Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AssociatedTaxon
}
\description{
AssociatedTaxon Class
}
\section{Fields}{

\describe{
\item{\code{name}}{character}

\item{\code{relationType}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor AssociatedTaxon object.

}
\item{\code{$fromList(AssociatedTaxonList)}}{

  Create AssociatedTaxon object from list.

}

\item{\code{$toList()}}{

  Get list representation of AssociatedTaxon.

}
\item{\code{fromJSONString(AssociatedTaxonJson)}}{

  Create AssociatedTaxon object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of AssociatedTaxon.

}
}
}

\keyword{datasets}
