% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioStratigraphy.r
\docType{data}
\name{BioStratigraphy}
\alias{BioStratigraphy}
\title{BioStratigraphy Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BioStratigraphy
}
\description{
BioStratigraphy Class
}
\section{Fields}{

\describe{
\item{\code{oldBioStratType}}{}
}}

\keyword{datasets}
