% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNKCJE.R
\name{checkActual}
\alias{checkActual}
\title{Check if game tables in database are up to date}
\usage{
checkActual(table, dbname, date = Sys.Date() - 1, schema = "rd",
  pghost = "localhost", pgport = 5432, pguser = "postgres",
  pgpassword = "postgres", calendartbl = "calendar", gamecol = "GAME_ID")
}
\description{
Check if game tables in database are up to date
}
