% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.r
\name{chronos_calib}
\alias{chronos_calib}
\title{Make calibration table}
\usage{
chronos_calib(specimens, tree, level = "genus")
}
\arguments{
\item{specimens}{list of objects of class Specimen, must have chronostratigraphy information}

\item{tree}{object of class 'phylo'}

\item{level}{character giving taxonomic level}
}
\description{
Make calibration table compatible with ape's 'chronos'
}
\details{
Given a list of specimen objects, and a phylogenetic tree,
makes a calibration table that is compatible with the 'chronos'
function from the 'ape' package. This can be done for various taxonomic levels
}
\examples{
# get specimen with chronostratigraphic data
sc <- SpecimenClient$new()
sp <- sc$find_by_unit_id("RGM.156532")$content
# load tree
library('ape')
path <- system.file('extdata', 'shark_tree.nex', package='nbaR')
tree <- read.nexus(path)
# make calibration table
chronos_calib(sp, tree, "genus")
}
\seealso{
Other utils: \code{\link{geo_age}}
}
