% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{colr_tbl}
\alias{colr_tbl}
\title{Coloured table}
\usage{
colr_tbl(DF, COL)
}
\arguments{
\item{DF}{The data frame that contains the data}

\item{COL}{the column containing the categories}
}
\description{
A basic table with a purple heading bar and coloured blocks around the
categories
This data should be in the same fomat as for the plot functions
}
\examples{
#test <- thr_tbl(bird_data, Protection_level)

}
