% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.r
\docType{class}
\name{Filter}
\alias{Filter}
\title{Filter Class}
\format{R6 class}
\usage{
# Filter$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for Filter objects.
}
\section{Fields}{

\describe{
\item{\code{acceptRegexp}}{character}

\item{\code{rejectRegexp}}{character}

\item{\code{acceptValues}}{list(character)}

\item{\code{rejectValues}}{list(character)}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Filter object.

}
\item{\code{$fromList(FilterList)}}{

  Create Filter object from list.

}

\item{\code{$toList()}}{

  Get list representation of Filter.

}
\item{\code{fromJSONString(FilterJson)}}{

  Create Filter object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Filter.

}
}
}

\keyword{datasets}
