% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatheringEvent.r
\docType{data}
\name{GatheringEvent}
\alias{GatheringEvent}
\title{GatheringEvent Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GatheringEvent
}
\description{
GatheringEvent Class
}
\section{Fields}{

\describe{
\item{\code{projectTitle}}{}

\item{\code{worldRegion}}{}

\item{\code{continent}}{}

\item{\code{country}}{}

\item{\code{iso3166Code}}{}

\item{\code{provinceState}}{}

\item{\code{island}}{}

\item{\code{locality}}{}

\item{\code{city}}{}

\item{\code{sublocality}}{}

\item{\code{localityText}}{}

\item{\code{dateTimeBegin}}{}

\item{\code{dateTimeEnd}}{}

\item{\code{method}}{}

\item{\code{altitude}}{}

\item{\code{altitudeUnifOfMeasurement}}{}

\item{\code{biotopeText}}{}

\item{\code{depth}}{}

\item{\code{depthUnitOfMeasurement}}{}

\item{\code{gatheringPersons}}{}

\item{\code{gatheringOrganizations}}{}

\item{\code{siteCoordinates}}{}

\item{\code{namedAreas}}{}

\item{\code{associatedTaxa}}{}

\item{\code{chronoStratigraphy}}{}

\item{\code{lithoStratigraphy}}{}

\item{\code{bioStratigraphic}}{}
}}

\keyword{datasets}
