% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatheringSiteCoordinates.r
\docType{class}
\name{GatheringSiteCoordinates}
\alias{GatheringSiteCoordinates}
\title{GatheringSiteCoordinates Class}
\format{R6 class}
\usage{
# GatheringSiteCoordinates$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for GatheringSiteCoordinates objects.
}
\section{Fields}{

\describe{
\item{\code{longitudeDecimal}}{numeric}

\item{\code{latitudeDecimal}}{numeric}

\item{\code{gridCellSystem}}{character}

\item{\code{gridLatitudeDecimal}}{numeric}

\item{\code{gridLongitudeDecimal}}{numeric}

\item{\code{gridCellCode}}{character}

\item{\code{gridQualifier}}{character}

\item{\code{geoShape}}{Point}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor GatheringSiteCoordinates object.

}
\item{\code{$fromList(GatheringSiteCoordinatesList)}}{

  Create GatheringSiteCoordinates object from list.

}

\item{\code{$toList()}}{

  Get list representation of GatheringSiteCoordinates.

}
\item{\code{fromJSONString(GatheringSiteCoordinatesJson)}}{

  Create GatheringSiteCoordinates object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of GatheringSiteCoordinates.

}
}
}

\keyword{datasets}
