% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.r
\name{geo_age}
\alias{geo_age}
\title{Geological Age}
\usage{
geo_age(geo_time)
}
\arguments{
\item{geo_time}{character giving a Geological timespan}
}
\value{
list
}
\description{
Get early and late age for geological time span
}
\details{
Uses the API from earthlifeconsortium.org to retreive
upper and lower bound for a geological age range (e.g. 'miocene').
Unit can be can be Eon, Era, System/Period, Series/Epoch.
Returns a list with items early_age and late_age. Gives a warning
if age range no found or if the API call times out.
}
\examples{
geo_age('lower miocene')
}
\seealso{
Other utils: \code{\link{chronos_calib}}
}
\concept{utils}
