% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoClient-wrapper.r
\name{geo_count_distinct_values_per_group}
\alias{geo_count_distinct_values_per_group}
\title{Count the distinct number of field values that exist per the given field to group by}
\usage{
geo_count_distinct_values_per_group(group = NULL, field = NULL,
  returnType = "data.frame", ...)
}
\arguments{
\item{group}{name of field in the geo area object to group by, type:}

\item{field}{name of field in the geo area object, type:}

\item{returnType}{Either \code{list} or \code{data.frame} (default)}

\item{...}{additional parameters passed to count_distinct_values_per_group from class nbaR.GeoClient}
}
\value{
scalar
}
\description{
This is a wrapper for the method \code{ count_distinct_values_per_group }
from class \code{ GeoClient}.
}
\details{
See also endpoint /getDistinctValues. See also getDistinctValuesPerGroup
}
\seealso{
Other nbaR.GeoClient-wrappers: \code{\link{geo_count_distinct_values}},
  \code{\link{geo_count}}, \code{\link{geo_find_by_ids}},
  \code{\link{geo_find}},
  \code{\link{geo_get_distinct_values_per_group}},
  \code{\link{geo_get_distinct_values}},
  \code{\link{geo_get_field_info}},
  \code{\link{geo_get_geo_json_for_locality}},
  \code{\link{geo_get_paths}},
  \code{\link{geo_get_settings}},
  \code{\link{geo_get_setting}},
  \code{\link{geo_is_operator_allowed}},
  \code{\link{geo_query}}
}
\concept{nbaR.GeoClient-wrappers}
