% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoClient-wrapper.r
\name{geo_find}
\alias{geo_find}
\title{Find a GEO area by id}
\usage{
geo_find(id = NULL, ...)
}
\arguments{
\item{id}{id of geo area, type:}

\item{...}{additional parameters passed to find from class class nbaR.GeoClient}
}
\value{
list
}
\description{
This is a wrapper function for find
from class nbaR.GeoClient.
}
\details{
Returns a GEO object containing a GEO json polygon
}
\seealso{
Other nbaR.GeoClient-wrappers: \code{\link{geo_count_distinct_values_per_group}},
  \code{\link{geo_count_distinct_values}},
  \code{\link{geo_count}}, \code{\link{geo_find_by_ids}},
  \code{\link{geo_get_distinct_values_per_group}},
  \code{\link{geo_get_distinct_values}},
  \code{\link{geo_get_field_info}},
  \code{\link{geo_get_geo_json_for_locality}},
  \code{\link{geo_get_paths}},
  \code{\link{geo_get_settings}},
  \code{\link{geo_get_setting}},
  \code{\link{geo_is_operator_allowed}},
  \code{\link{geo_query}}
}
\concept{nbaR.GeoClient-wrappers}
