% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoClient-wrapper.r
\name{geo_get_distinct_values}
\alias{geo_get_distinct_values}
\title{Get all different values that exist for a field}
\usage{
geo_get_distinct_values(field = NULL, returnType = "data.frame", ...)
}
\arguments{
\item{field}{name of field in geo area object, type:}

\item{returnType}{Either \code{list} or \code{data.frame} (default)}

\item{...}{additional parameters passed to get_distinct_values from class nbaR.GeoClient}
}
\value{
scalar
}
\description{
This is a wrapper for the method \code{ get_distinct_values }
from class \code{ GeoClient}.
}
\details{
A list of all fields for geo area documents can be retrieved with /metadata/getFieldInfo
}
\seealso{
Other nbaR.GeoClient-wrappers: 
\code{\link{geo_count_distinct_values_per_group}()},
\code{\link{geo_count_distinct_values}()},
\code{\link{geo_count}()},
\code{\link{geo_find_by_ids}()},
\code{\link{geo_find}()},
\code{\link{geo_get_distinct_values_per_group}()},
\code{\link{geo_get_field_info}()},
\code{\link{geo_get_geo_json_for_locality}()},
\code{\link{geo_get_paths}()},
\code{\link{geo_get_settings}()},
\code{\link{geo_get_setting}()},
\code{\link{geo_is_operator_allowed}()},
\code{\link{geo_query}()}
}
\concept{nbaR.GeoClient-wrappers}
