% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoClient-wrapper.r
\name{geo_get_distinct_values_per_group}
\alias{geo_get_distinct_values_per_group}
\title{Get all distinct values (and their document count) for the field given divided per distinct value of the field to group by}
\usage{
geo_get_distinct_values_per_group(group = NULL, field = NULL, ...)
}
\arguments{
\item{group}{name of field in the geo area object to group by, type:}

\item{field}{name of field in the geo area object, type:}

\item{...}{additional parameters passed to get_distinct_values_per_group from class class nbaR.GeoClient}
}
\value{
scalar
}
\description{
This is a wrapper function for get_distinct_values_per_group
from class nbaR.GeoClient.
}
\details{
See also endpoint /getDistinctValues
}
\seealso{
Other nbaR.GeoClient-wrappers: \code{\link{geo_count_distinct_values_per_group}},
  \code{\link{geo_count_distinct_values}},
  \code{\link{geo_count}}, \code{\link{geo_find_by_ids}},
  \code{\link{geo_find}},
  \code{\link{geo_get_distinct_values}},
  \code{\link{geo_get_field_info}},
  \code{\link{geo_get_geo_json_for_locality}},
  \code{\link{geo_get_paths}},
  \code{\link{geo_get_settings}},
  \code{\link{geo_get_setting}},
  \code{\link{geo_is_operator_allowed}},
  \code{\link{geo_query}}
}
\concept{nbaR.GeoClient-wrappers}
