% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoClient-wrapper.r
\name{geo_get_field_info}
\alias{geo_get_field_info}
\title{Returns extended information for each field of a specimen document}
\usage{
geo_get_field_info(...)
}
\arguments{
\item{...}{additional parameters passed to get_field_info from class class nbaR.GeoClient}
}
\value{
scalar
}
\description{
This is a wrapper function for get_field_info
from class nbaR.GeoClient.
}
\details{
Info consists of whether the fields is indexed, the ElasticSearch datatype and a list of allowed operators
}
\seealso{
Other nbaR.GeoClient-wrappers: \code{\link{geo_count_distinct_values_per_group}},
  \code{\link{geo_count_distinct_values}},
  \code{\link{geo_count}}, \code{\link{geo_find_by_ids}},
  \code{\link{geo_find}},
  \code{\link{geo_get_distinct_values_per_group}},
  \code{\link{geo_get_distinct_values}},
  \code{\link{geo_get_geo_json_for_locality}},
  \code{\link{geo_get_paths}},
  \code{\link{geo_get_settings}},
  \code{\link{geo_get_setting}},
  \code{\link{geo_is_operator_allowed}},
  \code{\link{geo_query}}
}
\concept{nbaR.GeoClient-wrappers}
