% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoClient-wrapper.r
\name{geo_get_geo_json_for_locality}
\alias{geo_get_geo_json_for_locality}
\title{Retrieve a GeoJson object for a given locality}
\usage{
geo_get_geo_json_for_locality(locality = NULL, ...)
}
\arguments{
\item{locality}{, type:}

\item{...}{additional parameters passed to get_geo_json_for_locality from class nbaR.GeoClient}
}
\value{
scalar
}
\description{
This is a wrapper for the method \code{ get_geo_json_for_locality }
from class \code{ GeoClient}.
}
\details{
Returns a GeoJson polygon
}
\seealso{
Other nbaR.GeoClient-wrappers: \code{\link{geo_count_distinct_values_per_group}},
  \code{\link{geo_count_distinct_values}},
  \code{\link{geo_count}}, \code{\link{geo_find_by_ids}},
  \code{\link{geo_find}},
  \code{\link{geo_get_distinct_values_per_group}},
  \code{\link{geo_get_distinct_values}},
  \code{\link{geo_get_field_info}},
  \code{\link{geo_get_paths}},
  \code{\link{geo_get_settings}},
  \code{\link{geo_get_setting}},
  \code{\link{geo_is_operator_allowed}},
  \code{\link{geo_query}}
}
\concept{nbaR.GeoClient-wrappers}
