% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoClient-wrapper.r
\name{geo_get_setting}
\alias{geo_get_setting}
\title{Get the value of an NBA setting}
\usage{
geo_get_setting(name = NULL, ...)
}
\arguments{
\item{name}{name of setting, type:}

\item{...}{additional parameters passed to get_setting from class nbaR.GeoClient}
}
\value{
scalar
}
\description{
This is a wrapper for the method \code{ get_setting }
from class \code{ GeoClient}.
}
\details{
All settings can be queried with /metadata/getSettings
}
\seealso{
Other nbaR.GeoClient-wrappers: 
\code{\link{geo_count_distinct_values_per_group}()},
\code{\link{geo_count_distinct_values}()},
\code{\link{geo_count}()},
\code{\link{geo_find_by_ids}()},
\code{\link{geo_find}()},
\code{\link{geo_get_distinct_values_per_group}()},
\code{\link{geo_get_distinct_values}()},
\code{\link{geo_get_field_info}()},
\code{\link{geo_get_geo_json_for_locality}()},
\code{\link{geo_get_paths}()},
\code{\link{geo_get_settings}()},
\code{\link{geo_is_operator_allowed}()},
\code{\link{geo_query}()}
}
\concept{nbaR.GeoClient-wrappers}
