% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoClient-wrapper.r
\name{geo_get_settings}
\alias{geo_get_settings}
\title{List all publicly available configuration settings for the NBA}
\usage{
geo_get_settings(returnType = "data.frame", ...)
}
\arguments{
\item{returnType}{Either \code{list} or \code{data.frame} (default)}

\item{...}{additional parameters passed to get_settings from class nbaR.GeoClient}
}
\value{
scalar
}
\description{
This is a wrapper for the method \code{ get_settings }
from class \code{ GeoClient}.
}
\details{
The value of a specific setting can be queried with metadata/getSetting/{name}
}
\seealso{
Other nbaR.GeoClient-wrappers: 
\code{\link{geo_count_distinct_values_per_group}()},
\code{\link{geo_count_distinct_values}()},
\code{\link{geo_count}()},
\code{\link{geo_find_by_ids}()},
\code{\link{geo_find}()},
\code{\link{geo_get_distinct_values_per_group}()},
\code{\link{geo_get_distinct_values}()},
\code{\link{geo_get_field_info}()},
\code{\link{geo_get_geo_json_for_locality}()},
\code{\link{geo_get_paths}()},
\code{\link{geo_get_setting}()},
\code{\link{geo_is_operator_allowed}()},
\code{\link{geo_query}()}
}
\concept{nbaR.GeoClient-wrappers}
