% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoClient-wrapper.r
\name{geo_is_operator_allowed}
\alias{geo_is_operator_allowed}
\title{Checks if a given operator is allowed for a given field}
\usage{
geo_is_operator_allowed(field = NULL, operator = NULL, ...)
}
\arguments{
\item{field}{Field in geo area document, type:}

\item{operator}{operator, type:}

\item{...}{additional parameters passed to is_operator_allowed from class class nbaR.GeoClient}
}
\value{
scalar
}
\description{
This is a wrapper function for is_operator_allowed
from class nbaR.GeoClient.
}
\details{
See also metadata/getFieldInfo
}
\seealso{
Other nbaR.GeoClient-wrappers: \code{\link{geo_count_distinct_values_per_group}},
  \code{\link{geo_count_distinct_values}},
  \code{\link{geo_count}}, \code{\link{geo_find_by_ids}},
  \code{\link{geo_find}},
  \code{\link{geo_get_distinct_values_per_group}},
  \code{\link{geo_get_distinct_values}},
  \code{\link{geo_get_field_info}},
  \code{\link{geo_get_geo_json_for_locality}},
  \code{\link{geo_get_paths}},
  \code{\link{geo_get_settings}},
  \code{\link{geo_get_setting}}, \code{\link{geo_query}}
}
\concept{nbaR.GeoClient-wrappers}
