% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoClient-wrapper.r
\name{geo_query}
\alias{geo_query}
\title{Query for geo areas}
\usage{
geo_query(queryParams = list(), ...)
}
\arguments{
\item{queryParams}{Named list or vector with names being the fields to be queried and values being the values to match}

\item{...}{additional parameters passed to query from class class nbaR.GeoClient}
}
\value{
list
}
\description{
This is a wrapper function for query
from class nbaR.GeoClient.
}
\details{
Query on searchable fields to retrieve matching geo areas
}
\seealso{
Other nbaR.GeoClient-wrappers: \code{\link{geo_count_distinct_values_per_group}},
  \code{\link{geo_count_distinct_values}},
  \code{\link{geo_count}}, \code{\link{geo_find_by_ids}},
  \code{\link{geo_find}},
  \code{\link{geo_get_distinct_values_per_group}},
  \code{\link{geo_get_distinct_values}},
  \code{\link{geo_get_field_info}},
  \code{\link{geo_get_geo_json_for_locality}},
  \code{\link{geo_get_paths}},
  \code{\link{geo_get_settings}},
  \code{\link{geo_get_setting}},
  \code{\link{geo_is_operator_allowed}}
}
\concept{nbaR.GeoClient-wrappers}
