% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoArea.r
\docType{class}
\name{GeoArea}
\alias{GeoArea}
\title{GeoArea Class}
\format{
R6 class

R6 class
}
\usage{
# GeoArea$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for GeoArea objects.
}
\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor GeoArea object.

}
\item{\code{$fromList(GeoAreaList)}}{

  Create GeoArea object from list.

}

\item{\code{$toList()}}{

  Get list representation of GeoArea.

}
\item{\code{fromJSONString(GeoAreaJson)}}{

  Create GeoArea object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of GeoArea.

}
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sourceSystem}}{SourceSystem}

\item{\code{sourceSystemId}}{character}

\item{\code{recordURI}}{character}

\item{\code{id}}{character}

\item{\code{areaType}}{character}

\item{\code{locality}}{character}

\item{\code{shape}}{list}

\item{\code{source}}{character}

\item{\code{isoCode}}{character}

\item{\code{countryNL}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeoArea-new}{\code{GeoArea$new()}}
\item \href{#method-GeoArea-toList}{\code{GeoArea$toList()}}
\item \href{#method-GeoArea-fromList}{\code{GeoArea$fromList()}}
\item \href{#method-GeoArea-toJSONString}{\code{GeoArea$toJSONString()}}
\item \href{#method-GeoArea-fromJSONString}{\code{GeoArea$fromJSONString()}}
\item \href{#method-GeoArea-print}{\code{GeoArea$print()}}
\item \href{#method-GeoArea-clone}{\code{GeoArea$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeoArea-new"></a>}}
\if{latex}{\out{\hypertarget{method-GeoArea-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeoArea$new(
  sourceSystem,
  sourceSystemId,
  recordURI,
  id,
  areaType,
  locality,
  shape,
  source,
  isoCode,
  countryNL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeoArea-toList"></a>}}
\if{latex}{\out{\hypertarget{method-GeoArea-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeoArea$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeoArea-fromList"></a>}}
\if{latex}{\out{\hypertarget{method-GeoArea-fromList}{}}}
\subsection{Method \code{fromList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeoArea$fromList(GeoAreaList, typeMapping = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeoArea-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GeoArea-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeoArea$toJSONString(pretty = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeoArea-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GeoArea-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeoArea$fromJSONString(GeoAreaJson, typeMapping = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeoArea-print"></a>}}
\if{latex}{\out{\hypertarget{method-GeoArea-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeoArea$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeoArea-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeoArea-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeoArea$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
