% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoJsonObject.r
\docType{data}
\name{GeoJsonObject}
\alias{GeoJsonObject}
\title{GeoJsonObject Class}
\format{}
\usage{

}
\description{
GeoJsonObject Class
}
\section{Fields}{

\describe{
\item{\code{crs}}{Crs}

\item{\code{bbox}}{list(numeric)}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor GeoJsonObject object.

}
\item{\code{$fromList(GeoJsonObjectList)}}{

  Create GeoJsonObject object from list.

}

\item{\code{$toList()}}{

  Get list representation of GeoJsonObject.

}
\item{\code{fromJSONString(GeoJsonObjectJson)}}{

  Create GeoJsonObject object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of GeoJsonObject.

}
}
}

\keyword{datasets}
