% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_defense.R
\name{get_all_defense}
\alias{get_all_defense}
\title{Download all SportVu player defense data at once for Players and Teams.}
\usage{
get_all_defense(
  season,
  type,
  defense_categories = c("Overall", "3+Pointers", "2+Pointers", "Less+Than+6Ft",
    "Less+Than+10Ft", "Greater+Than+15Ft"),
  ...
)
}
\arguments{
\item{season}{Number of the year in which season started}

\item{type}{Specify if data is for Team or Player   c("Player,"Team")}

\item{defense_categories}{Specify which defense categories group you want to download. c("Overall","3+Pointers","2+Pointers","Less+Than+6Ft","Less+Than+10Ft","Greater+Than+15Ft")}

\item{...}{Additional arguments for get_defense}
}
\value{
Dataset from stats.nba.com
}
\description{
Wrapper for processing NBA.com player defense data
}
\examples{

season <- 2019
type <- c("Player","Team")[1]

defense_cats <- c("Overall",
                 "3+Pointers"
                 )

df <- get_all_defense(season, type, defense_cats)

}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{NBAr,}
\keyword{defense}
\keyword{players,}
\keyword{sportvu,}
\keyword{teams,}
\keyword{tracking,}
