% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_lineups.R
\name{get_all_lineups}
\alias{get_all_lineups}
\title{Download all lineups combinations and merge into one table}
\usage{
get_all_lineups(
  season,
  group_quantities = c(2:5),
  measure_types = c("Base", "Advanced", "Misc", "Scoring", "Opponent"),
  ...
)
}
\arguments{
\item{season}{Number of the year in which season started}

\item{group_quantities}{Specify number of players in a lineup c(2:5)}

\item{measure_types}{Specify which statistics group you want to download. c('Base','Advanced','Misc','Scoring','Opponent')}

\item{...}{Additional arguments for get_lineups}
}
\value{
Dataset from stats.nba.com
}
\description{
Wrapper for processing NBA.com player lineups data
}
\examples{

season <- 2019
group_quantities <- c(2:3)
measure_types <- c('Base','Advanced')
df <- get_all_lineups(season, group_quantities, measure_types)

}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{NBAr,}
\keyword{lineups,}
\keyword{players,}
\keyword{sportvu,}
\keyword{teams,}
