% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_tracking.R
\name{get_all_tracking}
\alias{get_all_tracking}
\title{Download all SportVu player tracking data at once for Players and Teams.}
\usage{
get_all_tracking(
  season,
  type,
  measure_types = c("Drives", "Defense", "Passing", "Possessions", "PullUpShot",
    "Rebounding", "Efficiency", "SpeedDistance", "ElbowTouch", "PostTouch", "PaintTouch"),
  ...
)
}
\arguments{
\item{season}{Number of the year in which season started}

\item{type}{Specify if data is for Team or Player   c("Player,"Team")}

\item{measure_types}{Specify which statistics group you want to download. c("Drives","Defense","CatchShoot","Passing","Possessions","PullUpShot","Rebounding","Efficiency","SpeedDistance","ElbowTouch","PostTouch","PaintTouch")}

\item{...}{Additional arguments for get_tracking}
}
\value{
Dataset from stats.nba.com
}
\description{
Wrapper for processing NBA.com player tracking data from http://stats.nba.com/players/drives/, http://stats.nba.com/teams/passing/ etc.
}
\examples{

season <- 2019
type <- c("Player","Team")[1]

measure_types <- c("Drives",
                 "Defense"
                 )

df <- get_all_tracking(season, type, measure_types)

}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{NBAr,}
\keyword{players,}
\keyword{sportvu,}
\keyword{teams,}
\keyword{tracking,}
