% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boxscore.R
\name{get_boxscore}
\alias{get_boxscore}
\title{Download Boxscore data for NBA game}
\usage{
get_boxscore(game_id, boxscore_type, verbose = TRUE)
}
\arguments{
\item{game_id}{Game's ID in NBA.com DB}

\item{boxscore_type}{Type of boxscore, one from the list: c("advanced","misc","playertrack","traditional")}

\item{verbose}{Defalt TRUE - prints additional information}
}
\value{
Dataset containing data from one of the boxscore pages from NBA.com. Functions enable downloading Traditional, Advanced, Miscallenous or Player Tracking data.
}
\description{
Download and process NBA.com boxscore data for given game and season.
}
\examples{

type <- c("advanced","misc","playertrack","traditional")[1]
get_boxscore(21400001,type)


}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{Advanced,}
\keyword{Boxscores,}
\keyword{Game,}
\keyword{Misc,}
\keyword{NBAr,}
\keyword{Player}
\keyword{Tracking}
\keyword{Traditional,}
\keyword{players,}
