% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hustle.R
\name{get_hustle}
\alias{get_hustle}
\title{Download hustle data for NBA game}
\usage{
get_hustle(game_id, verbose = TRUE)
}
\arguments{
\item{game_id}{Game's ID in NBA.com DB}

\item{verbose}{Defalt TRUE - prints additional information}
}
\value{
Dataset containing data about specified game hustle plays
}
\description{
Download and process NBA.com hustle data for given game
}
\examples{

get_hustle(21701181)

}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{Boxscores,}
\keyword{Game,}
\keyword{NBAr,}
\keyword{hustle}
\keyword{players,}
