% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matchups.R
\name{get_matchups}
\alias{get_matchups}
\title{Download matchups data for NBA game}
\usage{
get_matchups(game_id, verbose = TRUE)
}
\arguments{
\item{game_id}{Game's ID in NBA.com DB}

\item{verbose}{Defalt TRUE - prints additional information}
}
\value{
Dataset containing data about specified game matchups
}
\description{
Download and process NBA.com matchups data for given game
}
\examples{

get_matchups(21701181)


}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{Boxscores,}
\keyword{Game,}
\keyword{NBAr,}
\keyword{matchups}
\keyword{players,}
