% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_news.R
\name{get_news}
\alias{get_news}
\title{Download NBA weekly news}
\usage{
get_news(limit = 100, verbose = TRUE)
}
\arguments{
\item{limit}{Default 100, choose how many recent news you want}

\item{verbose}{Defalt TRUE - prints additional information}
}
\value{
Dataset with weekly news
}
\description{
Download NBA weekly news
}
\examples{
df = get_news()

}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{NBAr,}
\keyword{injuries,}
\keyword{players,}
\keyword{rotowire}
\keyword{status,}
