% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_on_off.R
\name{get_on_off}
\alias{get_on_off}
\title{Download On/Off team data}
\usage{
get_on_off(
  season,
  team_id,
  measure_type = "Base",
  date_from = "",
  date_to = "",
  opponent_team_id = "0",
  outcome = "",
  per_mode = "Totals",
  period = 0,
  season_segment = "",
  game_segment = "",
  season_type = "Regular+Season",
  verbose = TRUE
)
}
\arguments{
\item{season}{Number of the year in which season started}

\item{team_id}{Specify team id from nba.com database. Default "0" means all teams.}

\item{measure_type}{Type of statistics c('Base','Advanced','Misc','Scoring','Opponent')}

\item{date_from}{Day from which data will be collected. It is set in MM/DD/YYYY format and by default is not specified, so data is calculated for whole season.}

\item{date_to}{Day to which data will be collected. It is set in MM/DD/YYYY format and by default is not specified, so data is calculated for whole season.}

\item{opponent_team_id}{Filter by opponent's team id from nba.com database. Default "0" means all teams.}

\item{outcome}{Filter by game result. It can be a loss (L) or a win (W). By default parameter is an empty string, so both are taken into account. c("","W","L")}

\item{per_mode}{Specify if you want data divided per game or totals. Default parameter is "PerGame". c("PerGame","Totals")}

\item{period}{Choose game period for the data. 0 means whole game and it is set by default. as.character(c(0:4))}

\item{season_segment}{Choose season half for the data. Empty string means whole season and it is set by default. c("","Post All-Star","Pre All-Star")}

\item{game_segment}{Choose game half for the data. Empty string means whole game and it is set by default. c("","First Half","Overtime","Second Half")}

\item{season_type}{Choose data for preseason, regular season or postseason. Default parameter is "Regular Season". c("Regular Season","Playoffs","Pre Season","All Star")}

\item{verbose}{Defalt TRUE - prints additional information}
}
\value{
Dataset from stats.nba.com
}
\description{
Download and process NBA.com on/off data from http://stats.nba.com/team/1610612743/onoffcourt-summary/
}
\examples{

season = '2017'
team_id = "1610612739"
date_from = '01/01/2018'
period = 0
date_to = '01/25/2018'
measure_type = 'Base'
outcome = c("","W","L")[1]
per_mode =  c("Totals","PerGame","MinutesPer","Per48","Per40","Per36")[1]
season_segment = c("","Post+All-Star","Pre+All-Star")[1]
game_segment = c("","First+Half","Overtime","Second+Half")[1]
season_type = c("Regular+Season","Playoffs","PreSeason","All+Star")[1]
opponent_team_id = "0"

df =     get_on_off(season = 2017,
                team_id = 1610612739,
                measure_type,
                date_from="",
                date_to="",
                opponent_team_id="0",
                outcome="",
                period = 0,
                per_mode=per_mode,
                season_segment="",
                game_segment="",
                season_type="Regular+Season")


}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{Jokic}
\keyword{NBAr,}
\keyword{be}
\keyword{lineups,}
\keyword{mason}
\keyword{not}
\keyword{paired}
\keyword{players,}
\keyword{plumlee}
\keyword{should}
\keyword{with}
