% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_playbyplay.R
\name{get_playbyplay}
\alias{get_playbyplay}
\title{Download Playbyplay data for NBA game}
\usage{
get_playbyplay(game_id, verbose = TRUE)
}
\arguments{
\item{game_id}{Game's ID in NBA.com DB}

\item{verbose}{Defalt TRUE - prints additional information}
}
\value{
Dataset containing data from Playbyplay pages. Play-by-play provides information about every play from NBA game, one action per row.
}
\description{
Download and process NBA.com play-by-play data for given game and season.
}
\examples{
get_playbyplay(21400001)

}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{NBAr,}
\keyword{game}
\keyword{play-by-play,}
