% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_movement.R
\name{get_player_movement}
\alias{get_player_movement}
\title{Download NBA players roster transactions since 2015}
\usage{
get_player_movement(verbose = TRUE)
}
\arguments{
\item{verbose}{Defalt TRUE - prints additional information}
}
\value{
Dataset containing data about roster changes
}
\description{
Download NBA players roster transactions since 2015
}
\examples{
df = get_player_movement()

}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{NBAr,}
\keyword{contracts}
\keyword{players,}
\keyword{roster,}
\keyword{signs,}
\keyword{trades,}
