% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_players.R
\name{get_players}
\alias{get_players}
\title{Download NBA player list for NBA season

Download NBA player list for NBA season

Download and process list of NBA players for a given season}
\usage{
get_players(season, only_current = 1, verbose = TRUE)
}
\arguments{
\item{season}{Number of the year in which season started}

\item{only_current}{0 To show historical list of all players 1 to show only current season}

\item{verbose}{Defalt TRUE - prints additional information}
}
\value{
Dataset containing data about NBA players who played in specified season.
}
\description{
Download NBA player list for NBA season

Download NBA player list for NBA season

Download and process list of NBA players for a given season
}
\examples{
get_players(2017)

}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{NBAr,}
\keyword{players}
