% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_playtype.R
\name{get_playtype}
\alias{get_playtype}
\title{Download Playtype data for Players and Teams}
\usage{
get_playtype(
  season,
  type,
  playtype,
  type_grouping = "offensive",
  season_type = "Regular+Season",
  per_mode = "PerGame",
  verbose = TRUE
)
}
\arguments{
\item{season}{Number of the year in which season started}

\item{type}{Specify if data is for Team or Player   c("P,"T")}

\item{playtype}{Pick one of the playtypes provided by stats.nba.com. c("Postup","Transition","Isolation","PRBallHandler","PRRollman","Spotup","Handoff","Cut","OffScreen","OffRebound","Misc")}

\item{type_grouping}{Specify if data is for offensive or defensive side of the floor. c("offensive","defensive")
Warning! 'Transition','Cut','OffRebound', 'Misc' do not work for defensive!}

\item{season_type}{Choose data for preseason, regular season or postseason ("Regular+Season","Playoffs")}

\item{per_mode}{Specify PerGame or Totals c("PerGame","Totals")}

\item{verbose}{Defalt TRUE - prints additional information}
}
\value{
Dataset from stats.nba.com
}
\description{
Download and process NBA.com Playtype data from http://stats.nba.com/players/playtype/isolation/
}
\examples{
get_playtype(2018,"P","Isolation")

season <- 2016
type_grouping <- c("offensive","defensive")[1]
type <- c("P","T")[1]
per_mode <- c("PerGame")[1]

playtype <- c("Postup",
              "Transition",
              "Isolation",
              "PRBallHandler",
              "PRRollman",
              "Spotup",
              "Handoff",
              "Cut",
              "OffScreen",
              "OffRebound",
              "Misc"
              )[1]

season_type <- c("Regular+Season","Playoffs")[1]

df <- get_playtype(Season,Type,Playtype,TypeGrouping,SeasonType,PerMode)

}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{NBAr,}
\keyword{Playtypes,}
\keyword{players,}
\keyword{teams,}
