% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rotowire_status.R
\name{get_rotowire_status}
\alias{get_rotowire_status}
\title{Download NBA players rotowire news and updates}
\usage{
get_rotowire_status(verbose = TRUE)
}
\arguments{
\item{verbose}{Defalt TRUE - prints additional information}
}
\value{
Dataset rotowire news and updates
}
\description{
Download NBA players rotowire news and updates
}
\examples{
df = get_rotowire_status()

}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{NBAr,}
\keyword{injuries,}
\keyword{players,}
\keyword{rotowire}
\keyword{status,}
