% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_schedule.R
\name{get_schedule}
\alias{get_schedule}
\title{Download NBA season games schedule}
\usage{
get_schedule(season, verbose = TRUE)
}
\arguments{
\item{season}{Number of the year in which season started}

\item{verbose}{Defalt TRUE - prints additional information}
}
\value{
Dataset Game ID, Game Date and Time, Arena, Home, Visitor team, series score for playoffs game.
}
\description{
Downloads and process NBA season games schedule from \url{http://NBA.com}. Works for seasons 2015-16 onward.
}
\examples{
get_schedule(2015)

}
\references{
\url{http://www.NBA.com}
}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{Calendar}
\keyword{NBAr,}
\keyword{Schedule,}
