% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_boxscore.R
\name{get_team_boxscore}
\alias{get_team_boxscore}
\title{Download boxscores data for Teams}
\usage{
get_team_boxscore(
  season,
  date_from = "",
  date_to = "",
  measure_type = "",
  outcome = "",
  season_type = "Regular+Season",
  opponent_team_id = "",
  period = "",
  game_segment = "",
  verbose = TRUE
)
}
\arguments{
\item{season}{Number of the year in which season started}

\item{date_from}{Day from which data will be collected. It is set in MM/DD/YYYY format and by default is not specified, so data is calculated for whole season.}

\item{date_to}{Day to which data will be collected. It is set in MM/DD/YYYY format and by default is not specified, so data is calculated for whole season.}

\item{measure_type}{Specify type of data   c('Base','Advanced','Four+Factors','Misc','Scoring')}

\item{outcome}{Filter by game result. It can be a loss (L) or a win (W). By default parameter is an empty string, so both are taken into account. c("","W","L")}

\item{season_type}{Choose data for preseason, regular season or postseason. Default parameter is "Regular+Season". c("Regular+Season","Playoffs","Pre+Season","All+Star")}

\item{opponent_team_id}{Filter by opponent's team id from nba.com database. Default "0" means all teams.}

\item{period}{Choose game period for the data. 0 means whole game and it is set by default. as.character(c(0:4))}

\item{game_segment}{Choose game half for the data. Empty string means whole game and it is set by default. c("","First+Half","Overtime","Second+Half")}

\item{verbose}{Defalt TRUE - prints additional information}
}
\value{
Dataset from stats.nba.com
}
\description{
Download and process NBA.com team boxscores data from https://stats.nba.com/teams/boxscores-traditional/
}
\examples{


get_team_boxscore(season = 2018, measure_type = 'Advanced')

}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{NBAr,}
\keyword{boxscores}
\keyword{teams,}
