% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracking.R
\name{get_tracking}
\alias{get_tracking}
\title{Download SportVu player tracking data for Players and Teams}
\usage{
get_tracking(
  season,
  type,
  measure_type,
  per_mode = "PerGame",
  season_type = "Regular+Season",
  season_segment = "",
  game_segment = "",
  date_from = "",
  date_to = "",
  outcome = "",
  period = 0,
  opponent_team_id = 0,
  verbose = TRUE
)
}
\arguments{
\item{season}{Number of the year in which season started}

\item{type}{Specify if data is for Team or Player   c("Player,"Team")}

\item{measure_type}{Specify which statistics group you want to download. c("Drives","Defense","CatchShoot","Passing","Possessions","PullUpShot","Rebounding","Efficiency","SpeedDistance","ElbowTouch","PostTouch","PaintTouch")[1]}

\item{per_mode}{Specify if you want data divided per game or totals. Default parameter is "PerGame". c("PerGame","Totals")}

\item{season_type}{Choose data for preseason, regular season or postseason. Default parameter is "Regular+Season". c("Regular+Season","Playoffs","Pre+Season","All+Star")}

\item{season_segment}{Choose season half for the data. Empty string means whole season and it is set by default. c("","Post+All-Star","Pre+All-Star")}

\item{game_segment}{Choose game half for the data. Empty string means whole game and it is set by default. c("","First+Half","Overtime","Second+Half")}

\item{date_from}{Day from which data will be collected. It is set in MM/DD/YYYY format and by default is not specified, so data is calculated for whole season.}

\item{date_to}{Day to which data will be collected. It is set in MM/DD/YYYY format and by default is not specified, so data is calculated for whole season.}

\item{outcome}{Filter by game result. It can be a loss (L) or a win (W). By default parameter is an empty string, so both are taken into account. c("","W","L")}

\item{period}{Choose game period for the data. 0 means whole game and it is set by default. as.character(c(0:4))}

\item{opponent_team_id}{Filter by opponent's team id from nba.com database. Default "0" means all teams.}

\item{verbose}{Defalt TRUE - prints additional information}
}
\value{
Dataset from stats.nba.com
}
\description{
Download and process NBA.com player tracking data from http://stats.nba.com/players/drives/, http://stats.nba.com/teams/passing/ etc.
}
\examples{

season <- 2016
type <- c("Player","Team")[1]

measure_type <- c("Drives",
                 "Defense",
                 "CatchShoot",
                 "Passing",
                 "Possessions",
                 "PullUpShot",
                 "Rebounding",
                 "Efficiency",
                 "SpeedDistance",
                 "ElbowTouch",
                 "PostTouch",
                 "PaintTouch"
                 )[1]

per_mode <- c("PerGame","Totals")[1]
season_type <- c("Regular+Season","Playoffs","Pre+Season","All+Star")[1]
season_segment <- c("","Post+All-Star","Pre All-Star")[1]
game_segment <- c("","First+Half","Overtime","Second+Half")[1]
period <- 0
date_from <- "01/01/2017"
date_to <- "04/30/2017"
outcome <- c("","W","L")[1]
opponent_team_id <- 0

df <- get_tracking(
                                season,
                                type,
                                measure_type,
                                per_mode,
                                season_type,
                                season_segment,
                                game_segment,
                                date_from,
                                date_to,
                                outcome,
                                period,
                                opponent_team_id
                                )

}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{NBAr,}
\keyword{players,}
\keyword{sportvu,}
\keyword{teams,}
\keyword{tracking,}
