% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoxscore.R
\name{getBoxscore}
\alias{getBoxscore}
\title{Download Boxscore data for NBA game}
\usage{
getBoxscore(gameID, Type)
}
\arguments{
\item{gameID}{Game's ID in NBA.com DB}

\item{Type}{Type of boxscore, one from the list: c("advanced","misc","playertrack","traditional")}
}
\value{
Dataset containing data from one of the boxscore pages from NBA.com. Functions enable downloading Traditional, Advanced, Miscallenous or Player Tracking data.
}
\description{
Download and process NBA.com boxscore data for given game and season.
}
\examples{

Type <- c("advanced","misc","playertrack","traditional")[1]
getBoxscore("0021400001",Type)

#playoffs:
getBoxscore("0041600101",Type)

gamelist<- paste("00",as.character(c(21400001:21401230)),sep="")
boxscores <- lapply(gamelist, getBoxscore, Type = "traditional")
player.tracking <- do.call(rbind.data.frame,boxscores)

}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{Advanced,}
\keyword{Boxscores,}
\keyword{Game,}
\keyword{Misc,}
\keyword{NBAr,}
\keyword{Player}
\keyword{Tracking}
\keyword{Traditional,}
\keyword{players,}
