% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNKCJE.R
\name{getDoneGames}
\alias{getDoneGames}
\title{Get the list of played games from calendar table from PostgreSQL}
\usage{
getDoneGames(dbname, date = Sys.Date() - 1, schema = "rd",
  pghost = "localhost", pgport = 5432, pguser = "postgres",
  pgpassword = "postgres", calendartbl = "calendar", gamecol = "GAME_ID")
}
\description{
Get the list of played games from calendar table from PostgreSQL
}
