% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneral.R
\name{getGeneral}
\alias{getGeneral}
\title{Download general data for Players and Teams}
\usage{
getGeneral(Season, Type, MeasureType, PerMode = "PerGame",
  SeasonType = "Regular Season", SeasonSegment = "", GameSegment = "",
  DateFrom = "", DateTo = "", Outcome = "", Period = "0",
  OpponentTeamID = "0")
}
\arguments{
\item{Season}{Number of the year in which season started}

\item{Type}{Specify if data is for Team or Player   c("Player,"Team")}

\item{MeasureType}{Specify which statistics group you want to download. c("Base","Advanced","Misc","Scoring","Usage","Opponent","Defense", "Four Factors")}

\item{PerMode}{Specify if you want data divided per game or totals. Default parameter is "PerGame". c("PerGame","Totals")}

\item{SeasonType}{Choose data for preseason, regular season or postseason. Default parameter is "Regular Season". c("Regular Season","Playoffs","Pre Season","All Star")}

\item{SeasonSegment}{Choose season half for the data. Empty string means whole season and it is set by default. c("","Post All-Star","Pre All-Star")}

\item{GameSegment}{Choose game half for the data. Empty string means whole game and it is set by default. c("","First Half","Overtime","Second Half")}

\item{DateFrom}{Day from which data will be collected. It is set in MM/DD/YYYY format and by default is not specified, so data is calculated for whole season.}

\item{DateTo}{Day to which data will be collected. It is set in MM/DD/YYYY format and by default is not specified, so data is calculated for whole season.}

\item{Outcome}{Filter by game result. It can be a loss (L) or a win (W). By default parameter is an empty string, so both are taken into account. c("","W","L")}

\item{Period}{Choose game period for the data. 0 means whole game and it is set by default. as.character(c(0:4))}

\item{OpponentTeamID}{Filter by opponent's team id from nba.com database. Default "0" means all teams.}
}
\value{
Dataset from stats.nba.com
}
\description{
Download and process NBA.com general data from http://stats.nba.com/players/traditional/, http://stats.nba.com/teams/advanced/ etc.
}
\examples{

Season <- "2016"
Type <- c("Player","Team")[1]

MeasureType <- c("Base",
                 "Advanced",
                 "Misc",
                 "Scoring",
                 "Usage",
                 "Opponent",
                 "Defense",
                 "Four Factors"
                 )[1]

PerMode <- c("PerGame","Totals")[1]
SeasonType <- c("Regular Season","Playoffs","Pre Season","All Star")[1]
SeasonSegment <- c("","Post All-Star","Pre All-Star")[1]
GameSegment <- c("","First Half","Overtime","Second Half")[1]
Period <- as.character(c(0:4))[1]
DateFrom <- "01/01/2017"
DateTo <- "04/30/2017"
Outcome <- c("","W","L")[1]
OpponentTeamID <- c("0")

general.dataset <- getGeneral(
                              Season,
                              Type,
                              MeasureType,
                              PerMode,
                              SeasonType,
                              SeasonSegment,
                              GameSegment,
                              DateFrom,
                              DateTo,
                              Outcome,
                              Period,
                              OpponentTeamID
                              )

}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{NBAr,}
\keyword{advanced}
\keyword{general,}
\keyword{players,}
\keyword{teams,}
\keyword{traditional,}
