% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNKCJE.R
\name{getLastGame}
\alias{getLastGame}
\title{Get last game from calendar table from PostgreSQL}
\usage{
getLastGame(dbname, date = Sys.Date() - 1, schema = "rd",
  pghost = "localhost", pgport = 5432, pguser = "postgres",
  pgpassword = "postgres", calendartbl = "calendar", gamecol = "GAME_ID")
}
\description{
Get last game from calendar table from PostgreSQL
}
