% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNBADB.R
\name{getNBADB}
\alias{getNBADB}
\title{Function Wrapper for getting all NBA datasets into R}
\usage{
getNBADB(dbname, schema, pghost = "localhost", pgport = 5432,
  pguser = "postgres", pgpassword = "postgres")
}
\arguments{
\item{dbname}{Name of the PostgreSQL database}

\item{schema}{Name of the PostgreSQL schema in the database}

\item{pghost}{Name or IP address of DB host. By default it is localhost.}

\item{pgport}{Port number of PostgreSQL server. By default it is 5432.}

\item{pguser}{Database user login. By default it is "postgres"}

\item{pgpassword}{Database user password By default it is "postgres"}
}
\value{
R datasets
}
\description{
Connect to PostgreSQL database to read all tables from schema into  R environment
}
\examples{
getNBADB("NBA","nba16","localhost",5432,"postgres","postgres")

}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{NBAr,}
\keyword{PostgreSQL}

