% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNKCJE.R
\name{getPgData}
\alias{getPgData}
\title{Get all tables from database}
\usage{
getPgData(dbname, schema, pghost = "localhost", pgport = 5432,
  pguser = "postgres", pgpassword = "postgres")
}
\description{
Get all tables from database
}
