% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlaybyplay.R
\name{getPlaybyplay}
\alias{getPlaybyplay}
\title{Download Playbyplay data for NBA game}
\usage{
getPlaybyplay(gameID)
}
\arguments{
\item{gameID}{Game's ID in NBA.com DB}
}
\value{
Dataset containing data from Playbyplay pages. Play-by-play provides information about every play from NBA game, one action per row.
}
\description{
Download and process NBA.com play-by-play data for given game and season.
}
\examples{
getPlaybyplay("0021400001")

gamelist<- paste("00",as.character(c(21400001:21401230)),sep="")
pbps <- lapply(gamelist, getPlaybyplay)
playbyplay <- do.call(rbind.data.frame,pbps)

}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{NBAr,}
\keyword{game}
\keyword{play-by-play,}

