% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerlist.R
\name{getPlayerlist}
\alias{getPlayerlist}
\title{Download NBA player list for NBA season}
\usage{
getPlayerlist(season)
}
\arguments{
\item{season}{Number of the year in which season started}
}
\value{
Dataset containing data about NBA players who played in specified season.
}
\description{
Download and process list of NBA players for a given season
}
\examples{
getPlayerlist("2014")


}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{NBAr,}
\keyword{players}

