% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNKCJE.R
\name{getPlayers}
\alias{getPlayers}
\title{Get list of players from a database}
\usage{
getPlayers(dbname, schema = "rd", pghost = "localhost", pgport = 5432,
  pguser = "postgres", pgpassword = "postgres", playertbl = "playerlist",
  playercol = "PLAYER_ID")
}
\description{
Get list of players from a database
}
