% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlaytype.R
\name{getPlaytype}
\alias{getPlaytype}
\title{Download Playtype data for Players and Teams}
\usage{
getPlaytype(Season = "", Type = "", Category = "", Names = "",
  SeasonType = "")
}
\arguments{
\item{Season}{Number of the year in which season started}

\item{Type}{Specify if data is for Team or Player   c("Player,"Team")}

\item{Category}{Pick one of the playtypes provided by stats.nba.com. These are Isolation, Transition, Postup, PRBallHandler, PRRollman, Spotup, Handoff,Postup, Cut, OffScreen, Putbacks and Misc.
c("Isolation","Transition","PRBallHandler","PRRollman","Postup","Spotup","Handoff","Cut","OffScreen","OffRebound","Misc")}

\item{Names}{Specify if data is for offensive or defensive side of the floor. c("offensive","defensive")}

\item{SeasonType}{Choose data for preseason, regular season or postseason c("Reg","Post","Pre")}
}
\value{
Dataset from stats.nba.com
}
\description{
Download and process NBA.com Playtype data from http://stats.nba.com/players/playtypeexample/ or http://stats.nba.com/teams/playtypeexample/
}
\examples{
getPlaytype("2016","Team","Handoff","offensive","Pre")
getPlaytype("2015","Player","Isolation","defensive","Post")

Season <- "2016"
Names <- c("offensive","defensive")[1]
Type <- c("Player","Team")[1]

Playtype <- c("PostTouch","ElbowTouch","PaintTouch"
              ,"CatchShoot","Drives","PullUpShot","Isolation",
              "PRBallHandler","PRRollman","Spotup","Handoff","Cut","OffScreen",
              "OffRebound")[1]

SeasonType <- c("Reg","Post","Pre")[1]

playtypes.dataset <- getPlaytype(Season,Type,Category,Names,SeasonType)

}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{NBAr,}
\keyword{Playtypes,}
\keyword{players,}
\keyword{teams,}

