% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSchedule.R
\name{getSchedule}
\alias{getSchedule}
\title{Download NBA season games schedule}
\usage{
getSchedule(month, season)
}
\arguments{
\item{month}{Character name of the month}

\item{season}{Number of the year in which season started}
}
\value{
None
}
\description{
Downloads and process NBA season games schedule from \url{http://www.basketball-reference.com}
}
\examples{
getSchedule("january","2016")

}
\author{
Patrick Chodowski, \email{kbroman@biostat.wisc.edu}
}
\references{
\url{http://www.basketball-reference.com}
}
\keyword{Calendar}
\keyword{NBAr,}
\keyword{Schedule,}

