% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{gt_theme_nba}
\alias{gt_theme_nba}
\title{gt table theme}
\usage{
gt_theme_nba(GT_TBL)
}
\arguments{
\item{GT_TBL}{The data frame that contains the data}
}
\description{
Function to apply gt styling to table
}
\examples{
test <- NBA_example_bar_plot |>
gt::gt() |>
NBA.package::gt_theme_nba()

}
