% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Iptc4xmpExt.r
\docType{class}
\name{Iptc4xmpExt}
\alias{Iptc4xmpExt}
\title{Iptc4xmpExt Class}
\format{R6 class}
\usage{
# Iptc4xmpExt$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for Iptc4xmpExt objects.
}
\section{Fields}{

\describe{
\item{\code{locationShown}}{character}

\item{\code{worldRegion}}{character}

\item{\code{countryCode}}{character}

\item{\code{countryName}}{character}

\item{\code{provinceState}}{character}

\item{\code{city}}{character}

\item{\code{sublocation}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Iptc4xmpExt object.

}
\item{\code{$fromList(Iptc4xmpExtList)}}{

  Create Iptc4xmpExt object from list.

}

\item{\code{$toList()}}{

  Get list representation of Iptc4xmpExt.

}
\item{\code{fromJSONString(Iptc4xmpExtJson)}}{

  Create Iptc4xmpExt object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Iptc4xmpExt.

}
}
}

\keyword{datasets}
