% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineString.r
\docType{class}
\name{LineString}
\alias{LineString}
\title{LineString Class}
\format{R6 class}
\usage{
# LineString$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for LineString objects.
}
\section{Fields}{

\describe{
\item{\code{crs}}{Crs}

\item{\code{bbox}}{list(numeric)}

\item{\code{coordinates}}{list(LngLatAlt)}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor LineString object.

}
\item{\code{$fromList(LineStringList)}}{

  Create LineString object from list.

}

\item{\code{$toList()}}{

  Get list representation of LineString.

}
\item{\code{fromJSONString(LineStringJson)}}{

  Create LineString object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of LineString.

}
}
}

\keyword{datasets}
