% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LithoStratigraphy.r
\docType{data}
\name{LithoStratigraphy}
\alias{LithoStratigraphy}
\title{LithoStratigraphy Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LithoStratigraphy
}
\description{
LithoStratigraphy Class
}
\section{Fields}{

\describe{
\item{\code{bed}}{}
}}

\keyword{datasets}
