% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LngLatAlt.r
\docType{data}
\name{LngLatAlt}
\alias{LngLatAlt}
\title{LngLatAlt Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LngLatAlt
}
\description{
LngLatAlt Class
}
\section{Fields}{

\describe{
\item{\code{longitude}}{}

\item{\code{latitude}}{}

\item{\code{altitude}}{}

\item{\code{additionalElements}}{}
}}

\keyword{datasets}
