% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataClient-wrapper.r
\name{metadata_get_controlled_list_phase_or_stage}
\alias{metadata_get_controlled_list_phase_or_stage}
\title{Get allowed values for the field &#39;PhaseOrStage&#39; in a specimen document}
\usage{
metadata_get_controlled_list_phase_or_stage(...)
}
\arguments{
\item{...}{additional parameters passed to get_controlled_list_phase_or_stage from class class nbaR.MetadataClient}
}
\value{
scalar
}
\description{
This is a wrapper function for get_controlled_list_phase_or_stage
from class nbaR.MetadataClient.
}
\details{
See also endpoint /getControlledLists
}
\seealso{
Other nbaR.MetadataClient-wrappers: \code{\link{metadata_get_allowed_date_formats}},
  \code{\link{metadata_get_controlled_list_sex}},
  \code{\link{metadata_get_controlled_list_specimen_type_status}},
  \code{\link{metadata_get_controlled_list_taxonomic_status}},
  \code{\link{metadata_get_controlled_lists}},
  \code{\link{metadata_get_rest_services}},
  \code{\link{metadata_get_settings}},
  \code{\link{metadata_get_setting}},
  \code{\link{metadata_get_source_systems}}
}
\concept{nbaR.MetadataClient-wrappers}
