% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataClient-wrapper.r
\name{metadata_get_controlled_list_spatial_datum}
\alias{metadata_get_controlled_list_spatial_datum}
\title{Get allowed values for the field &#39;SpatialDatum&#39;}
\usage{
metadata_get_controlled_list_spatial_datum(returnType = "data.frame", ...)
}
\arguments{
\item{returnType}{Either \code{list} or \code{data.frame} (default)}

\item{...}{additional parameters passed to get_controlled_list_spatial_datum from class nbaR.MetadataClient}
}
\value{
scalar
}
\description{
This is a wrapper for the method \code{ get_controlled_list_spatial_datum }
from class \code{ MetadataClient}.
}
\seealso{
Other nbaR.MetadataClient-wrappers: 
\code{\link{metadata_get_allowed_date_formats}()},
\code{\link{metadata_get_controlled_list_area_class}()},
\code{\link{metadata_get_controlled_list_license_type}()},
\code{\link{metadata_get_controlled_list_license}()},
\code{\link{metadata_get_controlled_list_sex}()},
\code{\link{metadata_get_controlled_list_specimen_type_status}()},
\code{\link{metadata_get_controlled_list_taxon_relation_type}()},
\code{\link{metadata_get_controlled_list_taxonomic_status}()},
\code{\link{metadata_get_controlled_lists}()},
\code{\link{metadata_get_rest_services}()},
\code{\link{metadata_get_settings}()},
\code{\link{metadata_get_setting}()},
\code{\link{metadata_get_source_systems}()}
}
\concept{nbaR.MetadataClient-wrappers}
