% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataClient-wrapper.r
\name{metadata_get_controlled_lists}
\alias{metadata_get_controlled_lists}
\title{Get the names of fields for which a controlled vocabulary exists}
\usage{
metadata_get_controlled_lists(...)
}
\arguments{
\item{...}{additional parameters passed to get_controlled_lists from class class nbaR.MetadataClient}
}
\value{
scalar
}
\description{
This is a wrapper function for get_controlled_lists
from class nbaR.MetadataClient.
}
\details{
Possible values for fields with controlled vocabularies can be queried with metadata/getControlledList/{field}
}
\seealso{
Other nbaR.MetadataClient-wrappers: \code{\link{metadata_get_allowed_date_formats}},
  \code{\link{metadata_get_controlled_list_phase_or_stage}},
  \code{\link{metadata_get_controlled_list_sex}},
  \code{\link{metadata_get_controlled_list_specimen_type_status}},
  \code{\link{metadata_get_controlled_list_taxonomic_status}},
  \code{\link{metadata_get_rest_services}},
  \code{\link{metadata_get_settings}},
  \code{\link{metadata_get_setting}},
  \code{\link{metadata_get_source_systems}}
}
\concept{nbaR.MetadataClient-wrappers}
