% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataClient-wrapper.r
\name{metadata_get_setting}
\alias{metadata_get_setting}
\title{Get the value of an NBA setting}
\usage{
metadata_get_setting(name = NULL, ...)
}
\arguments{
\item{name}{name of setting, type:}

\item{...}{additional parameters passed to get_setting from class nbaR.MetadataClient}
}
\value{
scalar
}
\description{
This is a wrapper for the method \code{ get_setting }
from class \code{ MetadataClient}.
}
\details{
All settings can be queried with /metadata/getSettings
}
\seealso{
Other nbaR.MetadataClient-wrappers: \code{\link{metadata_get_allowed_date_formats}},
  \code{\link{metadata_get_controlled_list_phase_or_stage}},
  \code{\link{metadata_get_controlled_list_sex}},
  \code{\link{metadata_get_controlled_list_specimen_type_status}},
  \code{\link{metadata_get_controlled_list_taxonomic_status}},
  \code{\link{metadata_get_controlled_lists}},
  \code{\link{metadata_get_rest_services}},
  \code{\link{metadata_get_settings}},
  \code{\link{metadata_get_source_systems}}
}
\concept{nbaR.MetadataClient-wrappers}
