% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataClient-wrapper.r
\name{metadata_get_settings}
\alias{metadata_get_settings}
\title{List all publicly available configuration settings for the NBA}
\usage{
metadata_get_settings(returnType = "data.frame", ...)
}
\arguments{
\item{returnType}{Either 'list' or 'data.frame', defaults to 'data.frame'}

\item{...}{additional parameters passed to get_settings from class nbaR.MetadataClient}
}
\value{
scalar
}
\description{
This is a wrapper function for get_settings
from class nbaR.MetadataClient.
}
\details{
The value of a specific setting can be queried with metadata/getSetting/{name}
}
\seealso{
Other nbaR.MetadataClient-wrappers: \code{\link{metadata_get_allowed_date_formats}},
  \code{\link{metadata_get_controlled_list_phase_or_stage}},
  \code{\link{metadata_get_controlled_list_sex}},
  \code{\link{metadata_get_controlled_list_specimen_type_status}},
  \code{\link{metadata_get_controlled_list_taxonomic_status}},
  \code{\link{metadata_get_controlled_lists}},
  \code{\link{metadata_get_rest_services}},
  \code{\link{metadata_get_setting}},
  \code{\link{metadata_get_source_systems}}
}
\concept{nbaR.MetadataClient-wrappers}
