% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataClient-wrapper.r
\name{metadata_get_source_systems}
\alias{metadata_get_source_systems}
\title{Get the data sources from which the data was retrieved}
\usage{
metadata_get_source_systems(returnType = "data.frame", ...)
}
\arguments{
\item{returnType}{Either \code{list} or \code{data.frame} (default)}

\item{...}{additional parameters passed to get_source_systems from class nbaR.MetadataClient}
}
\value{
list or data.frame, as specified by \code{returnType}
}
\description{
This is a wrapper for the method \code{ get_source_systems }
from class \code{ MetadataClient}.
}
\details{
Returns code and name of all source systems
}
\seealso{
Other nbaR.MetadataClient-wrappers: \code{\link{metadata_get_allowed_date_formats}},
  \code{\link{metadata_get_controlled_list_phase_or_stage}},
  \code{\link{metadata_get_controlled_list_sex}},
  \code{\link{metadata_get_controlled_list_specimen_type_status}},
  \code{\link{metadata_get_controlled_list_taxonomic_status}},
  \code{\link{metadata_get_controlled_lists}},
  \code{\link{metadata_get_rest_services}},
  \code{\link{metadata_get_settings}},
  \code{\link{metadata_get_setting}}
}
\concept{nbaR.MetadataClient-wrappers}
