% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataApi.r, R/MetadataClient.r
\docType{data}
\name{MetadataClient}
\alias{MetadataClient}
\alias{MetadataClient}
\title{Metadata operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MetadataClient

MetadataClient
}
\description{
nbaR.Metadata

nbaR.Metadata
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}

\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_allowed_date_formats Get allowed values for dates in queries


get_controlled_list_phase_or_stage Get allowed values for the field &#39;PhaseOrStage&#39; in a specimen document


get_controlled_list_sex Get allowed values for the field &#39;Sex&#39; in a specimen document


get_controlled_list_specimen_type_status Get allowed values for the field &#39;SpecimenTypeStatus&#39; in a specimen document


get_controlled_list_taxonomic_status Get allowed values for the field &#39;TaxonomicStatus&#39; in specimen and taxon documents


get_controlled_lists Get the names of fields for which a controlled vocabulary exists


get_rest_services List all available REST services and their parameters


get_setting Get the value of an NBA setting


get_settings List all publicly available configuration settings for the NBA


get_source_systems Get the data sources from which the data was retrieved

}


\describe{

get_allowed_date_formats Get allowed values for dates in queries


get_controlled_list_phase_or_stage Get allowed values for the field &#39;PhaseOrStage&#39; in a specimen document


get_controlled_list_sex Get allowed values for the field &#39;Sex&#39; in a specimen document


get_controlled_list_specimen_type_status Get allowed values for the field &#39;SpecimenTypeStatus&#39; in a specimen document


get_controlled_list_taxonomic_status Get allowed values for the field &#39;TaxonomicStatus&#39; in specimen and taxon documents


get_controlled_lists Get the names of fields for which a controlled vocabulary exists


get_rest_services List all available REST services and their parameters


get_setting Get the value of an NBA setting


get_settings List all publicly available configuration settings for the NBA


get_source_systems Get the data sources from which the data was retrieved

}
}

\keyword{datasets}
