% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monomial.r
\docType{data}
\name{Monomial}
\alias{Monomial}
\title{Monomial Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Monomial
}
\description{
Monomial Class
}
\section{Fields}{

\describe{
\item{\code{rank}}{character}

\item{\code{name}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Monomial object.

}
\item{\code{$fromList(MonomialList)}}{

  Create Monomial object from list.

}

\item{\code{$toList()}}{

  Get list representation of Monomial.

}
\item{\code{fromJSONString(MonomialJson)}}{

  Create Monomial object from list.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Monomial.

}
}
}

\keyword{datasets}
