% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultimediaClient-wrapper.r
\name{multimedia_download_query}
\alias{multimedia_download_query}
\title{Dynamic download service: Query for multimedia objects and return result as a stream ...}
\usage{
multimedia_download_query(queryParams = list(), ...)
}
\arguments{
\item{queryParams}{Named list or vector with names being the fields to be queried and values being the values to match}

\item{...}{additional parameters passed to download_query from class nbaR.MultimediaClient}
}
\description{
This is a wrapper for the method \code{ download_query }
from class \code{ MultimediaClient}.
}
\details{
Query with query parameters or querySpec JSON. ...
}
\seealso{
Other nbaR.MultimediaClient-wrappers: 
\code{\link{multimedia_count_distinct_values_per_group}()},
\code{\link{multimedia_count_distinct_values}()},
\code{\link{multimedia_count}()},
\code{\link{multimedia_find_by_ids}()},
\code{\link{multimedia_find}()},
\code{\link{multimedia_get_distinct_values_per_group}()},
\code{\link{multimedia_get_distinct_values}()},
\code{\link{multimedia_get_field_info}()},
\code{\link{multimedia_get_paths}()},
\code{\link{multimedia_get_settings}()},
\code{\link{multimedia_get_setting}()},
\code{\link{multimedia_is_operator_allowed}()},
\code{\link{multimedia_query_with_name_resolution}()},
\code{\link{multimedia_query}()}
}
\concept{nbaR.MultimediaClient-wrappers}
