% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultimediaClient-wrapper.r
\name{multimedia_find}
\alias{multimedia_find}
\title{Find a multimedia document by id}
\usage{
multimedia_find(id = NULL, ...)
}
\arguments{
\item{id}{id of multimedia document, type:}

\item{...}{additional parameters passed to find from class class nbaR.MultimediaClient}
}
\value{
list
}
\description{
This is a wrapper function for find
from class nbaR.MultimediaClient.
}
\details{
If found, returns a single multimedia document
}
\seealso{
Other nbaR.MultimediaClient-wrappers: \code{\link{multimedia_count_distinct_values_per_group}},
  \code{\link{multimedia_count_distinct_values}},
  \code{\link{multimedia_count}},
  \code{\link{multimedia_download_query}},
  \code{\link{multimedia_find_by_ids}},
  \code{\link{multimedia_get_distinct_values_per_group}},
  \code{\link{multimedia_get_distinct_values}},
  \code{\link{multimedia_get_field_info}},
  \code{\link{multimedia_get_paths}},
  \code{\link{multimedia_get_settings}},
  \code{\link{multimedia_get_setting}},
  \code{\link{multimedia_is_operator_allowed}},
  \code{\link{multimedia_query}}
}
\concept{nbaR.MultimediaClient-wrappers}
