% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultimediaClient-wrapper.r
\name{multimedia_get_paths}
\alias{multimedia_get_paths}
\title{Returns the full path of all fields within a document}
\usage{
multimedia_get_paths(returnType = "data.frame", ...)
}
\arguments{
\item{returnType}{Either 'list' or 'data.frame', defaults to 'data.frame'}

\item{...}{additional parameters passed to get_paths from class nbaR.MultimediaClient}
}
\value{
scalar
}
\description{
This is a wrapper function for get_paths
from class nbaR.MultimediaClient.
}
\details{
See also metadata/getFieldInfo for all allowed operators per field
}
\seealso{
Other nbaR.MultimediaClient-wrappers: \code{\link{multimedia_count_distinct_values_per_group}},
  \code{\link{multimedia_count_distinct_values}},
  \code{\link{multimedia_count}},
  \code{\link{multimedia_download_query}},
  \code{\link{multimedia_find_by_ids}},
  \code{\link{multimedia_find}},
  \code{\link{multimedia_get_distinct_values_per_group}},
  \code{\link{multimedia_get_distinct_values}},
  \code{\link{multimedia_get_field_info}},
  \code{\link{multimedia_get_settings}},
  \code{\link{multimedia_get_setting}},
  \code{\link{multimedia_is_operator_allowed}},
  \code{\link{multimedia_query}}
}
\concept{nbaR.MultimediaClient-wrappers}
