% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultimediaClient-wrapper.r
\name{multimedia_is_operator_allowed}
\alias{multimedia_is_operator_allowed}
\title{Checks if a given operator is allowed for a given field}
\usage{
multimedia_is_operator_allowed(field = NULL, operator = NULL, ...)
}
\arguments{
\item{field}{multimedia document field, type:}

\item{operator}{operator, type:}

\item{...}{additional parameters passed to is_operator_allowed from class nbaR.MultimediaClient}
}
\value{
scalar
}
\description{
This is a wrapper function for is_operator_allowed
from class nbaR.MultimediaClient.
}
\details{
See also metadata/getFieldInfo
}
\seealso{
Other nbaR.MultimediaClient-wrappers: \code{\link{multimedia_count_distinct_values_per_group}},
  \code{\link{multimedia_count_distinct_values}},
  \code{\link{multimedia_count}},
  \code{\link{multimedia_download_query}},
  \code{\link{multimedia_find_by_ids}},
  \code{\link{multimedia_find}},
  \code{\link{multimedia_get_distinct_values_per_group}},
  \code{\link{multimedia_get_distinct_values}},
  \code{\link{multimedia_get_field_info}},
  \code{\link{multimedia_get_paths}},
  \code{\link{multimedia_get_settings}},
  \code{\link{multimedia_get_setting}},
  \code{\link{multimedia_query}}
}
\concept{nbaR.MultimediaClient-wrappers}
