% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultimediaApi.r, R/MultimediaClient.r
\docType{data}
\name{MultimediaClient}
\alias{MultimediaClient}
\alias{MultimediaClient}
\title{Multimedia operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultimediaClient

MultimediaClient
}
\description{
nbaR.Multimedia

nbaR.Multimedia
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}

\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

count_http_get1 Get the number of multimedia documents matching a condition


count_http_post_json1 Get the number of multimedia documents matching a condition


find2 Find a multimedia document by id


find_by_ids2 Find multimedia document by ids


get_distinct_values_http_get1 Get all different values that can be found for one field


get_distinct_values_http_post_json1 Get all different values that exist for a field


get_field_info1 Returns extended information for each field of a multimedia document


get_paths1 Returns the full path of all fields within a document


get_settings2 List all publicly available configuration settings for the NBA


get_settings3 Get the value of an NBA setting


is_operator_allowed1 Checks if a given operator is allowed for a given field


query_http_get1 Query for multimedia documents


query_http_post_json1 Query for multimedia documents

}


\describe{

count_http_get1 Get the number of multimedia documents matching a condition


count_http_post_json1 Get the number of multimedia documents matching a condition


find2 Find a multimedia document by id


find_by_ids2 Find multimedia document by ids


get_distinct_values_http_get1 Get all different values that can be found for one field


get_distinct_values_http_post_json1 Get all different values that exist for a field


get_field_info1 Returns extended information for each field of a multimedia document


get_paths1 Returns the full path of all fields within a document


get_settings2 Get the value of an NBA setting


get_settings3 List all publicly available configuration settings for the NBA


is_operator_allowed1 Checks if a given operator is allowed for a given field


query_http_get1 Query for multimedia documents


query_http_post_json1 Query for multimedia documents

}
}

\keyword{datasets}
