% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiMediaContentIdentification.r
\docType{class}
\name{MultiMediaContentIdentification}
\alias{MultiMediaContentIdentification}
\title{MultiMediaContentIdentification Class}
\format{R6 class}
\usage{
# MultiMediaContentIdentification$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for MultiMediaContentIdentification objects.
}
\section{Fields}{

\describe{
\item{\code{taxonRank}}{character}

\item{\code{scientificName}}{ScientificName}

\item{\code{typeStatus}}{character}

\item{\code{dateIdentified}}{character}

\item{\code{defaultClassification}}{DefaultClassification}

\item{\code{systemClassification}}{list(Monomial)}

\item{\code{vernacularNames}}{list(VernacularName)}

\item{\code{identificationQualifiers}}{list(character)}

\item{\code{identifiers}}{list(Agent)}

\item{\code{taxonomicEnrichments}}{list(TaxonomicEnrichment)}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor MultiMediaContentIdentification object.

}
\item{\code{$fromList(MultiMediaContentIdentificationList)}}{

  Create MultiMediaContentIdentification object from list.

}

\item{\code{$toList()}}{

  Get list representation of MultiMediaContentIdentification.

}
\item{\code{fromJSONString(MultiMediaContentIdentificationJson)}}{

  Create MultiMediaContentIdentification object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of MultiMediaContentIdentification.

}
}
}

\keyword{datasets}
