% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPoint.r
\docType{data}
\name{MultiPoint}
\alias{MultiPoint}
\title{MultiPoint Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultiPoint
}
\description{
MultiPoint Class
}
\section{Fields}{

\describe{
\item{\code{coordinates}}{}
}}

\keyword{datasets}
