% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPolygon.r
\docType{class}
\name{MultiPolygon}
\alias{MultiPolygon}
\title{MultiPolygon Class}
\format{R6 class}
\usage{
# MultiPolygon$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for MultiPolygon objects.
}
\section{Fields}{

\describe{
\item{\code{crs}}{Crs}

\item{\code{bbox}}{list(numeric)}

\item{\code{coordinates}}{list(LngLatAlt)}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor MultiPolygon object.

}
\item{\code{$fromList(MultiPolygonList)}}{

  Create MultiPolygon object from list.

}

\item{\code{$toList()}}{

  Get list representation of MultiPolygon.

}
\item{\code{fromJSONString(MultiPolygonJson)}}{

  Create MultiPolygon object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of MultiPolygon.

}
}
}

\keyword{datasets}
