% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NamedArea.r
\docType{class}
\name{NamedArea}
\alias{NamedArea}
\title{NamedArea Class}
\format{R6 class}
\usage{
# NamedArea$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for NamedArea objects.
}
\section{Fields}{

\describe{
\item{\code{areaClass}}{character}

\item{\code{areaName}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor NamedArea object.

}
\item{\code{$fromList(NamedAreaList)}}{

  Create NamedArea object from list.

}

\item{\code{$toList()}}{

  Get list representation of NamedArea.

}
\item{\code{fromJSONString(NamedAreaJson)}}{

  Create NamedArea object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of NamedArea.

}
}
}

\keyword{datasets}
