% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NameResolutionQuerySpec.r
\docType{class}
\name{NameResolutionQuerySpec}
\alias{NameResolutionQuerySpec}
\title{NameResolutionQuerySpec Class}
\format{
R6 class

R6 class
}
\usage{
# NameResolutionQuerySpec$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for NameResolutionQuerySpec objects.
}
\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor NameResolutionQuerySpec object.

}
\item{\code{$fromList(NameResolutionQuerySpecList)}}{

  Create NameResolutionQuerySpec object from list.

}

\item{\code{$toList()}}{

  Get list representation of NameResolutionQuerySpec.

}
\item{\code{fromJSONString(NameResolutionQuerySpecJson)}}{

  Create NameResolutionQuerySpec object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of NameResolutionQuerySpec.

}
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{constantScore}}{logical}

\item{\code{fields}}{list(character)}

\item{\code{conditions}}{list(QueryCondition)}

\item{\code{logicalOperator}}{character}

\item{\code{sortFields}}{list(SortField)}

\item{\code{from}}{integer}

\item{\code{size}}{integer}

\item{\code{nameResolutionRequest}}{NameResolutionRequest}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NameResolutionQuerySpec-new}{\code{NameResolutionQuerySpec$new()}}
\item \href{#method-NameResolutionQuerySpec-toList}{\code{NameResolutionQuerySpec$toList()}}
\item \href{#method-NameResolutionQuerySpec-fromList}{\code{NameResolutionQuerySpec$fromList()}}
\item \href{#method-NameResolutionQuerySpec-toJSONString}{\code{NameResolutionQuerySpec$toJSONString()}}
\item \href{#method-NameResolutionQuerySpec-fromJSONString}{\code{NameResolutionQuerySpec$fromJSONString()}}
\item \href{#method-NameResolutionQuerySpec-print}{\code{NameResolutionQuerySpec$print()}}
\item \href{#method-NameResolutionQuerySpec-clone}{\code{NameResolutionQuerySpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NameResolutionQuerySpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-NameResolutionQuerySpec-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NameResolutionQuerySpec$new(
  constantScore,
  fields,
  conditions,
  logicalOperator,
  sortFields,
  from,
  size,
  nameResolutionRequest
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NameResolutionQuerySpec-toList"></a>}}
\if{latex}{\out{\hypertarget{method-NameResolutionQuerySpec-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NameResolutionQuerySpec$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NameResolutionQuerySpec-fromList"></a>}}
\if{latex}{\out{\hypertarget{method-NameResolutionQuerySpec-fromList}{}}}
\subsection{Method \code{fromList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NameResolutionQuerySpec$fromList(
  NameResolutionQuerySpecList,
  typeMapping = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NameResolutionQuerySpec-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-NameResolutionQuerySpec-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NameResolutionQuerySpec$toJSONString(pretty = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NameResolutionQuerySpec-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-NameResolutionQuerySpec-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NameResolutionQuerySpec$fromJSONString(
  NameResolutionQuerySpecJson,
  typeMapping = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NameResolutionQuerySpec-print"></a>}}
\if{latex}{\out{\hypertarget{method-NameResolutionQuerySpec-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NameResolutionQuerySpec$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NameResolutionQuerySpec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NameResolutionQuerySpec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NameResolutionQuerySpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
