% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NameResolutionRequest.r
\docType{class}
\name{NameResolutionRequest}
\alias{NameResolutionRequest}
\title{NameResolutionRequest Class}
\format{
R6 class

R6 class
}
\usage{
# NameResolutionRequest$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for NameResolutionRequest objects.
}
\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor NameResolutionRequest object.

}
\item{\code{$fromList(NameResolutionRequestList)}}{

  Create NameResolutionRequest object from list.

}

\item{\code{$toList()}}{

  Get list representation of NameResolutionRequest.

}
\item{\code{fromJSONString(NameResolutionRequestJson)}}{

  Create NameResolutionRequest object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of NameResolutionRequest.

}
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{useCoL}}{logical}

\item{\code{searchString}}{character}

\item{\code{searchType}}{character}

\item{\code{nameTypes}}{list(character)}

\item{\code{fuzzyMatching}}{logical}

\item{\code{from}}{integer}

\item{\code{size}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NameResolutionRequest-new}{\code{NameResolutionRequest$new()}}
\item \href{#method-NameResolutionRequest-toList}{\code{NameResolutionRequest$toList()}}
\item \href{#method-NameResolutionRequest-fromList}{\code{NameResolutionRequest$fromList()}}
\item \href{#method-NameResolutionRequest-toJSONString}{\code{NameResolutionRequest$toJSONString()}}
\item \href{#method-NameResolutionRequest-fromJSONString}{\code{NameResolutionRequest$fromJSONString()}}
\item \href{#method-NameResolutionRequest-print}{\code{NameResolutionRequest$print()}}
\item \href{#method-NameResolutionRequest-clone}{\code{NameResolutionRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NameResolutionRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-NameResolutionRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NameResolutionRequest$new(
  useCoL,
  searchString,
  searchType,
  nameTypes,
  fuzzyMatching,
  from,
  size
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NameResolutionRequest-toList"></a>}}
\if{latex}{\out{\hypertarget{method-NameResolutionRequest-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NameResolutionRequest$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NameResolutionRequest-fromList"></a>}}
\if{latex}{\out{\hypertarget{method-NameResolutionRequest-fromList}{}}}
\subsection{Method \code{fromList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NameResolutionRequest$fromList(NameResolutionRequestList, typeMapping = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NameResolutionRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-NameResolutionRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NameResolutionRequest$toJSONString(pretty = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NameResolutionRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-NameResolutionRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NameResolutionRequest$fromJSONString(
  NameResolutionRequestJson,
  typeMapping = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NameResolutionRequest-print"></a>}}
\if{latex}{\out{\hypertarget{method-NameResolutionRequest-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NameResolutionRequest$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NameResolutionRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NameResolutionRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NameResolutionRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
