% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBA_categories}
\alias{NBA_categories}
\title{NBA category words/ phrases used in threat status/ protection level/ and condition
categorisation}
\format{
\subsection{\code{NBA_categories}}{

A vector of names:
}
}
\source{
SANBI
}
\usage{
NBA_categories
}
\description{
This data is used to enable users to verify that they are using the
standard format for their headings.
}
\keyword{datasets}
