% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBA_colours}
\alias{NBA_colours}
\title{NBA colours for every category}
\format{
\subsection{\code{NBA_colours}}{

A vector of category colours:
}
}
\source{
SANBI
}
\usage{
NBA_colours
}
\description{
This data is used to enable users to use the standard colours for their figures
}
\keyword{datasets}
