% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBA_example_bar_plot}
\alias{NBA_example_bar_plot}
\title{NBA ecosystem type threat status data}
\format{
\subsection{\code{NBA_example_bar_plot}}{

A data frame with 8 rows and 6 columns:
\describe{
\item{OVERALL types}{Ecosystem functional group name}
\item{Critically Endangered,Endangered,Vulnerable,Least Concern}{Threat status}
\item{TOT}{Total number of ecosystems}
...
}
}
}
\source{
SANBI
}
\usage{
NBA_example_bar_plot
}
\description{
This data is used to make a bar plot of threat status of ecosystems within a functional group.
This dataset has a totals column, but this is not necessary.
}
\keyword{datasets}
