% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBA_example_bubble_data}
\alias{NBA_example_bubble_data}
\title{NBA bubble example data}
\format{
\subsection{\code{NBA_example_bubble_data}}{

A data frame with 13 rows and 3 columns:
\describe{
\item{taxon_group}{The names of the taxon groups the species fall into}
\item{pressure}{The pressures that could impact the taxa}
\item{sub_pressure}{The specific sub pressures within the pressure categories}
\item{perc_concern_under_press}{The percentage of species of ecological concern that are impacted by the pressure}
...
}
}
}
\source{
SANBI
}
\usage{
NBA_example_bubble_data
}
\description{
This data is used to make a bubble plot of the percentage of species of ecological concern
in a group that are affected by a pressure
}
\keyword{datasets}
