% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBA_example_comb_data}
\alias{NBA_example_comb_data}
\title{NBA combined ecosystem type protection level and threat level example data}
\format{
\subsection{\code{NBA_example_comb_data}}{

A data frame with 16 rows and 11 columns:
\describe{
\item{OVERALL types}{Ecosystem functional group name}
\item{Well Protected,Moderately Protected,Poorly Protected,Not Protected}{Protection levels}
\item{Critically Endangered,Endangered,Vulnerable,Least Concern}{Threat status}
\item{TOT}{Total number of ecosystems}
\item{metric}{Column to differentiate rows into groups}
...
}
}
}
\source{
SANBI
}
\usage{
NBA_example_comb_data
}
\description{
This data is used to make a combined plot of protection level and threat status of ecosystems
within a functional group.
}
\keyword{datasets}
