% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBA_example_epli2018_data}
\alias{NBA_example_epli2018_data}
\title{NBA index data (EPLI 2018)}
\format{
A data frame with 11 rows and 4 variables:
\describe{
\item{X}{row number}
\item{T_BIOME}{biome name}
\item{Total}{total number}
\item{EPLI}{EPLI}
...
}
}
\source{
SANBI
}
\usage{
NBA_example_epli2018_data
}
\description{
This data is used to make a line plot of the ecosystem protection level index across a number of years
}
\keyword{datasets}
