% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBA_example_map_data}
\alias{NBA_example_map_data}
\title{NBA map example data, mem (marine ecosystem map) 2023}
\format{
\subsection{\code{NBA_example_map_data}}{

A data frame with 13 rows and 3 columns:
\describe{
\item{P_EcosysType}{Ecosystem name}
\item{geometry}{polygons}
\item{protection_level}{protection_level}
...
}
}
}
\source{
SANBI
}
\usage{
NBA_example_map_data
}
\description{
This data is used to make a map of marine pelagic ecosystems and their protection level
}
\keyword{datasets}
