% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBA_example_press_bar_data}
\alias{NBA_example_press_bar_data}
\title{NBA pressure bar plot example data}
\format{
\code{NBA_example_press_bar_data}
A data frame with 3 variables:
\describe{
\item{Taxon}{Taxonomic group (e.g., Birds, Mammals, Plants)}
\item{pressure}{Name of the pressure category}
\item{percentage}{Percentage value for each pressure}
...
}
}
\source{
SANBI
}
\usage{
NBA_example_press_bar_data
}
\description{
A small example dataset used to demonstrate the \code{nba_pressure_bar_plot()} function.
}
\keyword{datasets}
