% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBA_example_pro_data}
\alias{NBA_example_pro_data}
\title{NBA ecosystem type protection level data}
\format{
\subsection{\code{NBA_example_pro_data}}{

A data frame with 8 rows and 5 columns:
\describe{
\item{OVERALL types}{Ecosystem functional group name}
\item{Well Protected,Moderately Protected,Poorly Protected,Not Protected}{Protection levels}
...
}
}
}
\source{
SANBI
}
\usage{
NBA_example_pro_data
}
\description{
This data is used to make a donut plot of protection level of ecosystems
within a functional group.
}
\keyword{datasets}
