% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBA_example_RLI_data}
\alias{NBA_example_RLI_data}
\title{NBA Redlist index data}
\format{
\subsection{\code{NBA_example_RLI_data}}{

A data frame with 26 rows and 4 columns:
\describe{
\item{Years}{years}
\item{min}{Minimum value}
\item{RLI}{Redlist index}
\item{max}{Maximum value}
...
}
}
}
\source{
SANBI
}
\usage{
NBA_example_RLI_data
}
\description{
This data is used to make a line plot of the redlist index across a number of years
}
\keyword{datasets}
