% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{NBA_gt_theme}
\alias{NBA_gt_theme}
\title{gt table theme}
\usage{
NBA_gt_theme(GT_TBL)
}
\arguments{
\item{GT_TBL}{The data frame that contains the data}
}
\description{
Function to apply gt styling to table
}
\examples{
#test <- NBA_example_bar_plot |>
#gt::gt() |>
#NBA.package::NBA_gt_theme()

}
