% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{nba_index_plot}
\alias{nba_index_plot}
\title{NBA index plot function}
\usage{
nba_index_plot(
  TYPE = c("RLIs", "RLIe", "EPLI"),
  DF = NULL,
  YEAR = NULL,
  RLI = NULL,
  MIN = NULL,
  MAX = NULL,
  ASSESSMENT_YEAR = NULL,
  GROUP = NULL,
  RLIE = NULL,
  BIOME = NULL,
  EPLI_list = NULL,
  PALETTE = "taxon",
  AGGREGATE = TRUE,
  SAVE = NULL
)
}
\arguments{
\item{TYPE}{Type of index to plot. Options: \code{"RLIs"}, \code{"RLIe"}, or \code{"EPLI"}.}

\item{DF}{Input dataset containing index data (depending on TYPE).}

\item{YEAR}{Column name for year variable.}

\item{RLI}{Column name for RLI (for TYPE = "RLIs").}

\item{MIN}{Column name for lower bound (for TYPE = "RLIs").}

\item{MAX}{Column name for upper bound (for TYPE = "RLIs").}

\item{ASSESSMENT_YEAR}{Year to highlight assessment points (optional).}

\item{GROUP}{Column name for taxon or biome grouping (for TYPE = "RLIs").}

\item{RLIE}{Column name for ecosystem RLI (for TYPE = "RLIe").}

\item{BIOME}{Column name for biome (for TYPE = "RLIe" or "EPLI").}

\item{EPLI_list}{List of EPLI datasets keyed by year (for TYPE = "EPLI").}

\item{PALETTE}{Choose color palette \code{"taxon"} or \code{"biome"}.}

\item{AGGREGATE}{Whether to include an aggregate line. Default = TRUE.}

\item{SAVE}{Optional filename to save the plot PNG in \verb{outputs/}.}
}
\value{
Returns a multiple line plot with points
}
\description{
Generates the Red List Index (RLI) plots for species and ecosystems, or the Ecosystem  Protection Level Index (EPLI).
Supports multiple index types: RLIs (species), RLIe (ecosystems), and EPLI (protection level).
}
\examples{
# Example 1: RLIs grouped by Biome
nba_index_plot(
  TYPE = "RLIs",
  DF = NBA_rlis_biome_example_data,
  YEAR = Year,
  RLI = RLI,
  ASSESSMENT_YEAR = Assessment_Year,
  GROUP = Biome,
  PALETTE = "biome",
  AGGREGATE = TRUE,
  SAVE = NULL
)

# Example 2: RLIs grouped by Taxon
nba_index_plot(
  TYPE = "RLIs",
  DF = NBA_rlis_example_data,
  YEAR = Year,
  RLI = RLI,
  ASSESSMENT_YEAR = Assessment_Year,
  GROUP = Taxon,
  PALETTE = "taxon",
  AGGREGATE = TRUE,
  SAVE = NULL
)

# Example 3: RLIe
nba_index_plot(
  TYPE = "RLIe",
  DF = NBA_rlie_example_data,
  YEAR = "Year",
  RLIE = "RLIE",
  BIOME = "Biome",
  PALETTE = "biome",
  SAVE = NULL
)

# Example 4: EPLI
nba_index_plot(
  TYPE = "EPLI",
  EPLI_list = list(
    "2018" = NBA_epli2018_example_data,
    "2024" = NBA_epli2024_example_data
  ),
  PALETTE = "biome",
  SAVE = NULL
)


}
