% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{nba_init_quarto_docs}
\alias{nba_init_quarto_docs}
\title{Create template quarto documents to get an NBA html page up and running}
\usage{
nba_init_quarto_docs(
  path = ".",
  overwrite = FALSE,
  files = c("_brand.yml", "_quarto.yml", "custom.scss", "sanbi.csl",
    "_title-meta-author.html", "title-metadata.html", "scientific.qmd", "basic.qmd",
    "references.bib"),
  rename = NULL
)
}
\arguments{
\item{path}{The destination directory. Defaults to current working directory. Just add the name of any folder you want the documents to be cpoied into as long as it is in your project folder.}

\item{overwrite}{Whether to overwrite existing files. Defaults to FALSE.}

\item{files}{Character vector of files to copy. Defaults to all template files.}

\item{rename}{Named character vector to rename files on copy, e.g., c("scientific.qmd" = "index.qmd")}
}
\value{
Invisibly returns the paths of copied files.
}
\description{
Copies a template.qmd, template.scss, and _quarto.yml into a Quarto project directory.
}
