% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{nba_map}
\alias{nba_map}
\title{NBA map function}
\usage{
nba_map(DF, COLS = NULL, GEOM, CAP, FILL)
}
\arguments{
\item{DF}{The data frame that contains the information}

\item{COLS}{The variables to group together with the ecosystem types if the data isn't already in the form of one row per ecosystem type.}

\item{GEOM}{The name of the geometry column}

\item{CAP}{The caption for the map (will be place at the bottom left of the plot)}

\item{FILL}{The column that contains the categories to colour the ecosystems by (e.g. protection level, threat status, or condition)}
}
\value{
Returns a map
}
\description{
NBA maps for protection level, threat status, and condition.
The function can take a dataset with either one polygon or
multipolygon per ecosystem (each row has a unique ecosystem) or
multiple rows per ecosystem type.
}
\details{
The function expects an sf object.

Please look at the example data NBA_example_map_data and NBA_example_donut_plot to see the correct
structure for the data.

The name of the fill column is irrelevant, but the categories must be
spelled correctly (there is a list of the standard spellings/ cases
of NBA categories named NBA_categories in this package,
which can be accessed for reference).
}
\examples{

map <- nba_map(DF = NBA_example_map_data,
              GEOM = geometry,
              CAP = "Figure 1. A map of the protection level of marine pelagic ecosystem types in South Africa",
              FILL = protection_level)

map

#map


}
