% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{NBA_plot}
\alias{NBA_plot}
\title{NBA plot function}
\usage{
NBA_plot(
  DF,
  GROUPS,
  COLS,
  CHRT = c("bar", "donut"),
  NUM = FALSE,
  LAB,
  GRP = TRUE,
  SAVE
)
}
\arguments{
\item{DF}{The data frame that contains the information on protection level}

\item{GROUPS}{The variables to graph (ecosystem type, taxa, etc)}

\item{COLS}{The categories to describe the groups (protection level, threat status, condition, etc)}

\item{CHRT}{A choice of either "bar" or "donut" plot}

\item{NUM}{A choice to show numbers in the plot, Flase for no numbers}

\item{LAB}{The x axis label of the plot}

\item{GRP}{A choice of whether or not to plot the donut graphs by group, TRUE will plot a donut plot for each group.}

\item{SAVE}{The name of the output file that will be saved to the output folder. If you do not have an outputs folder you will be prompted to make one.}
}
\value{
Returns a plot
}
\description{
NBA plots for protection level, threat status, and condition.
}
\examples{
#test <- NBA_plot(Fig1a,OVERALL_types,2:5,CHRT = "bar",NUM = TRUE,LAB = "Percentage of ecosystem types",SAVE = "Fig1a")

}
