% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{nba_plot_bubble}
\alias{nba_plot_bubble}
\title{Bubble plot}
\usage{
nba_plot_bubble(DF, GROUP, CAT, SUB_CAT, VALUE, SAVE = NULL)
}
\arguments{
\item{DF}{The data frame that contains the information}

\item{GROUP}{The grouping variables (taxa group, etc)}

\item{CAT}{The overall pressure category}

\item{SUB_CAT}{The sub pressure category}

\item{VALUE}{The percentage of taxa of ecological concern impacted by the pressure}

\item{SAVE}{The name of the output file that will be saved to the output folder. If you do not have an outputs folder you will be prompted to make one.}
}
\value{
Returns a bubble plot
}
\description{
This function will create bubble plot intended for creation of the
percentage of taxa of ecological concern impacted by various pressures
}
\details{
Please use the example dataset NBA_bubble_plot_example_data to see how
the data should be formatted.
}
\examples{
bubble_plot <- nba_plot_bubble(DF = NBA_example_bubble_data,
                        GROUP = taxon_group,
                       CAT = pressure,
                       SUB_CAT = sub_pressure,
                       VALUE = perc_concern_under_press,
                        SAVE = NULL)

bubble_plot

}
