% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{NBA_plot_RLI}
\alias{NBA_plot_RLI}
\title{NBA Redlist index plot function}
\usage{
NBA_plot_RLI(DF, YEAR, RLI, min, max, GRP = FALSE, SAVE = NULL)
}
\arguments{
\item{DF}{The data frame that contains the information on the Red List Index}

\item{YEAR}{The years}

\item{RLI}{The Red List Index}

\item{min}{The minimum values}

\item{max}{The maximum values}

\item{GRP}{A choice to group the plot, TRUE will group if, FALSE will not}

\item{SAVE}{The name of the output file that will be saved to the output folder. If you do not have an outputs folder you will be prompted to make one.}
}
\value{
Returns a RLI plot
}
\description{
This function will create a line graph indicating the Red list index per period/year.
Additionally, depending on the graph requirements it further creates a multiple line graph
grouping the plot according to species, and creates a buffer parallel to the RLI
line using the minimum and maximum values.
}
\details{
Note: The graph can be used on other datasets as well, as long as it contains the same
data structure as the Red List Index.
There is an example dataset (NBA_example_RLI_data), available for users
to reference if needed.
}
\examples{
RLI_plot <- NBA_plot_RLI(NBA_example_RLI_data,
Years,
RLI,
min,
max)

RLI_plot


}
