% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{nba_plot_theme}
\alias{nba_plot_theme}
\title{NBA theme for a ggplot object}
\usage{
nba_plot_theme()
}
\value{
Returns a ggplot object with the NBA_theme styling
}
\description{
This function will create a theme for all ggplot graphs to align them to the
NBA "look". All NBA functions already use this theme.
}
\examples{
#library(ggplot2)

#gg_plot <- ggplot(NBA_example_pro_data, aes(x = `OVERALL types`, y = `Well Protected`))+
#  ggplot2::geom_point()+
#  nba_plot_theme()

#gg_plot

}
