% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{nba_pressure_bar_plot}
\alias{nba_pressure_bar_plot}
\title{horizontal bar plot of pressures}
\usage{
nba_pressure_bar_plot(DF, TAXON = NULL, SAVE = NULL)
}
\arguments{
\item{DF}{A data frame containing at least the columns \code{Taxon}, \code{pressure}, and \code{percentage}.}

\item{TAXON}{The name of the taxon to filter for.}

\item{SAVE}{The name of the output file that will be saved to the output folder. If you do not have an outputs folder you will be prompted to make one.}
}
\value{
A horizontal bar plot.
}
\description{
This function generates a horizontal bar plot showing the percentage of pressures
affecting taxa.
}
\examples{

birds_press_bar_plot <- nba_pressure_bar_plot(
                          DF = NBA_example_press_bar_data,
                          TAXON = "Birds")

birds_press_bar_plot

}
