% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBA_rlie_example_data}
\alias{NBA_rlie_example_data}
\title{NBA index data (RLIe)}
\format{
A data frame with 44 rows and 8 variables:
\describe{
\item{Year}{assessment year}
\item{RLIE}{Red List Index value}
\item{lower}{minimum value}
\item{upper}{maximum value}
\item{criteria}{Overall}
\item{total_weight}{weight}
\item{total_count}{count}
\item{Biome}{biome name}
...
}
}
\source{
SANBI
}
\usage{
NBA_rlie_example_data
}
\description{
This data is used to make a line plot of the red list index across a number of years
}
\keyword{datasets}
