% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBA_rlis_example_data}
\alias{NBA_rlis_example_data}
\title{NBA index data (RLIs taxon)}
\format{
A data frame with 63 rows and 6 variables:
\describe{
\item{Year}{assessment year}
\item{RLI}{Red List Index value}
\item{min}{minimum value}
\item{max}{maximum value}
\item{Assessment_Year}{year of assessment}
\item{Taxon}{taxon name}
...
}
}
\source{
SANBI
}
\usage{
NBA_rlis_example_data
}
\description{
This data is used to make a line plot of the red list index across a number of years
}
\keyword{datasets}
