% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{nba_tbl}
\alias{nba_tbl}
\title{Basic NBA table}
\usage{
nba_tbl(DF)
}
\arguments{
\item{DF}{The data frame that contains the data}
}
\description{
A basic table with a purple heading bar.
This data should be in the same format as for the plot functions
}
\examples{
tbl <- nba_tbl(NBA_example_pro_data)

tbl

}
