% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{nba_tbl_colr}
\alias{nba_tbl_colr}
\title{Coloured table}
\usage{
nba_tbl_colr(
  DF,
  COL,
  HEADER = c("sanbi-green", "sanbi-orange", "sanbi-purple", "Freshwater", "Marine",
    "Coast", "Estuarine", "Terrestrial", "Genetics", "PEI")
)
}
\arguments{
\item{DF}{The data frame that contains the data}

\item{COL}{the column containing the categories}

\item{HEADER}{the name to determine the colour of the header row}
}
\description{
A basic table with a coloured heading bar and coloured blocks around the
categories
This data should be in the same format as for the plot functions.
}
\details{
This table has the same styling as set out in nba_tbl_theme. If you would like
to change anything you can just use a pipe \%>\%  to add gt styling onto the gt
object this function creates and it will override the styling set in the function.
}
\examples{
library(dplyr)
library(tidyr)

tbl <- NBA_example_pro_data \%>\%
 pivot_longer(2:5, names_to = "protection_level") \%>\%
 nba_tbl_colr(COL = protection_level, HEADER = "Coast")

tbl

}
