% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{nba_tbl_comb}
\alias{nba_tbl_comb}
\title{Table of protection level of threatened ecosystems.}
\usage{
nba_tbl_comb(DF, GROUP, THR, PRO, FILE = c("spatial", "csv"))
}
\arguments{
\item{DF}{The data frame that contains the data}

\item{GROUP}{The column that contains the name of the variable (the ecosystems or taxa names)}

\item{THR}{The column name of the threat statuses}

\item{PRO}{The column name of the protection levels}

\item{FILE}{An indication if the input file is a map (spatial file with a geom column) or a csv/ normal dataframe. If it is a normal dataframe the data must already be in the correct format as the function only formats data in a spatial file.}
}
\description{
Function to style a table with threat status on the horizontal rows
and protection level on the vertical columns, with number of ecosystems
or taxa that share those categories, with total and percentage çolumns added
}
\details{
This function is special because if you start from a spatial file you will need
to add the group, thr and pro, and file arguments, but if you start from a
dataframe (csv) you only need the DF, thr, and file arguments.
}
\examples{

thr_pro_tbl <- NBA_example_map_data |>
nba_tbl_comb(GROUP = P_EcosysType,
THR = threat_status,
PRO = protection_level,
FILE = "spatial")

thr_pro_tbl

}
