% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{nba_tbl_theme}
\alias{nba_tbl_theme}
\title{gt table theme}
\usage{
nba_tbl_theme(
  GT_TBL,
  HEADER = c("sanbi-green", "sanbi-orange", "sanbi-purple", "Freshwater", "Marine",
    "Coast", "Estuarine", "Terrestrial", "Genetics", "PEI")
)
}
\arguments{
\item{GT_TBL}{The data frame that contains the data}

\item{HEADER}{the name to determine the colour of the header row. Defaults to SANBI green}
}
\description{
Function to apply gt styling to table
}
\examples{
library(gt)
library(dplyr)


gt_tbl <- NBA_example_thr_data \%>\%
 gt() \%>\%
nba_tbl_theme()

gt_tbl

}
