% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaR-package.r
\docType{package}
\name{nbaR}
\alias{nbaR}
\alias{nbaR-package}
\title{nbaR : R Package Client for the Netherlands Biodiversity API}
\description{
nbaR : R Package Client for the Netherlands Biodiversity API
}
\details{
Access to the digitised Natural History collection at the Naturalis Biodiversity Center
}
\author{
Hannes  \email{ hannes.hettling@naturalis.nl }
}
\keyword{package}
