% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Organization.r
\docType{data}
\name{Organization}
\alias{Organization}
\title{Organization Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Organization
}
\description{
Organization Class
}
\section{Fields}{

\describe{
\item{\code{agentText}}{}

\item{\code{name}}{}
}}

\keyword{datasets}
