% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Person.r
\docType{data}
\name{Person}
\alias{Person}
\title{Person Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Person
}
\description{
Person Class
}
\section{Fields}{

\describe{
\item{\code{agentText}}{}

\item{\code{fullName}}{}

\item{\code{organization}}{}
}}

\keyword{datasets}
