% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Point.r
\docType{data}
\name{Point}
\alias{Point}
\title{Point Class}
\format{}
\usage{

}
\description{
Point Class
}
\section{Fields}{

\describe{
\item{\code{crs}}{Crs}

\item{\code{bbox}}{list(numeric)}

\item{\code{coordinates}}{LngLatAlt}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Point object.

}
\item{\code{$fromList(PointList)}}{

  Create Point object from list.

}

\item{\code{$toList()}}{

  Get list representation of Point.

}
\item{\code{fromJSONString(PointJson)}}{

  Create Point object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Point.

}
}
}

\keyword{datasets}
