% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polygon.r
\docType{class}
\name{Polygon}
\alias{Polygon}
\title{Polygon Class}
\format{R6 class}
\usage{
# Polygon$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for Polygon objects.
}
\section{Fields}{

\describe{
\item{\code{crs}}{Crs}

\item{\code{bbox}}{list(numeric)}

\item{\code{coordinates}}{list(LngLatAlt)}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Polygon object.

}
\item{\code{$fromList(PolygonList)}}{

  Create Polygon object from list.

}

\item{\code{$toList()}}{

  Get list representation of Polygon.

}
\item{\code{fromJSONString(PolygonJson)}}{

  Create Polygon object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Polygon.

}
}
}

\keyword{datasets}
