% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{pro_donut_plot}
\alias{pro_donut_plot}
\title{Protection level donut plot}
\usage{
pro_donut_plot(DF, YMAX, YMIN, FILL, COLOUR, LABEL)
}
\arguments{
\item{DF}{The data frame that contains the information on protection level}

\item{YMAX}{The protection level percentages}

\item{YMIN}{The groups}

\item{FILL}{The protection level categories}

\item{COLOUR}{colour for the protection level}
}
\value{
Returns a bar donut plot of protection level
}
\description{
Donut plot for the protection level of the ecosystem types
}
\examples{
#test <- pro_donut_plot(mydata, ecosystem_functional_grps, percentages, threat_status)

}
