% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{pro_donut_plot_test}
\alias{pro_donut_plot_test}
\title{Protection level donut plot}
\usage{
pro_donut_plot_test(DF, FILL)
}
\arguments{
\item{DF}{The data frame that contains the information on protection level}

\item{FILL}{The protection level categories}
}
\value{
Returns a bar donut plot of protection level
}
\description{
Donut plot for the protection level of the ecosystem types
}
\examples{
#test <- pro_donut_plot(mydata, ecosystem_functional_grps)

}
