% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{prot_efg}
\alias{prot_efg}
\title{Protection level bar plot}
\usage{
prot_efg(DF, X, Y, FILL, COUNT)
}
\arguments{
\item{DF}{The data frame that contains the information on protection level}

\item{X}{The groups}

\item{Y}{The protection level percentages}

\item{FILL}{The protection level categories}

\item{COUNT}{The frequency counts of the number of ecosystems within each protection level}
}
\value{
Returns a bar graph of protection level per functional group or taxa
}
\description{
Horizontal barplots for the protection level.
This function requires a dataframe that contains a columns of the
the ecosystem functional groups or taxa, the protection level,
and the number and percentage of ecosystems or species that fall within that
protection level.
}
\details{
The colours and axis titels are created within the function.
}
\examples{
#test <- prot_efg(mydata, ecosystem_functional_grps, percentages, protection level, number_of_ecosystems)

}
