% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryCondition.r
\docType{data}
\name{QueryCondition}
\alias{QueryCondition}
\title{QueryCondition Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
QueryCondition
}
\description{
QueryCondition Class
}
\section{Fields}{

\describe{
\item{\code{not}}{}

\item{\code{field}}{}

\item{\code{operator}}{}

\item{\code{value}}{}

\item{\code{and}}{}

\item{\code{or}}{}

\item{\code{constantScore}}{}

\item{\code{boost}}{}
}}

\keyword{datasets}
