% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryResult.r
\docType{data}
\name{QueryResult}
\alias{QueryResult}
\title{QueryResult Class}
\format{}
\usage{

}
\description{
QueryResult Class
}
\section{Fields}{

\describe{
\item{\code{totalSize}}{integer}

\item{\code{resultSet}}{list(QueryResultItemObject)}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor QueryResult object.

}
\item{\code{$fromList(QueryResultList)}}{

  Create QueryResult object from list.

}

\item{\code{$toList()}}{

  Get list representation of QueryResult.

}
\item{\code{fromJSONString(QueryResultJson)}}{

  Create QueryResult object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of QueryResult.

}
}
}

\keyword{datasets}
