% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryResultItemObject.r
\docType{data}
\name{QueryResultItemObject}
\alias{QueryResultItemObject}
\title{QueryResultItemObject Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
QueryResultItemObject
}
\description{
Contains an object returned from a QuerySpec query. Holds an item and a relevance score.
}
\section{Fields}{

\describe{
\item{\code{score}}{numeric}

\item{\code{item}}{list}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor QueryResultItemObject object.

}
\item{\code{$fromList(QueryResultItemObjectList)}}{

  Create QueryResultItemObject object from list.

}

\item{\code{$toList()}}{

  Get list representation of QueryResultItemObject.

}
\item{\code{fromJSONString(QueryResultItemObjectJson)}}{

  Create QueryResultItemObject object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of QueryResultItemObject.

}
}
}

\keyword{datasets}
