% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reference.r
\docType{class}
\name{Reference}
\alias{Reference}
\title{Reference Class}
\format{R6 class}
\usage{
# Reference$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for Reference objects.
}
\section{Fields}{

\describe{
\item{\code{titleCitation}}{character}

\item{\code{citationDetail}}{character}

\item{\code{uri}}{character}

\item{\code{author}}{Person}

\item{\code{publicationDate}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Reference object.

}
\item{\code{$fromList(ReferenceList)}}{

  Create Reference object from list.

}

\item{\code{$toList()}}{

  Get list representation of Reference.

}
\item{\code{fromJSONString(ReferenceJson)}}{

  Create Reference object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Reference.

}
}
}

\keyword{datasets}
