% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestService.r
\docType{data}
\name{RestService}
\alias{RestService}
\title{RestService Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RestService
}
\description{
RestService Class
}
\section{Fields}{

\describe{
\item{\code{url}}{}
}}

\keyword{datasets}
