% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{RLI_plot}
\alias{RLI_plot}
\title{RLI_plot}
\usage{
RLI_plot(DF, YEAR, RLI, min, max, GRP = FALSE)
}
\arguments{
\item{DF}{The data frame that contains the information on protection level}

\item{YEAR}{The protection level categories}

\item{RLI}{The protection level categories}

\item{min}{The protection level categories}

\item{max}{The protection level categories}

\item{GRP}{The protection level categories}
}
\value{
Returns a RLI plot
}
\description{
RLI_plot
}
\examples{
#test <- pro_donut_plot(mydata, ecosystem_functional_grps)

}
