% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificName.r
\docType{data}
\name{ScientificName}
\alias{ScientificName}
\title{ScientificName Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ScientificName
}
\description{
ScientificName Class
}
\section{Fields}{

\describe{
\item{\code{fullScientificName}}{}

\item{\code{taxonomicStatus}}{}

\item{\code{genusOrMonomial}}{}

\item{\code{subgenus}}{}

\item{\code{specificEpithet}}{}

\item{\code{infraspecificEpithet}}{}

\item{\code{infraspecificMarker}}{}

\item{\code{nameAddendum}}{}

\item{\code{authorshipVerbatim}}{}

\item{\code{author}}{}

\item{\code{year}}{}

\item{\code{scientificNameGroup}}{}

\item{\code{references}}{}

\item{\code{experts}}{}
}}

\keyword{datasets}
