% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServiceAccessPoint.r
\docType{data}
\name{ServiceAccessPoint}
\alias{ServiceAccessPoint}
\title{ServiceAccessPoint Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ServiceAccessPoint
}
\description{
ServiceAccessPoint Class
}
\section{Fields}{

\describe{
\item{\code{variant}}{}
}}

\keyword{datasets}
