% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shark_tree}
\alias{shark_tree}
\title{A molecular phylogeny of the sharks}
\format{An object of class \code{phylo}}
\usage{
shark_tree
}
\description{
An (uncalibrated) phylogenetic tree with 231 species from the eight
extant shark orders. Number of edges: 463. Branch lengths correspond
to molecular distances. Tips are labelled as follows: Genus_species.
Can be processed using the package \code{ape}. The data were kindly
provided by Ximena Vélez-Zuazo.
}
\examples{
data(shark_tree)
plot(shark_tree, cex=0.1)
}
\references{
Vélez-Zuazo & Agnarsson. Shark tales: A molecular species-level
phylogeny of sharks (Selachimorpha, Chondrichthyes).
Molecular Phylogenetics and Evolution 58:2, 207-217
(\href{http://www.sciencedirect.com/science/article/pii/S1055790310004537}{Link})
}
\keyword{datasets}
\keyword{phylogeny}
\keyword{sharks}
