% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNKCJE.R
\docType{data}
\name{shotchart}
\alias{shotchart}
\title{Download data from NBA.com/Stats/ and build your own database in local PostgreSQL instance}
\format{An object of class \code{NULL} of length 0.}
\usage{
shotchart
}
\description{
Download data from NBA.com/Stats/ and build your own database in local PostgreSQL instance
}
\keyword{datasets}

