% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceSystem.r
\docType{class}
\name{SourceSystem}
\alias{SourceSystem}
\title{SourceSystem Class}
\format{R6 class}
\usage{
# SourceSystem$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for SourceSystem objects.
}
\section{Fields}{

\describe{
\item{\code{code}}{character}

\item{\code{name}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor SourceSystem object.

}
\item{\code{$fromList(SourceSystemList)}}{

  Create SourceSystem object from list.

}

\item{\code{$toList()}}{

  Get list representation of SourceSystem.

}
\item{\code{fromJSONString(SourceSystemJson)}}{

  Create SourceSystem object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of SourceSystem.

}
}
}

\keyword{datasets}
