% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Specimen.r
\docType{class}
\name{Specimen}
\alias{Specimen}
\title{Specimen Class}
\format{
R6 class

R6 class
}
\usage{
# Specimen$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for Specimen objects.
}
\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Specimen object.

}
\item{\code{$fromList(SpecimenList)}}{

  Create Specimen object from list.

}

\item{\code{$toList()}}{

  Get list representation of Specimen.

}
\item{\code{fromJSONString(SpecimenJson)}}{

  Create Specimen object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Specimen.

}
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sourceSystem}}{SourceSystem}

\item{\code{sourceSystemId}}{character}

\item{\code{recordURI}}{character}

\item{\code{id}}{character}

\item{\code{unitID}}{character}

\item{\code{unitGUID}}{character}

\item{\code{collectorsFieldNumber}}{character}

\item{\code{assemblageID}}{character}

\item{\code{sourceInstitutionID}}{character}

\item{\code{sourceID}}{character}

\item{\code{previousSourceID}}{list(character)}

\item{\code{owner}}{character}

\item{\code{licenseType}}{character}

\item{\code{license}}{character}

\item{\code{recordBasis}}{character}

\item{\code{kindOfUnit}}{character}

\item{\code{collectionType}}{character}

\item{\code{sex}}{character}

\item{\code{phaseOrStage}}{character}

\item{\code{title}}{character}

\item{\code{notes}}{character}

\item{\code{preparationType}}{character}

\item{\code{previousUnitsText}}{character}

\item{\code{numberOfSpecimen}}{integer}

\item{\code{fromCaptivity}}{logical}

\item{\code{objectPublic}}{logical}

\item{\code{multiMediaPublic}}{logical}

\item{\code{acquiredFrom}}{Agent}

\item{\code{gatheringEvent}}{GatheringEvent}

\item{\code{dynamicProperties}}{character}

\item{\code{informationWithheld}}{character}

\item{\code{dataGeneralizations}}{character}

\item{\code{identifications}}{list(SpecimenIdentification)}

\item{\code{associatedMultiMediaUris}}{list(ServiceAccessPoint)}

\item{\code{theme}}{list(character)}

\item{\code{associatedMultiMediaObjects}}{list(MultiMediaObject)}

\item{\code{dateLastEdited}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Specimen-new}{\code{Specimen$new()}}
\item \href{#method-Specimen-toList}{\code{Specimen$toList()}}
\item \href{#method-Specimen-fromList}{\code{Specimen$fromList()}}
\item \href{#method-Specimen-toJSONString}{\code{Specimen$toJSONString()}}
\item \href{#method-Specimen-fromJSONString}{\code{Specimen$fromJSONString()}}
\item \href{#method-Specimen-print}{\code{Specimen$print()}}
\item \href{#method-Specimen-clone}{\code{Specimen$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Specimen-new"></a>}}
\if{latex}{\out{\hypertarget{method-Specimen-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Specimen$new(
  sourceSystem,
  sourceSystemId,
  recordURI,
  id,
  unitID,
  unitGUID,
  collectorsFieldNumber,
  assemblageID,
  sourceInstitutionID,
  sourceID,
  previousSourceID,
  owner,
  licenseType,
  license,
  recordBasis,
  kindOfUnit,
  collectionType,
  sex,
  phaseOrStage,
  title,
  notes,
  preparationType,
  previousUnitsText,
  numberOfSpecimen,
  fromCaptivity,
  objectPublic,
  multiMediaPublic,
  acquiredFrom,
  gatheringEvent,
  dynamicProperties,
  informationWithheld,
  dataGeneralizations,
  identifications,
  associatedMultiMediaUris,
  theme,
  associatedMultiMediaObjects,
  dateLastEdited
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Specimen-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Specimen-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Specimen$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Specimen-fromList"></a>}}
\if{latex}{\out{\hypertarget{method-Specimen-fromList}{}}}
\subsection{Method \code{fromList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Specimen$fromList(SpecimenList, typeMapping = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Specimen-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Specimen-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Specimen$toJSONString(pretty = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Specimen-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Specimen-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Specimen$fromJSONString(SpecimenJson, typeMapping = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Specimen-print"></a>}}
\if{latex}{\out{\hypertarget{method-Specimen-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Specimen$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Specimen-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Specimen-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Specimen$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
