% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecimenClient-wrapper.r
\name{specimen_find_by_unit_id}
\alias{specimen_find_by_unit_id}
\title{Find a specimen by unitID}
\usage{
specimen_find_by_unit_id(unitID = NULL, ...)
}
\arguments{
\item{unitID}{the unitID of the specimen to query, type:}

\item{...}{additional parameters passed to find_by_unit_id from class class nbaR.SpecimenClient}
}
\value{
list
}
\description{
This is a wrapper function for find_by_unit_id
from class nbaR.SpecimenClient.
}
\details{
Get a specimen by its unitID. Returns a list of specimens since unitIDs are not strictly unique
}
\seealso{
Other nbaR.SpecimenClient-wrappers: \code{\link{specimen_count_distinct_values_per_group}},
  \code{\link{specimen_count_distinct_values}},
  \code{\link{specimen_count}},
  \code{\link{specimen_download_query}},
  \code{\link{specimen_dwca_get_data_set_names}},
  \code{\link{specimen_dwca_get_data_set}},
  \code{\link{specimen_dwca_query}},
  \code{\link{specimen_exists}},
  \code{\link{specimen_find_by_ids}},
  \code{\link{specimen_find}},
  \code{\link{specimen_get_distinct_values_per_group}},
  \code{\link{specimen_get_distinct_values}},
  \code{\link{specimen_get_field_info}},
  \code{\link{specimen_get_ids_in_collection}},
  \code{\link{specimen_get_named_collections}},
  \code{\link{specimen_get_paths}},
  \code{\link{specimen_get_settings}},
  \code{\link{specimen_get_setting}},
  \code{\link{specimen_group_by_scientific_name}},
  \code{\link{specimen_is_operator_allowed}},
  \code{\link{specimen_query}}
}
\concept{nbaR.SpecimenClient-wrappers}
