% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecimenClient-wrapper.r
\name{specimen_get_distinct_values_per_group}
\alias{specimen_get_distinct_values_per_group}
\title{Get all distinct values (and their document count) for the field given divided per distinct value of the field to group by}
\usage{
specimen_get_distinct_values_per_group(
  group = NULL,
  field = NULL,
  returnType = "data.frame",
  ...
)
}
\arguments{
\item{group}{Name of field in the specimen object you want to group by, type:}

\item{field}{Name of field in the specimen object, type:}

\item{returnType}{Either \code{list} or \code{data.frame} (default)}

\item{...}{additional parameters passed to get_distinct_values_per_group from class nbaR.SpecimenClient}
}
\value{
scalar
}
\description{
This is a wrapper for the method \code{ get_distinct_values_per_group }
from class \code{ SpecimenClient}.
}
\seealso{
Other nbaR.SpecimenClient-wrappers: 
\code{\link{specimen_count_distinct_values_per_group}()},
\code{\link{specimen_count_distinct_values}()},
\code{\link{specimen_count}()},
\code{\link{specimen_download_query}()},
\code{\link{specimen_dwca_get_data_set_names}()},
\code{\link{specimen_dwca_get_data_set}()},
\code{\link{specimen_dwca_query}()},
\code{\link{specimen_exists}()},
\code{\link{specimen_find_by_ids}()},
\code{\link{specimen_find_by_unit_id}()},
\code{\link{specimen_find}()},
\code{\link{specimen_get_distinct_values}()},
\code{\link{specimen_get_field_info}()},
\code{\link{specimen_get_paths}()},
\code{\link{specimen_get_settings}()},
\code{\link{specimen_group_by_scientific_name}()},
\code{\link{specimen_is_operator_allowed}()},
\code{\link{specimen_query_with_name_resolution}()},
\code{\link{specimen_query}()}
}
\concept{nbaR.SpecimenClient-wrappers}
