% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecimenClient-wrapper.r
\name{specimen_get_ids_in_collection}
\alias{specimen_get_ids_in_collection}
\title{Retrieve all ids within a &#39;special collection&#39; of specimens}
\usage{
specimen_get_ids_in_collection(name = NULL, returnType = "data.frame",
  ...)
}
\arguments{
\item{name}{name of dataset, type:}

\item{returnType}{Either 'list' or 'data.frame', defaults to 'data.frame'}

\item{...}{additional parameters passed to get_ids_in_collection from class nbaR.SpecimenClient}
}
\value{
scalar
}
\description{
This is a wrapper function for get_ids_in_collection
from class nbaR.SpecimenClient.
}
\details{
Available collections can be queried with /getNamedCollections
}
\seealso{
Other nbaR.SpecimenClient-wrappers: \code{\link{specimen_count_distinct_values_per_group}},
  \code{\link{specimen_count_distinct_values}},
  \code{\link{specimen_count}},
  \code{\link{specimen_download_query}},
  \code{\link{specimen_dwca_get_data_set_names}},
  \code{\link{specimen_dwca_get_data_set}},
  \code{\link{specimen_dwca_query}},
  \code{\link{specimen_exists}},
  \code{\link{specimen_find_by_ids}},
  \code{\link{specimen_find_by_unit_id}},
  \code{\link{specimen_find}},
  \code{\link{specimen_get_distinct_values_per_group}},
  \code{\link{specimen_get_distinct_values}},
  \code{\link{specimen_get_field_info}},
  \code{\link{specimen_get_named_collections}},
  \code{\link{specimen_get_paths}},
  \code{\link{specimen_get_settings}},
  \code{\link{specimen_get_setting}},
  \code{\link{specimen_group_by_scientific_name}},
  \code{\link{specimen_is_operator_allowed}},
  \code{\link{specimen_query}}
}
\concept{nbaR.SpecimenClient-wrappers}
