% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecimenClient-wrapper.r
\name{specimen_get_settings}
\alias{specimen_get_settings}
\title{List all publicly available configuration settings for the NBA}
\usage{
specimen_get_settings(name = NULL, ...)

specimen_get_settings(name = NULL, ...)
}
\arguments{
\item{name}{name of setting, type:}

\item{...}{additional parameters passed to get_settings from class nbaR.SpecimenClient}

\item{returnType}{Either \code{list} or \code{data.frame} (default)}
}
\value{
scalar

scalar
}
\description{
This is a wrapper for the method \code{ get_settings }
from class \code{ SpecimenClient}.

This is a wrapper for the method \code{ get_settings }
from class \code{ SpecimenClient}.
}
\details{
The value of a specific setting can be queried with metadata/getSetting/{name}

All settings can be queried with /metadata/getSettings
}
\seealso{
Other nbaR.SpecimenClient-wrappers: 
\code{\link{specimen_count_distinct_values_per_group}()},
\code{\link{specimen_count_distinct_values}()},
\code{\link{specimen_count}()},
\code{\link{specimen_download_query}()},
\code{\link{specimen_dwca_get_data_set_names}()},
\code{\link{specimen_dwca_get_data_set}()},
\code{\link{specimen_dwca_query}()},
\code{\link{specimen_exists}()},
\code{\link{specimen_find_by_ids}()},
\code{\link{specimen_find_by_unit_id}()},
\code{\link{specimen_find}()},
\code{\link{specimen_get_distinct_values_per_group}()},
\code{\link{specimen_get_distinct_values}()},
\code{\link{specimen_get_field_info}()},
\code{\link{specimen_get_paths}()},
\code{\link{specimen_group_by_scientific_name}()},
\code{\link{specimen_is_operator_allowed}()},
\code{\link{specimen_query_with_name_resolution}()},
\code{\link{specimen_query}()}

Other nbaR.SpecimenClient-wrappers: 
\code{\link{specimen_count_distinct_values_per_group}()},
\code{\link{specimen_count_distinct_values}()},
\code{\link{specimen_count}()},
\code{\link{specimen_download_query}()},
\code{\link{specimen_dwca_get_data_set_names}()},
\code{\link{specimen_dwca_get_data_set}()},
\code{\link{specimen_dwca_query}()},
\code{\link{specimen_exists}()},
\code{\link{specimen_find_by_ids}()},
\code{\link{specimen_find_by_unit_id}()},
\code{\link{specimen_find}()},
\code{\link{specimen_get_distinct_values_per_group}()},
\code{\link{specimen_get_distinct_values}()},
\code{\link{specimen_get_field_info}()},
\code{\link{specimen_get_paths}()},
\code{\link{specimen_group_by_scientific_name}()},
\code{\link{specimen_is_operator_allowed}()},
\code{\link{specimen_query_with_name_resolution}()},
\code{\link{specimen_query}()}
}
\concept{nbaR.SpecimenClient-wrappers}
