% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecimenClient-wrapper.r
\name{specimen_is_operator_allowed}
\alias{specimen_is_operator_allowed}
\title{Checks if a given operator is allowed for a given field}
\usage{
specimen_is_operator_allowed(
  field = NULL,
  operator = NULL,
  returnType = "data.frame",
  ...
)
}
\arguments{
\item{field}{specimen document field, type:}

\item{operator}{operator, type:sc <- SpecimenClient$new()}

\item{returnType}{Either \code{list} or \code{data.frame} (default)}

\item{...}{additional parameters passed to is_operator_allowed from class nbaR.SpecimenClient}
}
\value{
scalar
}
\description{
This is a wrapper for the method \code{ is_operator_allowed }
from class \code{ SpecimenClient}.
}
\details{
See also metadata/getFieldInfo
}
\seealso{
Other nbaR.SpecimenClient-wrappers: 
\code{\link{specimen_count_distinct_values_per_group}()},
\code{\link{specimen_count_distinct_values}()},
\code{\link{specimen_count}()},
\code{\link{specimen_download_query}()},
\code{\link{specimen_dwca_get_data_set_names}()},
\code{\link{specimen_dwca_get_data_set}()},
\code{\link{specimen_dwca_query}()},
\code{\link{specimen_exists}()},
\code{\link{specimen_find_by_ids}()},
\code{\link{specimen_find_by_unit_id}()},
\code{\link{specimen_find}()},
\code{\link{specimen_get_distinct_values_per_group}()},
\code{\link{specimen_get_distinct_values}()},
\code{\link{specimen_get_field_info}()},
\code{\link{specimen_get_paths}()},
\code{\link{specimen_get_settings}()},
\code{\link{specimen_group_by_scientific_name}()},
\code{\link{specimen_query_with_name_resolution}()},
\code{\link{specimen_query}()}
}
\concept{nbaR.SpecimenClient-wrappers}
