% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecimenIdentification.r
\docType{data}
\name{SpecimenIdentification}
\alias{SpecimenIdentification}
\title{SpecimenIdentification Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SpecimenIdentification
}
\description{
SpecimenIdentification Class
}
\section{Fields}{

\describe{
\item{\code{remarks}}{}
}}

\keyword{datasets}
