% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarySourceSystem.r
\docType{data}
\name{SummarySourceSystem}
\alias{SummarySourceSystem}
\title{SummarySourceSystem Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SummarySourceSystem
}
\description{
SummarySourceSystem Class
}
\section{Fields}{

\describe{
\item{\code{code}}{}
}}

\keyword{datasets}
