% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryVernacularName.r
\docType{class}
\name{SummaryVernacularName}
\alias{SummaryVernacularName}
\title{SummaryVernacularName Class}
\format{R6 class}
\usage{
# SummaryVernacularName$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for SummaryVernacularName objects.
}
\section{Fields}{

\describe{
\item{\code{name}}{character}

\item{\code{language}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor SummaryVernacularName object.

}
\item{\code{$fromList(SummaryVernacularNameList)}}{

  Create SummaryVernacularName object from list.

}

\item{\code{$toList()}}{

  Get list representation of SummaryVernacularName.

}
\item{\code{fromJSONString(SummaryVernacularNameJson)}}{

  Create SummaryVernacularName object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of SummaryVernacularName.

}
}
}

\keyword{datasets}
