% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonClient-wrapper.r
\name{taxon_count_distinct_values}
\alias{taxon_count_distinct_values}
\title{Count the distinct number of values that exist for a given field}
\usage{
taxon_count_distinct_values(field = NULL, ...)
}
\arguments{
\item{field}{name of field in the taxon object, type:}

\item{...}{additional parameters passed to count_distinct_values from class class nbaR.TaxonClient}
}
\value{
scalar
}
\description{
This is a wrapper function for count_distinct_values
from class nbaR.TaxonClient.
}
\details{
See also endpoint /getDistinctValues
}
\seealso{
Other nbaR.TaxonClient-wrappers: \code{\link{taxon_count_distinct_values_per_group}},
  \code{\link{taxon_count}},
  \code{\link{taxon_download_query}},
  \code{\link{taxon_dwca_get_data_set_names}},
  \code{\link{taxon_dwca_get_data_set}},
  \code{\link{taxon_dwca_query}},
  \code{\link{taxon_find_by_ids}},
  \code{\link{taxon_find}},
  \code{\link{taxon_get_distinct_values_per_group}},
  \code{\link{taxon_get_distinct_values}},
  \code{\link{taxon_get_field_info}},
  \code{\link{taxon_get_paths}},
  \code{\link{taxon_get_settings}},
  \code{\link{taxon_get_setting}},
  \code{\link{taxon_group_by_scientific_name}},
  \code{\link{taxon_is_operator_allowed}},
  \code{\link{taxon_query}}
}
\concept{nbaR.TaxonClient-wrappers}
