% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonClient-wrapper.r
\name{taxon_find_by_ids}
\alias{taxon_find_by_ids}
\title{Find taxa by ids}
\usage{
taxon_find_by_ids(ids = NULL, returnType = "data.frame", ...)
}
\arguments{
\item{ids}{ids of multiple taxa, separated by comma, type: character}

\item{returnType}{Either \code{list} or \code{data.frame} (default)}

\item{...}{additional parameters passed to find_by_ids from class nbaR.TaxonClient}
}
\value{
list or data.frame, as specified by \code{returnType}
}
\description{
This is a wrapper for the method \code{ find_by_ids }
from class \code{ TaxonClient}.
}
\details{
Given multiple ids, returns a list of taxa
}
\seealso{
Other nbaR.TaxonClient-wrappers: \code{\link{taxon_count_distinct_values_per_group}},
  \code{\link{taxon_count_distinct_values}},
  \code{\link{taxon_count}},
  \code{\link{taxon_download_query}},
  \code{\link{taxon_dwca_get_data_set_names}},
  \code{\link{taxon_dwca_get_data_set}},
  \code{\link{taxon_dwca_query}}, \code{\link{taxon_find}},
  \code{\link{taxon_get_distinct_values_per_group}},
  \code{\link{taxon_get_distinct_values}},
  \code{\link{taxon_get_field_info}},
  \code{\link{taxon_get_paths}},
  \code{\link{taxon_get_settings}},
  \code{\link{taxon_get_setting}},
  \code{\link{taxon_group_by_scientific_name}},
  \code{\link{taxon_is_operator_allowed}},
  \code{\link{taxon_query}}
}
\concept{nbaR.TaxonClient-wrappers}
