% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonClient-wrapper.r
\name{taxon_get_distinct_values}
\alias{taxon_get_distinct_values}
\title{Get all different values that can be found for one field}
\usage{
taxon_get_distinct_values(field = NULL, returnType = "data.frame", ...)
}
\arguments{
\item{field}{name of field in a taxon object, type:}

\item{returnType}{Either \code{list} or \code{data.frame} (default)}

\item{...}{additional parameters passed to get_distinct_values from class nbaR.TaxonClient}
}
\value{
scalar
}
\description{
This is a wrapper for the method \code{ get_distinct_values }
from class \code{ TaxonClient}.
}
\details{
A list of all fields for taxon documents can be retrieved with /metadata/getFieldInfo
}
\seealso{
Other nbaR.TaxonClient-wrappers: 
\code{\link{taxon_count_distinct_values_per_group}()},
\code{\link{taxon_count_distinct_values}()},
\code{\link{taxon_count}()},
\code{\link{taxon_download_query}()},
\code{\link{taxon_dwca_get_data_set_names}()},
\code{\link{taxon_dwca_get_data_set}()},
\code{\link{taxon_dwca_query}()},
\code{\link{taxon_find_by_ids}()},
\code{\link{taxon_find}()},
\code{\link{taxon_get_distinct_values_per_group}()},
\code{\link{taxon_get_field_info}()},
\code{\link{taxon_get_paths}()},
\code{\link{taxon_get_settings}()},
\code{\link{taxon_group_by_scientific_name}()},
\code{\link{taxon_is_operator_allowed}()},
\code{\link{taxon_query}()}
}
\concept{nbaR.TaxonClient-wrappers}
