% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonClient-wrapper.r
\name{taxon_get_paths}
\alias{taxon_get_paths}
\title{Returns the full path of all fields within a document}
\usage{
taxon_get_paths(...)
}
\arguments{
\item{...}{additional parameters passed to get_paths from class class nbaR.TaxonClient}
}
\value{
scalar
}
\description{
This is a wrapper function for get_paths
from class nbaR.TaxonClient.
}
\details{
See also metadata/getFieldInfo for all allowed operators per field
}
\seealso{
Other nbaR.TaxonClient-wrappers: \code{\link{taxon_count_distinct_values_per_group}},
  \code{\link{taxon_count_distinct_values}},
  \code{\link{taxon_count}},
  \code{\link{taxon_download_query}},
  \code{\link{taxon_dwca_get_data_set_names}},
  \code{\link{taxon_dwca_get_data_set}},
  \code{\link{taxon_dwca_query}},
  \code{\link{taxon_find_by_ids}},
  \code{\link{taxon_find}},
  \code{\link{taxon_get_distinct_values_per_group}},
  \code{\link{taxon_get_distinct_values}},
  \code{\link{taxon_get_field_info}},
  \code{\link{taxon_get_settings}},
  \code{\link{taxon_get_setting}},
  \code{\link{taxon_group_by_scientific_name}},
  \code{\link{taxon_is_operator_allowed}},
  \code{\link{taxon_query}}
}
\concept{nbaR.TaxonClient-wrappers}
