% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonomicEnrichment.r
\docType{class}
\name{TaxonomicEnrichment}
\alias{TaxonomicEnrichment}
\title{TaxonomicEnrichment Class}
\format{R6 class}
\usage{
# TaxonomicEnrichment$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for TaxonomicEnrichment objects.
}
\section{Fields}{

\describe{
\item{\code{vernacularNames}}{list(SummaryVernacularName)}

\item{\code{synonyms}}{list(SummaryScientificName)}

\item{\code{sourceSystem}}{SummarySourceSystem}

\item{\code{taxonId}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor TaxonomicEnrichment object.

}
\item{\code{$fromList(TaxonomicEnrichmentList)}}{

  Create TaxonomicEnrichment object from list.

}

\item{\code{$toList()}}{

  Get list representation of TaxonomicEnrichment.

}
\item{\code{fromJSONString(TaxonomicEnrichmentJson)}}{

  Create TaxonomicEnrichment object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of TaxonomicEnrichment.

}
}
}

\keyword{datasets}
