% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{thr_donut_plot}
\alias{thr_donut_plot}
\title{Threat status donut plot}
\usage{
thr_donut_plot(DF, FILL)
}
\arguments{
\item{DF}{The data frame that contains the information on threat status}

\item{FILL}{The groups}
}
\value{
Returns a bar donut plot of threat status
}
\description{
Donut plots for the threat status.
This function requires a dataframe that contains a columns of the
the ecosystem function groups or taxa, ecosystem types or species, and their threat status.
}
\details{
The function will group the ecosystem types or species by the functional
group or taxa and threat status, and calculate the number of ecosystem types or species within each functional
group or taxa with that threat status.
It will also create the colours and axis titles.
}
\examples{
#test <- thr_donut_plot(mydata, threat_status)

}
