% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{thr_donut_plot_test}
\alias{thr_donut_plot_test}
\title{Threat status donut plot}
\usage{
thr_donut_plot_test(DF, FILL)
}
\arguments{
\item{DF}{The data frame that contains the information on threat status}

\item{FILL}{The groups}
}
\value{
Returns a bar donut plot of threat status
}
\description{
Donut plots for the threat status of the ecosystem types
}
\examples{
#test <- thr_donut_plot(mydata, ecosystem_functional_grps, percentages, threat_status)

}
