% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{thr_tbl}
\alias{thr_tbl}
\title{Threat status table}
\usage{
thr_tbl(DF, COL)
}
\arguments{
\item{DF}{The data frame that contains the data}

\item{COL}{the column containing the threat status categories}
}
\description{
A basic table with a purple heading bar and coloured blocks around the
threat status categories
}
\examples{
#test <- thr_tbl(bird_data, Status)

}
