% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeNBADB.R
\name{writeNBADB}
\alias{writeNBADB}
\title{Function Wrapper for writing NBA data to PostgreSQL database}
\usage{
writeNBADB(dataset, dbname, schema, pghost = "localhost", pgport = 5432,
  pguser = "postgres", pgpassword = "postgres", owrite = FALSE)
}
\arguments{
\item{dataset}{Character name of the dataset. It maybe a R dataset or .RDS file}

\item{dbname}{Name of the PostgreSQL database}

\item{schema}{Name of the PostgreSQL schema in the database}

\item{pghost}{Name or IP address of DB host. By default it is localhost.}

\item{pgport}{Port number of PostgreSQL server. By default it is 5432.}

\item{pguser}{Database user login. By default it is "postgres"}

\item{pgpassword}{Database user password By default it is "postgres"}

\item{owrite}{TRUE/FALSE. Decides if database table should by overwritten or appended to existing table. By default new rows will be appended.}
}
\value{
PostgreSQL table
}
\description{
Provide a dataset name or .RDS file name and write the data dircetly and easy to PostgreSQL database
}
\examples{
writeNBADB("GameCalendar2016","NBA","nba16","localhost",5432,"postgres","postgres",overwrite = TRUE)

}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{NBAr,}
\keyword{PostgreSQL}
