% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNKCJE.R
\name{writePG}
\alias{writePG}
\title{Download data from NBA.com/Stats/ and build your own database in local PostgreSQL instance}
\usage{
writePG(dataset, dbname, schema, pghost = "localhost", pgport = 5432,
  pguser = "postgres", pgpassword = "postgres", owrite = FALSE)
}
\description{
Download data from NBA.com/Stats/ and build your own database in local PostgreSQL instance
}

