% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agent.r
\docType{data}
\name{Agent}
\alias{Agent}
\title{Agent Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Agent
}
\description{
Agent Class
}
\section{Fields}{

\describe{
\item{\code{agentText}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Agent object.

}
\item{\code{$fromList(AgentList)}}{

  Create Agent object from list.

}

\item{\code{$toList()}}{

  Get list representation of Agent.

}
\item{\code{fromJSONString(AgentJson)}}{

  Create Agent object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Agent.

}
}
}

\keyword{datasets}
