% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agent.r
\docType{data}
\name{Agent}
\alias{Agent}
\title{Agent Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Agent
}
\description{
Agent Class
}
\section{Fields}{

\describe{
\item{\code{agentText}}{}
}}

\keyword{datasets}
