% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiClient.r
\docType{data}
\name{ApiClient}
\alias{ApiClient}
\title{operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApiClient
}
\description{
Generic class for API client calls
}
\section{Fields}{

\describe{
\item{\code{basePath}}{specifies the base URL of the API, defaults to
http://api.biodiversitydata.nl/v2}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{
}
ApiClient Class

Generic API client for Swagger client library builds.
Swagger generic API client. This client handles the client-
server communication, and is invariant across implementations. Specifics of
the methods and models for each application are generated from the Swagger
templates.

NOTE: This class is auto generated by the swagger code generator program.
Ref: https://github.com/swagger-api/swagger-codegen
Do not edit the class manually.
}

\keyword{datasets}
