% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiClient.r
\docType{data}
\name{ApiClient}
\alias{ApiClient}
\title{ApiClient class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApiClient
}
\description{
Generic class for API client calls, all nbaR client classes
derive from this class
}
\section{Fields}{

\describe{
\item{\code{basePath}}{specifies the base URL of the API, defaults to
http://api.biodiversitydata.nl/v2}

\item{\code{userAgent}}{Set the user agent of the request, defaults to
nbaR/0.0.1}
}}

\keyword{datasets}
