% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssociatedTaxon.r
\docType{class}
\name{AssociatedTaxon}
\alias{AssociatedTaxon}
\title{AssociatedTaxon Class}
\format{R6 class}
\usage{
# AssociatedTaxon$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for AssociatedTaxon objects.
}
\section{Fields}{

\describe{
\item{\code{name}}{character}

\item{\code{relationType}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor AssociatedTaxon object.

}
\item{\code{$fromList(AssociatedTaxonList)}}{

  Create AssociatedTaxon object from list.

}

\item{\code{$toList()}}{

  Get list representation of AssociatedTaxon.

}
\item{\code{fromJSONString(AssociatedTaxonJson)}}{

  Create AssociatedTaxon object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of AssociatedTaxon.

}
}
}

\keyword{datasets}
