% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{basic_tbl}
\alias{basic_tbl}
\title{Basic table}
\usage{
basic_tbl(DF)
}
\arguments{
\item{DF}{The data frame that contains the data}
}
\description{
A basic table with a purple heading bar.
This data should be in the same fomat as for the plot functions
}
\examples{
#test <- basic_tbl(bird_data)

}
