% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioStratigraphy.r
\docType{data}
\name{BioStratigraphy}
\alias{BioStratigraphy}
\title{BioStratigraphy Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BioStratigraphy
}
\description{
BioStratigraphy Class
}
\section{Fields}{

\describe{
\item{\code{youngBioDatingQualifier}}{}

\item{\code{youngBioName}}{}

\item{\code{youngFossilZone}}{}

\item{\code{youngFossilSubZone}}{}

\item{\code{youngBioCertainty}}{}

\item{\code{youngStratType}}{}

\item{\code{bioDatingQualifier}}{}

\item{\code{bioPreferredFlag}}{}

\item{\code{rangePosition}}{}

\item{\code{oldBioName}}{}

\item{\code{bioIdentifier}}{}

\item{\code{oldFossilzone}}{}

\item{\code{oldFossilSubzone}}{}

\item{\code{oldBioCertainty}}{}

\item{\code{oldBioStratType}}{}
}}

\keyword{datasets}
