% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioStratigraphy.r
\docType{class}
\name{BioStratigraphy}
\alias{BioStratigraphy}
\title{BioStratigraphy Class}
\format{R6 class}
\usage{
# BioStratigraphy$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for BioStratigraphy objects.
}
\section{Fields}{

\describe{
\item{\code{youngBioDatingQualifier}}{character}

\item{\code{youngBioName}}{character}

\item{\code{youngFossilZone}}{character}

\item{\code{youngFossilSubZone}}{character}

\item{\code{youngBioCertainty}}{character}

\item{\code{youngStratType}}{character}

\item{\code{bioDatingQualifier}}{character}

\item{\code{bioPreferredFlag}}{logical}

\item{\code{rangePosition}}{character}

\item{\code{oldBioName}}{character}

\item{\code{bioIdentifier}}{character}

\item{\code{oldFossilzone}}{character}

\item{\code{oldFossilSubzone}}{character}

\item{\code{oldBioCertainty}}{character}

\item{\code{oldBioStratType}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor BioStratigraphy object.

}
\item{\code{$fromList(BioStratigraphyList)}}{

  Create BioStratigraphy object from list.

}

\item{\code{$toList()}}{

  Get list representation of BioStratigraphy.

}
\item{\code{fromJSONString(BioStratigraphyJson)}}{

  Create BioStratigraphy object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of BioStratigraphy.

}
}
}

\keyword{datasets}
