% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChronoStratigraphy.r
\docType{data}
\name{ChronoStratigraphy}
\alias{ChronoStratigraphy}
\title{ChronoStratigraphy Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ChronoStratigraphy
}
\description{
ChronoStratigraphy Class
}
\section{Fields}{

\describe{
\item{\code{youngRegionalSubstage}}{}

\item{\code{youngRegionalStage}}{}

\item{\code{youngRegionalSeries}}{}

\item{\code{youngDatingQualifier}}{}

\item{\code{youngInternSystem}}{}

\item{\code{youngInternSubstage}}{}

\item{\code{youngInternStage}}{}

\item{\code{youngInternSeries}}{}

\item{\code{youngInternErathem}}{}

\item{\code{youngInternEonothem}}{}

\item{\code{youngChronoName}}{}

\item{\code{youngCertainty}}{}

\item{\code{oldDatingQualifier}}{}

\item{\code{chronoPreferredFlag}}{}

\item{\code{oldRegionalSubstage}}{}

\item{\code{oldRegionalStage}}{}

\item{\code{oldRegionalSeries}}{}

\item{\code{oldInternSystem}}{}

\item{\code{oldInternSubstage}}{}

\item{\code{oldInternStage}}{}

\item{\code{oldInternSeries}}{}

\item{\code{oldInternErathem}}{}

\item{\code{oldInternEonothem}}{}

\item{\code{oldChronoName}}{}

\item{\code{chronoIdentifier}}{}

\item{\code{oldCertainty}}{}
}}

\keyword{datasets}
