% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChronoStratigraphy.r
\docType{data}
\name{ChronoStratigraphy}
\alias{ChronoStratigraphy}
\title{ChronoStratigraphy Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ChronoStratigraphy
}
\description{
ChronoStratigraphy Class
}
\section{Fields}{

\describe{
\item{\code{oldCertainty}}{}
}}

\keyword{datasets}
