% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChronoStratigraphy.r
\docType{data}
\name{ChronoStratigraphy}
\alias{ChronoStratigraphy}
\title{ChronoStratigraphy Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ChronoStratigraphy
}
\description{
ChronoStratigraphy Class
}
\section{Fields}{

\describe{
\item{\code{youngRegionalSubstage}}{character}

\item{\code{youngRegionalStage}}{character}

\item{\code{youngRegionalSeries}}{character}

\item{\code{youngDatingQualifier}}{character}

\item{\code{youngInternSystem}}{character}

\item{\code{youngInternSubstage}}{character}

\item{\code{youngInternStage}}{character}

\item{\code{youngInternSeries}}{character}

\item{\code{youngInternErathem}}{character}

\item{\code{youngInternEonothem}}{character}

\item{\code{youngChronoName}}{character}

\item{\code{youngCertainty}}{character}

\item{\code{oldDatingQualifier}}{character}

\item{\code{chronoPreferredFlag}}{logical}

\item{\code{oldRegionalSubstage}}{character}

\item{\code{oldRegionalStage}}{character}

\item{\code{oldRegionalSeries}}{character}

\item{\code{oldInternSystem}}{character}

\item{\code{oldInternSubstage}}{character}

\item{\code{oldInternStage}}{character}

\item{\code{oldInternSeries}}{character}

\item{\code{oldInternErathem}}{character}

\item{\code{oldInternEonothem}}{character}

\item{\code{oldChronoName}}{character}

\item{\code{chronoIdentifier}}{character}

\item{\code{oldCertainty}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor ChronoStratigraphy object.

}
\item{\code{$fromList(ChronoStratigraphyList)}}{

  Create ChronoStratigraphy object from list.

}

\item{\code{$toList()}}{

  Get list representation of ChronoStratigraphy.

}
\item{\code{fromJSONString(ChronoStratigraphyJson)}}{

  Create ChronoStratigraphy object from list.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of ChronoStratigraphy.

}
}
}

\keyword{datasets}
