% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crs.r
\docType{class}
\name{Crs}
\alias{Crs}
\title{Crs Class}
\format{R6 class}
\usage{
# Crs$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for Crs objects.
}
\section{Fields}{

\describe{
\item{\code{type}}{character}

\item{\code{properties}}{list}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Crs object.

}
\item{\code{$fromList(CrsList)}}{

  Create Crs object from list.

}

\item{\code{$toList()}}{

  Get list representation of Crs.

}
\item{\code{fromJSONString(CrsJson)}}{

  Create Crs object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Crs.

}
}
}

\keyword{datasets}
