% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crs.r
\docType{data}
\name{Crs}
\alias{Crs}
\title{Crs Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Crs
}
\description{
Crs Class
}
\section{Fields}{

\describe{
\item{\code{properties}}{}
}}

\keyword{datasets}
