% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crs.r
\docType{data}
\name{Crs}
\alias{Crs}
\title{Crs Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Crs
}
\description{
Crs Class
}
\section{Fields}{

\describe{
\item{\code{type}}{character}

\item{\code{properties}}{list}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Crs object.

}
\item{\code{$fromList(CrsList)}}{

  Create Crs object from list.

}

\item{\code{$toList()}}{

  Get list representation of Crs.

}
\item{\code{fromJSONString(CrsJson)}}{

  Create Crs object from list.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Crs.

}
}
}

\keyword{datasets}
