% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultClassification.r
\docType{data}
\name{DefaultClassification}
\alias{DefaultClassification}
\title{DefaultClassification Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DefaultClassification
}
\description{
DefaultClassification Class
}
\section{Fields}{

\describe{
\item{\code{infraspecificRank}}{}
}}

\keyword{datasets}
