% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expert.r
\docType{data}
\name{Expert}
\alias{Expert}
\title{Expert Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Expert
}
\description{
Expert Class
}
\section{Fields}{

\describe{
\item{\code{agentText}}{}

\item{\code{fullName}}{}

\item{\code{organization}}{}
}}

\keyword{datasets}
