% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expert.r
\docType{class}
\name{Expert}
\alias{Expert}
\title{Expert Class}
\format{R6 class}
\usage{
# Expert$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for Expert objects.
}
\section{Fields}{

\describe{
\item{\code{agentText}}{character}

\item{\code{fullName}}{character}

\item{\code{organization}}{Organization}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Expert object.

}
\item{\code{$fromList(ExpertList)}}{

  Create Expert object from list.

}

\item{\code{$toList()}}{

  Get list representation of Expert.

}
\item{\code{fromJSONString(ExpertJson)}}{

  Create Expert object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Expert.

}
}
}

\keyword{datasets}
