% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expert.r
\docType{data}
\name{Expert}
\alias{Expert}
\title{Expert Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Expert
}
\description{
Expert Class
}
\section{Fields}{

\describe{
\item{\code{agentText}}{character}

\item{\code{fullName}}{character}

\item{\code{organization}}{Organization}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Expert object.

}
\item{\code{$fromList(ExpertList)}}{

  Create Expert object from list.

}

\item{\code{$toList()}}{

  Get list representation of Expert.

}
\item{\code{fromJSONString(ExpertJson)}}{

  Create Expert object from list.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Expert.

}
}
}

\keyword{datasets}
