% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature.r
\docType{class}
\name{Feature}
\alias{Feature}
\title{Feature Class}
\format{R6 class}
\usage{
# Feature$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for Feature objects.
}
\section{Fields}{

\describe{
\item{\code{crs}}{Crs}

\item{\code{bbox}}{list(numeric)}

\item{\code{properties}}{list}

\item{\code{geometry}}{list}

\item{\code{id}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Feature object.

}
\item{\code{$fromList(FeatureList)}}{

  Create Feature object from list.

}

\item{\code{$toList()}}{

  Get list representation of Feature.

}
\item{\code{fromJSONString(FeatureJson)}}{

  Create Feature object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Feature.

}
}
}

\keyword{datasets}
