% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature.r
\docType{data}
\name{Feature}
\alias{Feature}
\title{Feature Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Feature
}
\description{
Feature Class
}
\section{Fields}{

\describe{
\item{\code{crs}}{}

\item{\code{bbox}}{}

\item{\code{properties}}{}

\item{\code{geometry}}{}

\item{\code{id}}{}
}}

\keyword{datasets}
