% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature.r
\docType{data}
\name{Feature}
\alias{Feature}
\title{Feature Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Feature
}
\description{
Feature Class
}
\section{Fields}{

\describe{
\item{\code{crs}}{Crs}

\item{\code{bbox}}{list(numeric)}

\item{\code{properties}}{list}

\item{\code{geometry}}{list}

\item{\code{id}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Feature object.

}
\item{\code{$fromList(FeatureList)}}{

  Create Feature object from list.

}

\item{\code{$toList()}}{

  Get list representation of Feature.

}
\item{\code{fromJSONString(FeatureJson)}}{

  Create Feature object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Feature.

}
}
}

\keyword{datasets}
