% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureCollection.r
\docType{data}
\name{FeatureCollection}
\alias{FeatureCollection}
\title{FeatureCollection Class}
\format{}
\usage{

}
\description{
FeatureCollection Class
}
\section{Fields}{

\describe{
\item{\code{crs}}{Crs}

\item{\code{bbox}}{list(numeric)}

\item{\code{features}}{list(Feature)}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor FeatureCollection object.

}
\item{\code{$fromList(FeatureCollectionList)}}{

  Create FeatureCollection object from list.

}

\item{\code{$toList()}}{

  Get list representation of FeatureCollection.

}
\item{\code{fromJSONString(FeatureCollectionJson)}}{

  Create FeatureCollection object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of FeatureCollection.

}
}
}

\keyword{datasets}
