% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureCollection.r
\docType{data}
\name{FeatureCollection}
\alias{FeatureCollection}
\title{FeatureCollection Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FeatureCollection
}
\description{
FeatureCollection Class
}
\section{Fields}{

\describe{
\item{\code{crs}}{}

\item{\code{bbox}}{}

\item{\code{features}}{}
}}

\keyword{datasets}
