% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureCollection.r
\docType{class}
\name{FeatureCollection}
\alias{FeatureCollection}
\title{FeatureCollection Class}
\format{R6 class}
\usage{
# FeatureCollection$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for FeatureCollection objects.
}
\section{Fields}{

\describe{
\item{\code{crs}}{Crs}

\item{\code{bbox}}{list(numeric)}

\item{\code{features}}{list(Feature)}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor FeatureCollection object.

}
\item{\code{$fromList(FeatureCollectionList)}}{

  Create FeatureCollection object from list.

}

\item{\code{$toList()}}{

  Get list representation of FeatureCollection.

}
\item{\code{fromJSONString(FeatureCollectionJson)}}{

  Create FeatureCollection object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of FeatureCollection.

}
}
}

\keyword{datasets}
