% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.r
\docType{data}
\name{Filter}
\alias{Filter}
\title{Filter Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Filter
}
\description{
Filter Class
}
\section{Fields}{

\describe{
\item{\code{acceptRegexp}}{character}

\item{\code{rejectRegexp}}{character}

\item{\code{acceptValues}}{list(character)}

\item{\code{rejectValues}}{list(character)}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Filter object.

}
\item{\code{$fromList(FilterList)}}{

  Create Filter object from list.

}

\item{\code{$toList()}}{

  Get list representation of Filter.

}
\item{\code{fromJSONString(FilterJson)}}{

  Create Filter object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Filter.

}
}
}

\keyword{datasets}
