% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.r
\docType{data}
\name{Filter}
\alias{Filter}
\title{Filter Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Filter
}
\description{
Filter Class
}
\section{Fields}{

\describe{
\item{\code{acceptRegexp}}{}

\item{\code{rejectRegexp}}{}

\item{\code{acceptValues}}{}

\item{\code{rejectValues}}{}
}}

\keyword{datasets}
