% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatheringEvent.r
\docType{data}
\name{GatheringEvent}
\alias{GatheringEvent}
\title{GatheringEvent Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GatheringEvent
}
\description{
GatheringEvent Class
}
\section{Fields}{

\describe{
\item{\code{bioStratigraphic}}{}
}}

\keyword{datasets}
