% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatheringEvent.r
\docType{class}
\name{GatheringEvent}
\alias{GatheringEvent}
\title{GatheringEvent Class}
\format{R6 class}
\usage{
# GatheringEvent$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for GatheringEvent objects.
}
\section{Fields}{

\describe{
\item{\code{projectTitle}}{character}

\item{\code{worldRegion}}{character}

\item{\code{continent}}{character}

\item{\code{country}}{character}

\item{\code{iso3166Code}}{character}

\item{\code{provinceState}}{character}

\item{\code{island}}{character}

\item{\code{locality}}{character}

\item{\code{city}}{character}

\item{\code{sublocality}}{character}

\item{\code{localityText}}{character}

\item{\code{dateTimeBegin}}{character}

\item{\code{dateTimeEnd}}{character}

\item{\code{method}}{character}

\item{\code{altitude}}{character}

\item{\code{altitudeUnifOfMeasurement}}{character}

\item{\code{biotopeText}}{character}

\item{\code{depth}}{character}

\item{\code{depthUnitOfMeasurement}}{character}

\item{\code{gatheringPersons}}{list(Person)}

\item{\code{gatheringOrganizations}}{list(Organization)}

\item{\code{siteCoordinates}}{list(GatheringSiteCoordinates)}

\item{\code{namedAreas}}{list(NamedArea)}

\item{\code{associatedTaxa}}{list(AssociatedTaxon)}

\item{\code{chronoStratigraphy}}{list(ChronoStratigraphy)}

\item{\code{lithoStratigraphy}}{list(LithoStratigraphy)}

\item{\code{bioStratigraphic}}{list(BioStratigraphy)}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor GatheringEvent object.

}
\item{\code{$fromList(GatheringEventList)}}{

  Create GatheringEvent object from list.

}

\item{\code{$toList()}}{

  Get list representation of GatheringEvent.

}
\item{\code{fromJSONString(GatheringEventJson)}}{

  Create GatheringEvent object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of GatheringEvent.

}
}
}

\keyword{datasets}
