% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatheringSiteCoordinates.r
\docType{data}
\name{GatheringSiteCoordinates}
\alias{GatheringSiteCoordinates}
\title{GatheringSiteCoordinates Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GatheringSiteCoordinates
}
\description{
GatheringSiteCoordinates Class
}
\section{Fields}{

\describe{
\item{\code{geoShape}}{}
}}

\keyword{datasets}
