% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatheringSiteCoordinates.r
\docType{data}
\name{GatheringSiteCoordinates}
\alias{GatheringSiteCoordinates}
\title{GatheringSiteCoordinates Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GatheringSiteCoordinates
}
\description{
GatheringSiteCoordinates Class
}
\section{Fields}{

\describe{
\item{\code{longitudeDecimal}}{}

\item{\code{latitudeDecimal}}{}

\item{\code{gridCellSystem}}{}

\item{\code{gridLatitudeDecimal}}{}

\item{\code{gridLongitudeDecimal}}{}

\item{\code{gridCellCode}}{}

\item{\code{gridQualifier}}{}

\item{\code{geoShape}}{}
}}

\keyword{datasets}
