% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatheringSiteCoordinates.r
\docType{data}
\name{GatheringSiteCoordinates}
\alias{GatheringSiteCoordinates}
\title{GatheringSiteCoordinates Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GatheringSiteCoordinates
}
\description{
GatheringSiteCoordinates Class
}
\section{Fields}{

\describe{
\item{\code{longitudeDecimal}}{numeric}

\item{\code{latitudeDecimal}}{numeric}

\item{\code{gridCellSystem}}{character}

\item{\code{gridLatitudeDecimal}}{numeric}

\item{\code{gridLongitudeDecimal}}{numeric}

\item{\code{gridCellCode}}{character}

\item{\code{gridQualifier}}{character}

\item{\code{geoShape}}{Point}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor GatheringSiteCoordinates object.

}
\item{\code{$fromList(GatheringSiteCoordinatesList)}}{

  Create GatheringSiteCoordinates object from list.

}

\item{\code{$toList()}}{

  Get list representation of GatheringSiteCoordinates.

}
\item{\code{fromJSONString(GatheringSiteCoordinatesJson)}}{

  Create GatheringSiteCoordinates object from list.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of GatheringSiteCoordinates.

}
}
}

\keyword{datasets}
