% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoArea.r
\docType{data}
\name{GeoArea}
\alias{GeoArea}
\title{GeoArea Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GeoArea
}
\description{
GeoArea Class
}
\section{Fields}{

\describe{
\item{\code{sourceSystem}}{}

\item{\code{sourceSystemId}}{}

\item{\code{recordURI}}{}

\item{\code{id}}{}

\item{\code{areaType}}{}

\item{\code{locality}}{}

\item{\code{shape}}{}

\item{\code{source}}{}

\item{\code{isoCode}}{}

\item{\code{countryNL}}{}
}}

\keyword{datasets}
