% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoArea.r
\docType{class}
\name{GeoArea}
\alias{GeoArea}
\title{GeoArea Class}
\format{R6 class}
\usage{
# GeoArea$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for GeoArea objects. Geographical mapping of species occurrences is essential in biogeographical analyses. More than 1.2 Million of our specimen records are geo-referenced, i.e. they have latitude and longitude coordinates reported for the sita at which they were collected. The data type geo stores geographical areas to which the geo-referenced species can be mapped. This enables querying for specimens found in a specific region, to e.g. retrieve all the primate species collected on Madagascar. The areas are represented as polygons of longitude/latitude pairs. All polygons are coded in the GeoJSON format.
}
\section{Fields}{

\describe{
\item{\code{sourceSystem}}{SourceSystem}

\item{\code{sourceSystemId}}{character}

\item{\code{recordURI}}{character}

\item{\code{id}}{character}

\item{\code{areaType}}{character}

\item{\code{locality}}{character}

\item{\code{shape}}{list}

\item{\code{source}}{character}

\item{\code{isoCode}}{character}

\item{\code{countryNL}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor GeoArea object.

}
\item{\code{$fromList(GeoAreaList)}}{

  Create GeoArea object from list.

}

\item{\code{$toList()}}{

  Get list representation of GeoArea.

}
\item{\code{fromJSONString(GeoAreaJson)}}{

  Create GeoArea object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of GeoArea.

}
}
}

\keyword{datasets}
