% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoArea.r
\docType{data}
\name{GeoArea}
\alias{GeoArea}
\title{GeoArea Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GeoArea
}
\description{
GeoArea Class
}
\section{Fields}{

\describe{
\item{\code{countryNL}}{}
}}

\keyword{datasets}
