% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoApi.r, R/GeoClient.r
\docType{data}
\name{GeoClient}
\alias{GeoClient}
\alias{GeoClient}
\title{Geo operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GeoClient

GeoClient
}
\description{
nbaR.Geo

nbaR.Geo
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}

\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

count_http_get Get the number of geo areas matching a condition


count_http_post_json Get the number of geo areas matching a condition


find1 Find a GEO area by id


find_by_ids1 Find geo areas by ids


get_distinct_values_http_get Get all different values that exist for a field


get_distinct_values_http_post_json Get all different values that exist for a field


get_field_info Returns extended information for each field of a specimen document


get_geo_json_for_locality Retrieve a GeoJson object for a given locality


get_paths Returns the full path of all fields within a document


get_settings List all publicly available configuration settings for the NBA


get_settings1 Get the value of an NBA setting


is_operator_allowed Checks if a given operator is allowed for a given field


query_http_get Query for geo areas


query_http_post_json Query for geo areas

}


\describe{

count_http_get Get the number of geo areas matching a condition


count_http_post_json Get the number of geo areas matching a condition


find1 Find a GEO area by id


find_by_ids1 Find geo areas by ids


get_distinct_values_http_get Get all different values that exist for a field


get_distinct_values_http_post_json Get all different values that exist for a field


get_field_info Returns extended information for each field of a specimen document


get_geo_json_for_locality Retrieve a GeoJson object for a given locality


get_paths Returns the full path of all fields within a document


get_settings Get the value of an NBA setting


get_settings1 List all publicly available configuration settings for the NBA


is_operator_allowed Checks if a given operator is allowed for a given field


query_http_get Query for geo areas


query_http_post_json Query for geo areas

}
}

\keyword{datasets}
