% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoJsonObject.r
\docType{data}
\name{GeoJsonObject}
\alias{GeoJsonObject}
\title{GeoJsonObject Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GeoJsonObject
}
\description{
GeoJsonObject Class
}
\section{Fields}{

\describe{
\item{\code{crs}}{}

\item{\code{bbox}}{}
}}

\keyword{datasets}
