% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoJsonObject.r
\docType{class}
\name{GeoJsonObject}
\alias{GeoJsonObject}
\title{GeoJsonObject Class}
\format{R6 class}
\usage{
# GeoJsonObject$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for GeoJsonObject objects.
}
\section{Fields}{

\describe{
\item{\code{crs}}{Crs}

\item{\code{bbox}}{list(numeric)}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor GeoJsonObject object.

}
\item{\code{$fromList(GeoJsonObjectList)}}{

  Create GeoJsonObject object from list.

}

\item{\code{$toList()}}{

  Get list representation of GeoJsonObject.

}
\item{\code{fromJSONString(GeoJsonObjectJson)}}{

  Create GeoJsonObject object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of GeoJsonObject.

}
}
}

\keyword{datasets}
