% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeometryCollection.r
\docType{data}
\name{GeometryCollection}
\alias{GeometryCollection}
\title{GeometryCollection Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GeometryCollection
}
\description{
GeometryCollection Class
}
\section{Fields}{

\describe{
\item{\code{geometries}}{}
}}

\keyword{datasets}
