% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeometryCollection.r
\docType{data}
\name{GeometryCollection}
\alias{GeometryCollection}
\title{GeometryCollection Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GeometryCollection
}
\description{
GeometryCollection Class
}
\section{Fields}{

\describe{
\item{\code{crs}}{Crs}

\item{\code{bbox}}{list(numeric)}

\item{\code{geometries}}{list(list)}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor GeometryCollection object.

}
\item{\code{$fromList(GeometryCollectionList)}}{

  Create GeometryCollection object from list.

}

\item{\code{$toList()}}{

  Get list representation of GeometryCollection.

}
\item{\code{fromJSONString(GeometryCollectionJson)}}{

  Create GeometryCollection object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of GeometryCollection.

}
}
}

\keyword{datasets}
