% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoxscoreAdv.R
\name{getBoxscore}
\alias{getBoxscore}
\alias{getBoxscoreAdv}
\alias{getBoxscoreMisc}
\alias{getBoxscorePt}
\alias{getBoxscoreTrad}
\title{Download Boxscore data for NBA game}
\usage{
getBoxscoreAdv(gameID, season)

getBoxscoreMisc(gameID, season)

getBoxscorePt(gameID, season)

getBoxscoreTrad(gameID, season)
}
\arguments{
\item{gameID}{Game's ID in NBA.com DB}

\item{season}{Number of the year in which season started}
}
\value{
Dataset containing data from one of the boxscore pages from NBA.com. Functions enable downloading Traditional, Advanced, Miscallenous or Player Tracking data.
}
\description{
Download and process NBA.com boxscore data for given game and season.
}
\examples{
getBoxscoreTrad("0021400001","2014")
getBoxscoreMisc("0021400001","2014")
getBoxscoreAdv("0021400001","2014")
getBoxscorePt("0021400001","2014")

gamelist<- paste("00",as.character(c(21400001:21401230)),sep="")
boxscores <- lapply(gamelist, getBoxscorePt, season="2014")
player.tracking <- do.call(rbind.data.frame,boxscores)

}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{Advanced,}
\keyword{Boxscores,}
\keyword{Game,}
\keyword{Misc,}
\keyword{NBAr,}
\keyword{Player}
\keyword{Tracking}
\keyword{Traditional,}
\keyword{players,}

