% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlaytype.R
\name{getPlaytype}
\alias{getPlaytype}
\title{Download Playtype data for Players and Teams}
\usage{
getPlaytype(Season, Type, Playtype, Names = "offensive", SeasonType = "Reg")
}
\arguments{
\item{Season}{Number of the year in which season started}

\item{Type}{Specify if data is for Team or Player   c("Player,"Team")}

\item{Playtype}{Pick one of the playtypes provided by stats.nba.com. c("Postup","Transition","Isolation","PRBallHandler","PRRollman","Spotup","Handoff","Cut","OffScreen","OffRebound","Misc")}

\item{Names}{Specify if data is for offensive or defensive side of the floor. c("offensive","defensive")}

\item{SeasonType}{Choose data for preseason, regular season or postseason c("Reg","Post","Pre")}
}
\value{
Dataset from stats.nba.com
}
\description{
Download and process NBA.com Playtype data from http://stats.nba.com/players/playtypeexample/ or http://stats.nba.com/teams/playtypeexample/
}
\examples{
getPlaytype("2016","Team","Handoff","offensive","Pre")
getPlaytype("2015","Player","Isolation","defensive","Post")

Season <- "2016"
Names <- c("offensive","defensive")[1]
Type <- c("Player","Team")[1]

Playtype <- c("Postup",
              "Transition",
              "Isolation",
              "PRBallHandler",
              "PRRollman",
              "Spotup",
              "Handoff",
              "Cut",
              "OffScreen",
              "OffRebound",
              "Misc"
              )[1]

SeasonType <- c("Reg","Post","Pre")[1]

playtypes.dataset <- getPlaytype(Season,Type,Playtype,Names,SeasonType)

}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{NBAr,}
\keyword{Playtypes,}
\keyword{players,}
\keyword{teams,}
