% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSchedule.R
\name{getSchedule}
\alias{getSchedule}
\title{Download NBA season games schedule}
\usage{
getSchedule(month, season)
}
\arguments{
\item{month}{Character name of the month}

\item{season}{Number of the year in which season started}
}
\value{
Dataset containing Game Date, Visitor team name, Home team name, Visitor points,Home points and OT information
}
\description{
Downloads and process NBA season games schedule from \url{http://www.basketball-reference.com}
}
\examples{
getSchedule("january","2016")

season.no <- "2016"
months <- c("october","november","december","january","february","march","april","may")
ds.months <- lapply(months,getSchedule, season = season.no)
schedule <- do.call(rbind.data.frame,ds.months)

}
\references{
\url{http://www.basketball-reference.com}
}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{Calendar}
\keyword{NBAr,}
\keyword{Schedule,}
