% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShotchart.R
\name{getShotchart}
\alias{getShotchart}
\title{Download shotchart data for player}
\usage{
getShotchart(playerID, season)
}
\arguments{
\item{playerID}{Player's ID in NBA.com DB}

\item{season}{Number of the year in which season started}
}
\value{
Dataset containing shot information such as location on the floor, type of shot and result per player and game.
}
\description{
Downloads and process NBA.com shotchart data for given player and season.
}
\examples{
getShotchart("2853","2014")

shotcharts<- lapply(players, getShotchart, season = "2014")
shotchart <- do.call(rbind.data.frame,shotcharts)

}
\author{
Patrick Chodowski, \email{Chodowski.Patrick@gmail.com}
}
\keyword{NBAr,}
\keyword{Shotchart,}
\keyword{players}
