% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupByScientificNameQuerySpec.r
\docType{class}
\name{GroupByScientificNameQuerySpec}
\alias{GroupByScientificNameQuerySpec}
\title{GroupByScientificNameQuerySpec Class}
\format{R6 class}
\usage{
# GroupByScientificNameQuerySpec$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for GroupByScientificNameQuerySpec objects.
}
\section{Fields}{

\describe{
\item{\code{constantScore}}{logical
If true, no relevance scores will be calculated for the returned documents. By default Elasticsearch not only determines whether a document matches your search criteria, but also how well it matches them, expressed as a so-called relevance score. If you are not interested in relevance scores, set constantScore to true, as there is some performance overhead associated with calculating relevance scores.}

\item{\code{fields}}{list(character)
The fields to be returned. Specimen, Taxon and MultiMediaObject documents are large documents containing lots of fields. If you are only interested in a few fields, use the fields property to specify them.}

\item{\code{conditions}}{list(QueryCondition)
List of QueryCondition objetcs}

\item{\code{logicalOperator}}{character
(AND/OR) Specifies whether a document must satisfy all search criteria or just one in order to be returned.}

\item{\code{sortFields}}{list(SortField)
Specifies the field(s) on which to sort the documents.}

\item{\code{from}}{integer
The offset in the result set from which to return the documents.}

\item{\code{size}}{integer
The number of documents to return.}

\item{\code{groupSort}}{character}

\item{\code{groupFilter}}{Filter}

\item{\code{specimensFrom}}{integer}

\item{\code{specimensSize}}{integer}

\item{\code{specimensSortFields}}{list(SortField)}

\item{\code{noTaxa}}{logical}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor GroupByScientificNameQuerySpec object.

}
\item{\code{$fromList(GroupByScientificNameQuerySpecList)}}{

  Create GroupByScientificNameQuerySpec object from list.

}

\item{\code{$toList()}}{

  Get list representation of GroupByScientificNameQuerySpec.

}
\item{\code{fromJSONString(GroupByScientificNameQuerySpecJson)}}{

  Create GroupByScientificNameQuerySpec object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of GroupByScientificNameQuerySpec.

}
}
}

\keyword{datasets}
