% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupByScientificNameQuerySpec.r
\docType{data}
\name{GroupByScientificNameQuerySpec}
\alias{GroupByScientificNameQuerySpec}
\title{GroupByScientificNameQuerySpec Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GroupByScientificNameQuerySpec
}
\description{
GroupByScientificNameQuerySpec Class
}
\section{Fields}{

\describe{
\item{\code{noTaxa}}{}
}}

\keyword{datasets}
