% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupByScientificNameQuerySpec.r
\docType{data}
\name{GroupByScientificNameQuerySpec}
\alias{GroupByScientificNameQuerySpec}
\title{GroupByScientificNameQuerySpec Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GroupByScientificNameQuerySpec
}
\description{
GroupByScientificNameQuerySpec Class
}
\section{Fields}{

\describe{
\item{\code{constantScore}}{}

\item{\code{fields}}{}

\item{\code{conditions}}{}

\item{\code{logicalOperator}}{}

\item{\code{sortFields}}{}

\item{\code{from}}{}

\item{\code{size}}{}

\item{\code{groupSort}}{}

\item{\code{groupFilter}}{}

\item{\code{specimensFrom}}{}

\item{\code{specimensSize}}{}

\item{\code{specimensSortFields}}{}

\item{\code{noTaxa}}{}
}}

\keyword{datasets}
