% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Iptc4xmpExt.r
\docType{data}
\name{Iptc4xmpExt}
\alias{Iptc4xmpExt}
\title{Iptc4xmpExt Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Iptc4xmpExt
}
\description{
Iptc4xmpExt Class
}
\section{Fields}{

\describe{
\item{\code{locationShown}}{}

\item{\code{worldRegion}}{}

\item{\code{countryCode}}{}

\item{\code{countryName}}{}

\item{\code{provinceState}}{}

\item{\code{city}}{}

\item{\code{sublocation}}{}
}}

\keyword{datasets}
