% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Iptc4xmpExt.r
\docType{data}
\name{Iptc4xmpExt}
\alias{Iptc4xmpExt}
\title{Iptc4xmpExt Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Iptc4xmpExt
}
\description{
Iptc4xmpExt Class
}
\section{Fields}{

\describe{
\item{\code{locationShown}}{character}

\item{\code{worldRegion}}{character}

\item{\code{countryCode}}{character}

\item{\code{countryName}}{character}

\item{\code{provinceState}}{character}

\item{\code{city}}{character}

\item{\code{sublocation}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Iptc4xmpExt object.

}
\item{\code{$fromList(Iptc4xmpExtList)}}{

  Create Iptc4xmpExt object from list.

}

\item{\code{$toList()}}{

  Get list representation of Iptc4xmpExt.

}
\item{\code{fromJSONString(Iptc4xmpExtJson)}}{

  Create Iptc4xmpExt object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Iptc4xmpExt.

}
}
}

\keyword{datasets}
