% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineString.r
\docType{data}
\name{LineString}
\alias{LineString}
\title{LineString Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LineString
}
\description{
LineString Class
}
\section{Fields}{

\describe{
\item{\code{crs}}{Crs}

\item{\code{bbox}}{list(numeric)}

\item{\code{coordinates}}{list(LngLatAlt)}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor LineString object.

}
\item{\code{$fromList(LineStringList)}}{

  Create LineString object from list.

}

\item{\code{$toList()}}{

  Get list representation of LineString.

}
\item{\code{fromJSONString(LineStringJson)}}{

  Create LineString object from list.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of LineString.

}
}
}

\keyword{datasets}
