% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineString.r
\docType{data}
\name{LineString}
\alias{LineString}
\title{LineString Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LineString
}
\description{
LineString Class
}
\section{Fields}{

\describe{
\item{\code{crs}}{}

\item{\code{bbox}}{}

\item{\code{coordinates}}{}
}}

\keyword{datasets}
