% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LithoStratigraphy.r
\docType{class}
\name{LithoStratigraphy}
\alias{LithoStratigraphy}
\title{LithoStratigraphy Class}
\format{
R6 class

R6 class
}
\usage{
# LithoStratigraphy$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for LithoStratigraphy objects.
}
\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor LithoStratigraphy object.

}
\item{\code{$fromList(LithoStratigraphyList)}}{

  Create LithoStratigraphy object from list.

}

\item{\code{$toList()}}{

  Get list representation of LithoStratigraphy.

}
\item{\code{fromJSONString(LithoStratigraphyJson)}}{

  Create LithoStratigraphy object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of LithoStratigraphy.

}
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{qualifier}}{character}

\item{\code{preferredFlag}}{logical}

\item{\code{member2}}{character}

\item{\code{member}}{character}

\item{\code{informalName2}}{character}

\item{\code{informalName}}{character}

\item{\code{importedName2}}{character}

\item{\code{importedName1}}{character}

\item{\code{lithoIdentifier}}{character}

\item{\code{formation2}}{character}

\item{\code{formationGroup2}}{character}

\item{\code{formationGroup}}{character}

\item{\code{formation}}{character}

\item{\code{certainty2}}{character}

\item{\code{certainty}}{character}

\item{\code{bed2}}{character}

\item{\code{bed}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LithoStratigraphy-new}{\code{LithoStratigraphy$new()}}
\item \href{#method-LithoStratigraphy-toList}{\code{LithoStratigraphy$toList()}}
\item \href{#method-LithoStratigraphy-fromList}{\code{LithoStratigraphy$fromList()}}
\item \href{#method-LithoStratigraphy-toJSONString}{\code{LithoStratigraphy$toJSONString()}}
\item \href{#method-LithoStratigraphy-fromJSONString}{\code{LithoStratigraphy$fromJSONString()}}
\item \href{#method-LithoStratigraphy-print}{\code{LithoStratigraphy$print()}}
\item \href{#method-LithoStratigraphy-clone}{\code{LithoStratigraphy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LithoStratigraphy-new"></a>}}
\if{latex}{\out{\hypertarget{method-LithoStratigraphy-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LithoStratigraphy$new(
  qualifier,
  preferredFlag,
  member2,
  member,
  informalName2,
  informalName,
  importedName2,
  importedName1,
  lithoIdentifier,
  formation2,
  formationGroup2,
  formationGroup,
  formation,
  certainty2,
  certainty,
  bed2,
  bed
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LithoStratigraphy-toList"></a>}}
\if{latex}{\out{\hypertarget{method-LithoStratigraphy-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LithoStratigraphy$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LithoStratigraphy-fromList"></a>}}
\if{latex}{\out{\hypertarget{method-LithoStratigraphy-fromList}{}}}
\subsection{Method \code{fromList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LithoStratigraphy$fromList(LithoStratigraphyList, typeMapping = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LithoStratigraphy-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-LithoStratigraphy-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LithoStratigraphy$toJSONString(pretty = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LithoStratigraphy-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-LithoStratigraphy-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LithoStratigraphy$fromJSONString(LithoStratigraphyJson, typeMapping = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LithoStratigraphy-print"></a>}}
\if{latex}{\out{\hypertarget{method-LithoStratigraphy-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LithoStratigraphy$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LithoStratigraphy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LithoStratigraphy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LithoStratigraphy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
