% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LithoStratigraphy.r
\docType{data}
\name{LithoStratigraphy}
\alias{LithoStratigraphy}
\title{LithoStratigraphy Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LithoStratigraphy
}
\description{
LithoStratigraphy Class
}
\section{Fields}{

\describe{
\item{\code{qualifier}}{}

\item{\code{preferredFlag}}{}

\item{\code{member2}}{}

\item{\code{member}}{}

\item{\code{informalName2}}{}

\item{\code{informalName}}{}

\item{\code{importedName2}}{}

\item{\code{importedName1}}{}

\item{\code{lithoIdentifier}}{}

\item{\code{formation2}}{}

\item{\code{formationGroup2}}{}

\item{\code{formationGroup}}{}

\item{\code{formation}}{}

\item{\code{certainty2}}{}

\item{\code{certainty}}{}

\item{\code{bed2}}{}

\item{\code{bed}}{}
}}

\keyword{datasets}
