% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LngLatAlt.r
\docType{data}
\name{LngLatAlt}
\alias{LngLatAlt}
\title{LngLatAlt Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LngLatAlt
}
\description{
LngLatAlt Class
}
\section{Fields}{

\describe{
\item{\code{longitude}}{numeric}

\item{\code{latitude}}{numeric}

\item{\code{altitude}}{numeric}

\item{\code{additionalElements}}{list(numeric)}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor LngLatAlt object.

}
\item{\code{$fromList(LngLatAltList)}}{

  Create LngLatAlt object from list.

}

\item{\code{$toList()}}{

  Get list representation of LngLatAlt.

}
\item{\code{fromJSONString(LngLatAltJson)}}{

  Create LngLatAlt object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of LngLatAlt.

}
}
}

\keyword{datasets}
