% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LngLatAlt.r
\docType{class}
\name{LngLatAlt}
\alias{LngLatAlt}
\title{LngLatAlt Class}
\format{R6 class}
\usage{
# LngLatAlt$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for LngLatAlt objects.
}
\section{Fields}{

\describe{
\item{\code{longitude}}{numeric}

\item{\code{latitude}}{numeric}

\item{\code{altitude}}{numeric}

\item{\code{additionalElements}}{list(numeric)}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor LngLatAlt object.

}
\item{\code{$fromList(LngLatAltList)}}{

  Create LngLatAlt object from list.

}

\item{\code{$toList()}}{

  Get list representation of LngLatAlt.

}
\item{\code{fromJSONString(LngLatAltJson)}}{

  Create LngLatAlt object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of LngLatAlt.

}
}
}

\keyword{datasets}
