% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataClient-wrapper.r
\name{metadata_get_allowed_date_formats}
\alias{metadata_get_allowed_date_formats}
\title{Get allowed values for dates in queries}
\usage{
metadata_get_allowed_date_formats(returnType = "data.frame", ...)
}
\arguments{
\item{returnType}{Either 'list' or 'data.frame', defaults to 'data.frame'}

\item{...}{additional parameters passed to get_allowed_date_formats from class nbaR.MetadataClient}
}
\value{
scalar
}
\description{
This is a wrapper function for get_allowed_date_formats
from class nbaR.MetadataClient.
}
\details{
Queries with other formatted dates will result in a query error
}
\seealso{
Other nbaR.MetadataClient-wrappers: \code{\link{metadata_get_controlled_list_phase_or_stage}},
  \code{\link{metadata_get_controlled_list_sex}},
  \code{\link{metadata_get_controlled_list_specimen_type_status}},
  \code{\link{metadata_get_controlled_list_taxonomic_status}},
  \code{\link{metadata_get_controlled_lists}},
  \code{\link{metadata_get_rest_services}},
  \code{\link{metadata_get_settings}},
  \code{\link{metadata_get_setting}},
  \code{\link{metadata_get_source_systems}}
}
\concept{nbaR.MetadataClient-wrappers}
