% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataClient.r
\docType{data}
\name{MetadataClient}
\alias{MetadataClient}
\title{Metadata operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MetadataClient
}
\description{
Client class for nbaR.Metadata objects
}
\section{Fields}{

\describe{
\item{\code{basePath}}{Stores url path of the request.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{
\item{\code{ get_allowed_date_formats }}{

  Get allowed values for dates in queries;
  Queries with other formatted dates will result in a query error.

    Parameters:
    \itemize{




        \item \code{ ... } : additional parameters passed to httr::GET
    }
    Returns:
        \code{ character }
}
\item{\code{ get_controlled_list_phase_or_stage }}{

  Get allowed values for the field &#39;PhaseOrStage&#39; in a specimen document;
  .

    Parameters:
    \itemize{




        \item \code{ ... } : additional parameters passed to httr::GET
    }
    Returns:
        \code{ character }
}
\item{\code{ get_controlled_list_sex }}{

  Get allowed values for the field &#39;Sex&#39; in a specimen document;
  .

    Parameters:
    \itemize{




        \item \code{ ... } : additional parameters passed to httr::GET
    }
    Returns:
        \code{ character }
}
\item{\code{ get_controlled_list_specimen_type_status }}{

  Get allowed values for the field &#39;SpecimenTypeStatus&#39; in a specimen document;
  .

    Parameters:
    \itemize{




        \item \code{ ... } : additional parameters passed to httr::GET
    }
    Returns:
        \code{ character }
}
\item{\code{ get_controlled_list_taxonomic_status }}{

  Get allowed values for the field &#39;TaxonomicStatus&#39; in specimen and taxon documents;
  .

    Parameters:
    \itemize{




        \item \code{ ... } : additional parameters passed to httr::GET
    }
    Returns:
        \code{ character }
}
\item{\code{ get_controlled_lists }}{

  Get the names of fields for which a controlled vocabulary exists;
  Possible values for fields with controlled vocabularies can be queried with metadata/getControlledList/{field}.

    Parameters:
    \itemize{




        \item \code{ ... } : additional parameters passed to httr::GET
    }
    Returns:
        \code{ character }
}
\item{\code{ get_rest_services }}{

  List all available REST services and their parameters;
  Lists end point name, http method, response type, and URL.

    Parameters:
    \itemize{




        \item \code{ ... } : additional parameters passed to httr::GET
    }
    Returns:
        \code{ RestService }
}
\item{\code{ get_setting }}{

  Get the value of an NBA setting;
  All settings can be queried with /metadata/getSettings.

    Parameters:
    \itemize{

        \item \code{ name } : name of setting


        \item \code{ ... } : additional parameters passed to httr::GET
    }
    Returns:
        \code{ list }
}
\item{\code{ get_settings }}{

  List all publicly available configuration settings for the NBA;
  The value of a specific setting can be queried with metadata/getSetting/{name}.

    Parameters:
    \itemize{




        \item \code{ ... } : additional parameters passed to httr::GET
    }
    Returns:
        \code{ list }
}
\item{\code{ get_source_systems }}{

  Get the data sources from which the data was retrieved;
  Returns code and name of all source systems.

    Parameters:
    \itemize{




        \item \code{ ... } : additional parameters passed to httr::GET
    }
    Returns:
        \code{ SourceSystem }
}
}
}

\keyword{datasets}
