% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monomial.r
\docType{class}
\name{Monomial}
\alias{Monomial}
\title{Monomial Class}
\format{R6 class}
\usage{
# Monomial$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for Monomial objects.
}
\section{Fields}{

\describe{
\item{\code{rank}}{character}

\item{\code{name}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Monomial object.

}
\item{\code{$fromList(MonomialList)}}{

  Create Monomial object from list.

}

\item{\code{$toList()}}{

  Get list representation of Monomial.

}
\item{\code{fromJSONString(MonomialJson)}}{

  Create Monomial object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Monomial.

}
}
}

\keyword{datasets}
