% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monomial.r
\docType{data}
\name{Monomial}
\alias{Monomial}
\title{Monomial Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Monomial
}
\description{
Monomial Class
}
\section{Fields}{

\describe{
\item{\code{rank}}{}

\item{\code{name}}{}
}}

\keyword{datasets}
