% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLineString.r
\docType{data}
\name{MultiLineString}
\alias{MultiLineString}
\title{MultiLineString Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultiLineString
}
\description{
MultiLineString Class
}
\section{Fields}{

\describe{
\item{\code{crs}}{}

\item{\code{bbox}}{}

\item{\code{coordinates}}{}
}}

\keyword{datasets}
