% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLineString.r
\docType{data}
\name{MultiLineString}
\alias{MultiLineString}
\title{MultiLineString Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultiLineString
}
\description{
MultiLineString Class
}
\section{Fields}{

\describe{
\item{\code{crs}}{Crs}

\item{\code{bbox}}{list(numeric)}

\item{\code{coordinates}}{list(LngLatAlt)}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor MultiLineString object.

}
\item{\code{$fromList(MultiLineStringList)}}{

  Create MultiLineString object from list.

}

\item{\code{$toList()}}{

  Get list representation of MultiLineString.

}
\item{\code{fromJSONString(MultiLineStringJson)}}{

  Create MultiLineString object from list.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of MultiLineString.

}
}
}

\keyword{datasets}
