% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultimediaClient-wrapper.r
\name{multimedia_find_by_ids}
\alias{multimedia_find_by_ids}
\title{Find multimedia document by ids}
\usage{
multimedia_find_by_ids(ids = NULL, returnType = "data.frame", ...)
}
\arguments{
\item{ids}{ids of multiple multimedia documents, separated by comma, type: character}

\item{returnType}{Either \code{list} or \code{data.frame} (default)}

\item{...}{additional parameters passed to find_by_ids from class nbaR.MultimediaClient}
}
\value{
list or data.frame, as specified by \code{returnType}
}
\description{
This is a wrapper for the method \code{ find_by_ids }
from class \code{ MultimediaClient}.
}
\details{
Given multiple ids, returns a list of multimedia documents
}
\seealso{
Other nbaR.MultimediaClient-wrappers: 
\code{\link{multimedia_count_distinct_values_per_group}()},
\code{\link{multimedia_count_distinct_values}()},
\code{\link{multimedia_count}()},
\code{\link{multimedia_download_query}()},
\code{\link{multimedia_find}()},
\code{\link{multimedia_get_distinct_values_per_group}()},
\code{\link{multimedia_get_distinct_values}()},
\code{\link{multimedia_get_field_info}()},
\code{\link{multimedia_get_paths}()},
\code{\link{multimedia_get_settings}()},
\code{\link{multimedia_get_setting}()},
\code{\link{multimedia_is_operator_allowed}()},
\code{\link{multimedia_query_with_name_resolution}()},
\code{\link{multimedia_query}()}
}
\concept{nbaR.MultimediaClient-wrappers}
