% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultimediaClient-wrapper.r
\name{multimedia_get_field_info}
\alias{multimedia_get_field_info}
\title{Returns extended information for each field of a multimedia document}
\usage{
multimedia_get_field_info(returnType = "data.frame", ...)
}
\arguments{
\item{returnType}{Either \code{list} or \code{data.frame} (default)}

\item{...}{additional parameters passed to get_field_info from class nbaR.MultimediaClient}
}
\value{
scalar
}
\description{
This is a wrapper for the method \code{ get_field_info }
from class \code{ MultimediaClient}.
}
\details{
Info consists of whether the fields is indexed, the ElasticSearch datatype and a list of allowed operators
}
\seealso{
Other nbaR.MultimediaClient-wrappers: 
\code{\link{multimedia_count_distinct_values_per_group}()},
\code{\link{multimedia_count_distinct_values}()},
\code{\link{multimedia_count}()},
\code{\link{multimedia_download_query}()},
\code{\link{multimedia_find_by_ids}()},
\code{\link{multimedia_find}()},
\code{\link{multimedia_get_distinct_values_per_group}()},
\code{\link{multimedia_get_distinct_values}()},
\code{\link{multimedia_get_paths}()},
\code{\link{multimedia_get_settings}()},
\code{\link{multimedia_get_setting}()},
\code{\link{multimedia_is_operator_allowed}()},
\code{\link{multimedia_query_with_name_resolution}()},
\code{\link{multimedia_query}()}
}
\concept{nbaR.MultimediaClient-wrappers}
