% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultimediaClient.r
\docType{data}
\name{MultimediaClient}
\alias{MultimediaClient}
\title{Multimedia operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultimediaClient
}
\description{
nbaR.Multimedia
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

count Get the number of multimedia documents matching a given condition


count_distinct_values Count the distinct number of values that exist for a given field


count_distinct_values_per_group Count the distinct number of field values that exist per the given field to group by


download_query Dynamic download service: Query for multimedia objects and return result as a stream ...


find Find a multimedia document by id


find_by_ids Find multimedia document by ids


get_distinct_values Get all different values that can be found for one field


get_distinct_values_per_group Get all distinct values (and their document count) for the field given divided per distinct value of the field to group by


get_field_info Returns extended information for each field of a multimedia document


get_paths Returns the full path of all fields within a document


get_setting Get the value of an NBA setting


get_settings List all publicly available configuration settings for the NBA


is_operator_allowed Checks if a given operator is allowed for a given field


query Query for multimedia documents

}
}

\keyword{datasets}
