% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiMediaContentIdentification.r
\docType{data}
\name{MultiMediaContentIdentification}
\alias{MultiMediaContentIdentification}
\title{MultiMediaContentIdentification Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultiMediaContentIdentification
}
\description{
MultiMediaContentIdentification Class
}
\section{Fields}{

\describe{
\item{\code{taxonRank}}{}

\item{\code{scientificName}}{}

\item{\code{typeStatus}}{}

\item{\code{dateIdentified}}{}

\item{\code{defaultClassification}}{}

\item{\code{systemClassification}}{}

\item{\code{vernacularNames}}{}

\item{\code{identificationQualifiers}}{}

\item{\code{identifiers}}{}

\item{\code{taxonomicEnrichments}}{}
}}

\keyword{datasets}
