% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiMediaContentIdentification.r
\docType{data}
\name{MultiMediaContentIdentification}
\alias{MultiMediaContentIdentification}
\title{MultiMediaContentIdentification Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultiMediaContentIdentification
}
\description{
MultiMediaContentIdentification Class
}
\section{Fields}{

\describe{
\item{\code{taxonomicEnrichments}}{}
}}

\keyword{datasets}
