% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiMediaGatheringEvent.r
\docType{data}
\name{MultiMediaGatheringEvent}
\alias{MultiMediaGatheringEvent}
\title{MultiMediaGatheringEvent Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultiMediaGatheringEvent
}
\description{
MultiMediaGatheringEvent Class
}
\section{Fields}{

\describe{
\item{\code{bioStratigraphic}}{}
}}

\keyword{datasets}
