% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiMediaObject.r
\docType{data}
\name{MultiMediaObject}
\alias{MultiMediaObject}
\title{MultiMediaObject Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultiMediaObject
}
\description{
MultiMediaObject Class
}
\section{Fields}{

\describe{
\item{\code{associatedTaxon}}{}
}}

\keyword{datasets}
