% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiMediaObject.r
\docType{data}
\name{MultiMediaObject}
\alias{MultiMediaObject}
\title{MultiMediaObject Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultiMediaObject
}
\description{
MultiMediaObject Class
}
\section{Fields}{

\describe{
\item{\code{sourceSystem}}{}

\item{\code{sourceSystemId}}{}

\item{\code{recordURI}}{}

\item{\code{id}}{}

\item{\code{sourceInstitutionID}}{}

\item{\code{sourceID}}{}

\item{\code{owner}}{}

\item{\code{licenseType}}{}

\item{\code{license}}{}

\item{\code{unitID}}{}

\item{\code{collectionType}}{}

\item{\code{title}}{}

\item{\code{caption}}{}

\item{\code{description}}{}

\item{\code{serviceAccessPoints}}{}

\item{\code{type}}{}

\item{\code{taxonCount}}{}

\item{\code{creator}}{}

\item{\code{copyrightText}}{}

\item{\code{associatedSpecimenReference}}{}

\item{\code{associatedTaxonReference}}{}

\item{\code{multiMediaPublic}}{}

\item{\code{subjectParts}}{}

\item{\code{subjectOrientations}}{}

\item{\code{phasesOrStages}}{}

\item{\code{sexes}}{}

\item{\code{gatheringEvents}}{}

\item{\code{identifications}}{}

\item{\code{theme}}{}

\item{\code{associatedSpecimen}}{}

\item{\code{associatedTaxon}}{}
}}

\keyword{datasets}
