% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPoint.r
\docType{class}
\name{MultiPoint}
\alias{MultiPoint}
\title{MultiPoint Class}
\format{R6 class}
\usage{
# MultiPoint$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for MultiPoint objects.
}
\section{Fields}{

\describe{
\item{\code{crs}}{Crs}

\item{\code{bbox}}{list(numeric)}

\item{\code{coordinates}}{list(LngLatAlt)}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor MultiPoint object.

}
\item{\code{$fromList(MultiPointList)}}{

  Create MultiPoint object from list.

}

\item{\code{$toList()}}{

  Get list representation of MultiPoint.

}
\item{\code{fromJSONString(MultiPointJson)}}{

  Create MultiPoint object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of MultiPoint.

}
}
}

\keyword{datasets}
