% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPolygon.r
\docType{data}
\name{MultiPolygon}
\alias{MultiPolygon}
\title{MultiPolygon Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultiPolygon
}
\description{
MultiPolygon Class
}
\section{Fields}{

\describe{
\item{\code{coordinates}}{}
}}

\keyword{datasets}
