% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPolygon.r
\docType{class}
\name{MultiPolygon}
\alias{MultiPolygon}
\title{MultiPolygon Class}
\format{
R6 class

R6 class
}
\usage{
# MultiPolygon$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for MultiPolygon objects.
}
\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor MultiPolygon object.

}
\item{\code{$fromList(MultiPolygonList)}}{

  Create MultiPolygon object from list.

}

\item{\code{$toList()}}{

  Get list representation of MultiPolygon.

}
\item{\code{fromJSONString(MultiPolygonJson)}}{

  Create MultiPolygon object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of MultiPolygon.

}
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{crs}}{Crs}

\item{\code{bbox}}{list(numeric)}

\item{\code{coordinates}}{list(LngLatAlt)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultiPolygon-new}{\code{MultiPolygon$new()}}
\item \href{#method-MultiPolygon-toList}{\code{MultiPolygon$toList()}}
\item \href{#method-MultiPolygon-fromList}{\code{MultiPolygon$fromList()}}
\item \href{#method-MultiPolygon-toJSONString}{\code{MultiPolygon$toJSONString()}}
\item \href{#method-MultiPolygon-fromJSONString}{\code{MultiPolygon$fromJSONString()}}
\item \href{#method-MultiPolygon-print}{\code{MultiPolygon$print()}}
\item \href{#method-MultiPolygon-clone}{\code{MultiPolygon$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiPolygon-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultiPolygon-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiPolygon$new(crs, bbox, coordinates)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiPolygon-toList"></a>}}
\if{latex}{\out{\hypertarget{method-MultiPolygon-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiPolygon$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiPolygon-fromList"></a>}}
\if{latex}{\out{\hypertarget{method-MultiPolygon-fromList}{}}}
\subsection{Method \code{fromList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiPolygon$fromList(MultiPolygonList, typeMapping = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiPolygon-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MultiPolygon-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiPolygon$toJSONString(pretty = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiPolygon-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MultiPolygon-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiPolygon$fromJSONString(MultiPolygonJson, typeMapping = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiPolygon-print"></a>}}
\if{latex}{\out{\hypertarget{method-MultiPolygon-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiPolygon$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiPolygon-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultiPolygon-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiPolygon$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
