% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPolygon.r
\docType{data}
\name{MultiPolygon}
\alias{MultiPolygon}
\title{MultiPolygon Class}
\format{}
\usage{

}
\description{
MultiPolygon Class
}
\section{Fields}{

\describe{
\item{\code{crs}}{Crs}

\item{\code{bbox}}{list(numeric)}

\item{\code{coordinates}}{list(LngLatAlt)}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor MultiPolygon object.

}
\item{\code{$fromList(MultiPolygonList)}}{

  Create MultiPolygon object from list.

}

\item{\code{$toList()}}{

  Get list representation of MultiPolygon.

}
\item{\code{fromJSONString(MultiPolygonJson)}}{

  Create MultiPolygon object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of MultiPolygon.

}
}
}

\keyword{datasets}
