% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NamedArea.r
\docType{data}
\name{NamedArea}
\alias{NamedArea}
\title{NamedArea Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NamedArea
}
\description{
NamedArea Class
}
\section{Fields}{

\describe{
\item{\code{areaClass}}{character}

\item{\code{areaName}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor NamedArea object.

}
\item{\code{$fromList(NamedAreaList)}}{

  Create NamedArea object from list.

}

\item{\code{$toList()}}{

  Get list representation of NamedArea.

}
\item{\code{fromJSONString(NamedAreaJson)}}{

  Create NamedArea object from list.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of NamedArea.

}
}
}

\keyword{datasets}
