% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBA_colours}
\alias{NBA_colours}
\title{NBA colours for every category}
\format{
\subsection{\code{NBA_colours}}{

A vector of category colours:
}
}
\source{
SANBI
}
\usage{
NBA_colours
}
\description{
This data is used to enable users to use the standard colours for their figures
#Threat staus
"Critically Endangered",
"Endangered" ,
"Vulnerable",
"Near Threatened" ,
"Least Concern" ,
"Data Deficient" ,
"Rare" ,
"Extinct" ,
"Extinct in the Wild" ,
"Regionally Extinct",
"Critically Endangered (Possibly Extinct)",
}
\details{
#Protection level
"Well Protected" ,
"Moderately Protected" ,
"Poorly Protected" ,
"Not Protected" ,

#Pressures
"Low" ,
"Medium" ,
"High" ,
"Very high",

"No threats" ,
"Pollution",
"Transportation & service corridors" ,
"Agriculture" ,
"Agriculture and aquaculture",
"Geological events",
"Biological resource use" ,
"Other threats",
"Human intrusions & disturbance" ,
"Human intrusions and disturbance" ,
"Climate change" ,
"Climate change & severe weather",
"Energy production & mining" ,
"Energy production and mining" ,
"Natural system modifications" ,
"Invasive and other problematic species, genes & diseases",
"Invasive and other problematic species, genes and diseases",
"Residential & commercial development" ,
}
\section{Condition}{
"Natural" ,
"Natural / near natural",
"Near natural" ,
"Moderately modified" ,
"Heavily / intensively modified" ,
"Permanently / irreversibly modified" ,
}

\section{Responses}{
"No response" ,
"Some kind of response" ,
"Gazetted" ,
"Signed off" ,
}

\section{Priority areas}{
"Land-based Protected Areas",
"Marine Protected Areas",
"Critical Biodiversity Areas",
"Ecologically Sensitive Areas" ,

#priority ecosystems

"Critically Endangered & Not Protected",
"Critically Endangered & Poorly Protected",
"Endangered & Not Protected",
"Endangered & Poorly Protected",
}

\section{Built up areas}{
"Cropland",
"Plantation",
"Built up",
"Mine",
"Artificial waterbody",
"Landcover Natural"
}

\keyword{datasets}
