% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{NBA_colr_tbl}
\alias{NBA_colr_tbl}
\title{Coloured table}
\usage{
NBA_colr_tbl(DF, COL)
}
\arguments{
\item{DF}{The data frame that contains the data}

\item{COL}{the column containing the categories}
}
\description{
A basic table with a purple heading bar and coloured blocks around the
categories
This data should be in the same fomat as for the plot functions
}
\examples{
library(dplyr)
library(tidyr)

tbl <- NBA_example_pro_data \%>\%
 pivot_longer(2:5, names_to = "protection_level") \%>\%
 NBA_colr_tbl(COL = protection_level)

tbl

}
