% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{NBA_plot}
\alias{NBA_plot}
\title{NBA plot function}
\usage{
NBA_plot(
  DF,
  GROUPS,
  COLS,
  CHRT = c("bar", "donut"),
  NUM = FALSE,
  LAB,
  GRP = TRUE,
  SAVE
)
}
\arguments{
\item{DF}{The data frame that contains the information}

\item{GROUPS}{The categorized variables (ecosystem functional group, taxa, etc)}

\item{COLS}{The categories to describe the variables (protection level, threat status, condition, etc). You can use any tidyselect method to select these columns e.g. 2:4/ Endangered:Vulnerable/ c(Well Protected,Moderately Protected,Poorly Protected,Not Protected) etc}

\item{CHRT}{A choice of either "bar" or "donut" plot}

\item{NUM}{A choice to show numbers in the plot, False to show no numbers}

\item{LAB}{The x axis label of the plot}

\item{GRP}{A choice of whether or not to plot the donut graphs by group, TRUE will plot a donut plot for each group.}

\item{SAVE}{The name of the output file that will be saved to the output folder. If you do not have an outputs folder you will be prompted to make one.}
}
\value{
Returns a plot
}
\description{
NBA plots for protection level, threat status, and condition.
The function expects that the groups are in a column, with one
group per row, and the protection level, threat status, or condition
categories are the headings of each column. Please look at the example
data NBA_example_bar_plot and NBA_example_donut_plot to see the correct
structure for the data. Please note that both of these datasets have the
same structure, whether it will be used to make a bar or donut plot is
irrelevant.
This function will plot the data as either a bar or donut plot
depending on what you require. You can also decide if you want the donut
plot to be split by ecosystem functional group or not and choose if
you want the number of ecosystems to be displayed within the plot.
}
\details{
The name of the groups column is irrelevant, but the categories must be
spelled correctly (there is a list of the standard spellings/ cases
of NBA categories named NBA_categories in this package,
which can be accessed for reference).
}
\examples{
#bar_plot <- NBA_plot(Fig1a,
#                   OVERALL_types,
#                   2:5,
#                   CHRT = "bar",
#                   NUM = TRUE,
#                   LAB = "Percentage of ecosystem types",
#                   SAVE = "Fig1a")
#
#donut_plot <- NBA_plot(Fig61mapinset,
#                        OVERALL_types,
#                         COLS = 2:5,
#                         NUM = T,
#                         GRP = T,
#                         CHRT = "donut",
#                         LAB = "Protection level",
#                         SAVE = "Fig61mapinset")


}
