% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{nba_plot_RLI}
\alias{nba_plot_RLI}
\title{NBA Redlist index plot function}
\usage{
nba_plot_RLI(
  DF,
  YEAR,
  RLI,
  MIN,
  MAX,
  GROUP = NULL,
  summarise_by_year = TRUE,
  SAVE = NULL
)
}
\arguments{
\item{DF}{The data frame that contains the information on the Red List Index}

\item{YEAR}{The years}

\item{RLI}{The Red List Index}

\item{MIN}{The minimum values}

\item{MAX}{The maximum values}

\item{GROUP}{A choice to group the plot, if a column name is supplied will groupd, if left NULL will not group}

\item{summarise_by_year}{\strong{must be added}}

\item{SAVE}{The name of the output file that will be saved to the output folder. If you do not have an outputs folder you will be prompted to make one.}
}
\value{
Returns a RLI plot
}
\description{
This function will create a line graph indicating the Red list index per period/year.
Additionally, depending on the graph requirements it further creates a multiple line graph
grouping the plot according to species, and creates a buffer parallel to the RLI
line using the minimum and maximum values.
}
\details{
Note: The graph can be used on other datasets as well, as long as it contains the same
data structure as the Red List Index.
There is an example dataset (NBA_example_RLI_data), available for users
to reference if needed.
}
\examples{
#RLI_plot <- nba_plot_RLI(NBA_example_RLI_data,
#Years,
#RLI,
#MIN,
#MAX)

#RLI_plot


}
