% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{nba_tbl_colr}
\alias{nba_tbl_colr}
\title{Coloured table}
\usage{
nba_tbl_colr(DF, COL)
}
\arguments{
\item{DF}{The data frame that contains the data}

\item{COL}{the column containing the categories}
}
\description{
A basic table with a purple heading bar and coloured blocks around the
categories
This data should be in the same fomat as for the plot functions
}
\examples{
library(dplyr)
library(tidyr)

tbl <- NBA_example_pro_data \%>\%
 pivot_longer(2:5, names_to = "protection_level") \%>\%
 nba_tbl_colr(COL = protection_level)

tbl

}
