% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{nba_tbl_comb}
\alias{nba_tbl_comb}
\title{Table of protection level of threatened ecosystems.}
\usage{
nba_tbl_comb(DF, GROUP, THR, PRO)
}
\arguments{
\item{DF}{The data frame that contains the data}

\item{GROUP}{The column that contains the name of the variable (the ecosystems or taxa names)}

\item{THR}{The column name of the threat statuses}

\item{PRO}{The column name of the protection levels}
}
\description{
Function to style a table with threat status on the horizontal rows
and protection level on the vertical columns, with number of ecosystems
or taxa that share those categories, with total and percentage çolumns added
}
\examples{

thr_pro_tbl <- NBA_example_map_data |>
sf::st_drop_geometry() |>
nba_tbl_comb(GROUP = P_EcosysType,
THR = threat_status,
PRO = protection_level)

thr_pro_tbl

}
