% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{NBA_tbl_theme}
\alias{NBA_tbl_theme}
\title{gt table theme}
\usage{
NBA_tbl_theme(GT_TBL)
}
\arguments{
\item{GT_TBL}{The data frame that contains the data}
}
\description{
Function to apply gt styling to table
}
\examples{
library(gt)
library(dplyr)


gt_tbl <- NBA_example_thr_data \%>\%
 gt() \%>\%
NBA_tbl_theme()

gt_tbl

}
