% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{NBA_theme}
\alias{NBA_theme}
\title{NBA_theme}
\usage{
NBA_theme()
}
\value{
Returns a NBA_theme
}
\description{
This function will create a theme for all ggplot graphs to align them to the
NBA "look". All NBA functions already use this theme.
}
\examples{
#library(ggplot2)

#gg_plot <- ggplot(NBA_example_pro_data, aes(x = `OVERALL types`, y = `Well Protected`))+
#  ggplot2::geom_point()+
#  NBA_theme()

#gg_plot

}
