% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Organization.r
\docType{data}
\name{Organization}
\alias{Organization}
\title{Organization Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Organization
}
\description{
Organization Class
}
\section{Fields}{

\describe{
\item{\code{agentText}}{character}

\item{\code{name}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Organization object.

}
\item{\code{$fromList(OrganizationList)}}{

  Create Organization object from list.

}

\item{\code{$toList()}}{

  Get list representation of Organization.

}
\item{\code{fromJSONString(OrganizationJson)}}{

  Create Organization object from list.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Organization.

}
}
}

\keyword{datasets}
