% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Organization.r
\docType{class}
\name{Organization}
\alias{Organization}
\title{Organization Class}
\format{R6 class}
\usage{
# Organization$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for Organization objects.
}
\section{Fields}{

\describe{
\item{\code{agentText}}{character}

\item{\code{name}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Organization object.

}
\item{\code{$fromList(OrganizationList)}}{

  Create Organization object from list.

}

\item{\code{$toList()}}{

  Get list representation of Organization.

}
\item{\code{fromJSONString(OrganizationJson)}}{

  Create Organization object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Organization.

}
}
}

\keyword{datasets}
