% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path.r
\docType{class}
\name{Path}
\alias{Path}
\title{Path Class}
\format{R6 class}
\usage{
# Path$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for Path objects.
}
\section{Fields}{

\describe{
\item{\code{purePath}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Path object.

}
\item{\code{$fromList(PathList)}}{

  Create Path object from list.

}

\item{\code{$toList()}}{

  Get list representation of Path.

}
\item{\code{fromJSONString(PathJson)}}{

  Create Path object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Path.

}
}
}

\keyword{datasets}
