% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path.r
\docType{data}
\name{Path}
\alias{Path}
\title{Path Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Path
}
\description{
Path Class
}
\section{Fields}{

\describe{
\item{\code{purePath}}{}
}}

\keyword{datasets}
