% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path.r
\docType{data}
\name{Path}
\alias{Path}
\title{Path Class}
\format{}
\usage{

}
\description{
Path Class
}
\section{Fields}{

\describe{
\item{\code{purePath}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Path object.

}
\item{\code{$fromList(PathList)}}{

  Create Path object from list.

}

\item{\code{$toList()}}{

  Get list representation of Path.

}
\item{\code{fromJSONString(PathJson)}}{

  Create Path object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Path.

}
}
}

\keyword{datasets}
