% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Person.r
\docType{class}
\name{Person}
\alias{Person}
\title{Person Class}
\format{R6 class}
\usage{
# Person$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for Person objects.
}
\section{Fields}{

\describe{
\item{\code{agentText}}{character}

\item{\code{fullName}}{character}

\item{\code{organization}}{Organization}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Person object.

}
\item{\code{$fromList(PersonList)}}{

  Create Person object from list.

}

\item{\code{$toList()}}{

  Get list representation of Person.

}
\item{\code{fromJSONString(PersonJson)}}{

  Create Person object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Person.

}
}
}

\keyword{datasets}
