% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Person.r
\docType{data}
\name{Person}
\alias{Person}
\title{Person Class}
\format{}
\usage{

}
\description{
Person Class
}
\section{Fields}{

\describe{
\item{\code{agentText}}{character}

\item{\code{fullName}}{character}

\item{\code{organization}}{Organization}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Person object.

}
\item{\code{$fromList(PersonList)}}{

  Create Person object from list.

}

\item{\code{$toList()}}{

  Get list representation of Person.

}
\item{\code{fromJSONString(PersonJson)}}{

  Create Person object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Person.

}
}
}

\keyword{datasets}
