% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Point.r
\docType{data}
\name{Point}
\alias{Point}
\title{Point Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Point
}
\description{
Point Class
}
\section{Fields}{

\describe{
\item{\code{crs}}{}

\item{\code{bbox}}{}

\item{\code{coordinates}}{}
}}

\keyword{datasets}
