% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Point.r
\docType{class}
\name{Point}
\alias{Point}
\title{Point Class}
\format{R6 class}
\usage{
# Point$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for Point objects.
}
\section{Fields}{

\describe{
\item{\code{crs}}{Crs}

\item{\code{bbox}}{list(numeric)}

\item{\code{coordinates}}{LngLatAlt}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Point object.

}
\item{\code{$fromList(PointList)}}{

  Create Point object from list.

}

\item{\code{$toList()}}{

  Get list representation of Point.

}
\item{\code{fromJSONString(PointJson)}}{

  Create Point object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Point.

}
}
}

\keyword{datasets}
