% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polygon.r
\docType{data}
\name{Polygon}
\alias{Polygon}
\title{Polygon Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Polygon
}
\description{
Polygon Class
}
\section{Fields}{

\describe{
\item{\code{coordinates}}{}
}}

\keyword{datasets}
