% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polygon.r
\docType{data}
\name{Polygon}
\alias{Polygon}
\title{Polygon Class}
\format{}
\usage{

}
\description{
Polygon Class
}
\section{Fields}{

\describe{
\item{\code{crs}}{Crs}

\item{\code{bbox}}{list(numeric)}

\item{\code{coordinates}}{list(LngLatAlt)}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Polygon object.

}
\item{\code{$fromList(PolygonList)}}{

  Create Polygon object from list.

}

\item{\code{$toList()}}{

  Get list representation of Polygon.

}
\item{\code{fromJSONString(PolygonJson)}}{

  Create Polygon object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Polygon.

}
}
}

\keyword{datasets}
