% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{pro_donut_plot_test}
\alias{pro_donut_plot_test}
\title{Protection level donut plot}
\usage{
pro_donut_plot_test(DF, FILL)
}
\arguments{
\item{DF}{The data frame that contains the information on protection level}

\item{FILL}{The protection level categories}
}
\value{
Returns a bar donut plot of protection level
}
\description{
Donut plot for the protection level.
This function requires a dataframe that contains a columns of the
the ecosystem function groups or taxa, ecosystem types or species, and their protection level.
}
\details{
The function will group the ecosystem types or species by the functional
group or taxa and protection level, and calculate the number of ecosystem types or species within each functional
group or taxa with that protection level.
It will also create the colours and axis titles.
}
\examples{
#test <- pro_donut_plot(mydata, ecosystem_functional_grps)

}
