% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{prot_efg}
\alias{prot_efg}
\title{Protection level bar plot}
\usage{
prot_efg(DF, X, Y, FILL, COUNTS)
}
\arguments{
\item{DF}{The data frame that contains the information on protection level}

\item{X}{The groups}

\item{Y}{The protection level percentages}

\item{FILL}{The protection level categories}

\item{COUNTS}{The frequency counts of the number of ecosystems within each protection level}
}
\value{
Returns a bar graph of protection level
}
\description{
Horizontal barplots for the protection level of ecosystem function groups (efgs)
}
\examples{
#test <- prot_efg(mydata, ecosystem_functional_grps, percentages, threat_status)

}
