% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryCondition.r
\docType{class}
\name{QueryCondition}
\alias{QueryCondition}
\title{QueryCondition Class}
\format{
R6 class

R6 class
}
\usage{
# QueryCondition$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for QueryCondition objects.
}
\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor QueryCondition object.

}
\item{\code{$fromList(QueryConditionList)}}{

  Create QueryCondition object from list.

}

\item{\code{$toList()}}{

  Get list representation of QueryCondition.

}
\item{\code{fromJSONString(QueryConditionJson)}}{

  Create QueryCondition object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of QueryCondition.

}
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{not}}{character}

\item{\code{field}}{character}

\item{\code{operator}}{character}

\item{\code{value}}{list}

\item{\code{and}}{list(QueryCondition)}

\item{\code{or}}{list(QueryCondition)}

\item{\code{constantScore}}{logical}

\item{\code{boost}}{numeric}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QueryCondition-new}{\code{QueryCondition$new()}}
\item \href{#method-QueryCondition-toList}{\code{QueryCondition$toList()}}
\item \href{#method-QueryCondition-fromList}{\code{QueryCondition$fromList()}}
\item \href{#method-QueryCondition-toJSONString}{\code{QueryCondition$toJSONString()}}
\item \href{#method-QueryCondition-fromJSONString}{\code{QueryCondition$fromJSONString()}}
\item \href{#method-QueryCondition-print}{\code{QueryCondition$print()}}
\item \href{#method-QueryCondition-clone}{\code{QueryCondition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryCondition-new"></a>}}
\if{latex}{\out{\hypertarget{method-QueryCondition-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryCondition$new(not, field, operator, value, and, or, constantScore, boost)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryCondition-toList"></a>}}
\if{latex}{\out{\hypertarget{method-QueryCondition-toList}{}}}
\subsection{Method \code{toList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryCondition$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryCondition-fromList"></a>}}
\if{latex}{\out{\hypertarget{method-QueryCondition-fromList}{}}}
\subsection{Method \code{fromList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryCondition$fromList(QueryConditionList, typeMapping = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryCondition-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QueryCondition-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryCondition$toJSONString(pretty = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryCondition-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QueryCondition-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryCondition$fromJSONString(QueryConditionJson, typeMapping = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryCondition-print"></a>}}
\if{latex}{\out{\hypertarget{method-QueryCondition-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryCondition$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryCondition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QueryCondition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryCondition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
