% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryCondition.r
\docType{data}
\name{QueryCondition}
\alias{QueryCondition}
\title{QueryCondition Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
QueryCondition
}
\description{
QueryCondition Class
}
\section{Fields}{

\describe{
\item{\code{not}}{character}

\item{\code{field}}{character}

\item{\code{operator}}{character}

\item{\code{value}}{list}

\item{\code{and}}{list(QueryCondition)}

\item{\code{or}}{list(QueryCondition)}

\item{\code{constantScore}}{logical}

\item{\code{boost}}{numeric}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor QueryCondition object.

}
\item{\code{$fromList(QueryConditionList)}}{

  Create QueryCondition object from list.

}

\item{\code{$toList()}}{

  Get list representation of QueryCondition.

}
\item{\code{fromJSONString(QueryConditionJson)}}{

  Create QueryCondition object from list.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of QueryCondition.

}
}
}

\keyword{datasets}
