% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryResult.r
\docType{data}
\name{QueryResult}
\alias{QueryResult}
\title{QueryResult Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
QueryResult
}
\description{
QueryResult Class
}
\section{Fields}{

\describe{
\item{\code{resultSet}}{}
}}

\keyword{datasets}
