% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryResult.r
\docType{class}
\name{QueryResult}
\alias{QueryResult}
\title{QueryResult Class}
\format{R6 class}
\usage{
# QueryResult$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for QueryResult objects.
}
\section{Fields}{

\describe{
\item{\code{totalSize}}{integer}

\item{\code{resultSet}}{list(QueryResultItemObject)}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor QueryResult object.

}
\item{\code{$fromList(QueryResultList)}}{

  Create QueryResult object from list.

}

\item{\code{$toList()}}{

  Get list representation of QueryResult.

}
\item{\code{fromJSONString(QueryResultJson)}}{

  Create QueryResult object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of QueryResult.

}
}
}

\keyword{datasets}
