% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryResultItem.r
\docType{data}
\name{QueryResultItem}
\alias{QueryResultItem}
\title{QueryResultItem Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
QueryResultItem
}
\description{
Contains an object returned from a QuerySpec query. Holds an item and a relevance score.
}
\section{Fields}{

\describe{
\item{\code{score}}{numeric}

\item{\code{item}}{list}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor QueryResultItem object.

}
\item{\code{$fromList(QueryResultItemList)}}{

  Create QueryResultItem object from list.

}

\item{\code{$toList()}}{

  Get list representation of QueryResultItem.

}
\item{\code{fromJSONString(QueryResultItemJson)}}{

  Create QueryResultItem object from list.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of QueryResultItem.

}
}
}

\keyword{datasets}
