% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryResultItem.r
\docType{data}
\name{QueryResultItem}
\alias{QueryResultItem}
\title{QueryResultItem Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
QueryResultItem
}
\description{
QueryResultItem Class
}
\section{Fields}{

\describe{
\item{\code{item}}{}
}}

\keyword{datasets}
