% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryResultItemObject.r
\docType{data}
\name{QueryResultItemObject}
\alias{QueryResultItemObject}
\title{QueryResultItemObject Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
QueryResultItemObject
}
\description{
QueryResultItemObject Class
}
\section{Fields}{

\describe{
\item{\code{item}}{}
}}

\keyword{datasets}
