% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuerySpec.r
\docType{data}
\name{QuerySpec}
\alias{QuerySpec}
\title{QuerySpec Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
QuerySpec
}
\description{
QuerySpec Class
}
\section{Fields}{

\describe{
\item{\code{size}}{}
}}

\keyword{datasets}
