% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reference.r
\docType{data}
\name{Reference}
\alias{Reference}
\title{Reference Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Reference
}
\description{
Reference Class
}
\section{Fields}{

\describe{
\item{\code{publicationDate}}{}
}}

\keyword{datasets}
