% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reference.r
\docType{data}
\name{Reference}
\alias{Reference}
\title{Reference Class}
\format{}
\usage{

}
\description{
Reference Class
}
\section{Fields}{

\describe{
\item{\code{titleCitation}}{character}

\item{\code{citationDetail}}{character}

\item{\code{uri}}{character}

\item{\code{author}}{Person}

\item{\code{publicationDate}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Reference object.

}
\item{\code{$fromList(ReferenceList)}}{

  Create Reference object from list.

}

\item{\code{$toList()}}{

  Get list representation of Reference.

}
\item{\code{fromJSONString(ReferenceJson)}}{

  Create Reference object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Reference.

}
}
}

\keyword{datasets}
