% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestService.r
\docType{class}
\name{RestService}
\alias{RestService}
\title{RestService Class}
\format{R6 class}
\usage{
# RestService$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for RestService objects.
}
\section{Fields}{

\describe{
\item{\code{endPoint}}{character}

\item{\code{method}}{character}

\item{\code{consumes}}{character}

\item{\code{produces}}{character}

\item{\code{url}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor RestService object.

}
\item{\code{$fromList(RestServiceList)}}{

  Create RestService object from list.

}

\item{\code{$toList()}}{

  Get list representation of RestService.

}
\item{\code{fromJSONString(RestServiceJson)}}{

  Create RestService object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of RestService.

}
}
}

\keyword{datasets}
