% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{RLI_plot}
\alias{RLI_plot}
\title{RLI_plot}
\usage{
RLI_plot(DF, YEAR, RLI, min, max, GRP = FALSE)
}
\arguments{
\item{DF}{The data frame that contains the information on protection level}

\item{YEAR}{The years}

\item{RLI}{The Redlist index}

\item{min}{The minimum values}

\item{max}{The maximum values}

\item{GRP}{A choice to group the plot, TRUE will group if, FALSE will not.}
}
\value{
Returns a RLI plot
}
\description{
This function will create a Redlist index plot.
There is an example dataset (NBA_example_RLI_plot), available for users
to refernce if needed.
}
\examples{
#test(Fig6, Years, RLI, min, max)

}
