% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{RLI_plot}
\alias{RLI_plot}
\title{RLI_plot}
\usage{
RLI_plot(DF, YEAR, RLI, min, max, GRP = FALSE)
}
\arguments{
\item{DF}{The data frame that contains the information on the Red List Index}

\item{YEAR}{The years}

\item{RLI}{The Red List Index}

\item{min}{The minimum values}

\item{max}{The maximum values}

\item{GRP}{A choice to group the plot, TRUE will group if, FALSE will not.}
}
\value{
Returns a RLI plot
}
\description{
This function will create a line graph indicating the Red list index per period/year.
Additionally, depending on the graph requirements it further creates a multiple line graph
grouping the plot according to species,
and creates a buffer parallel to the RLI line using the minimum and maximum values.
Note: The graph can be used on other datasets as well, as long as it contains the same
data structure as the Red List Index.
There is an example dataset (NBA_example_RLI_plot), available for users
to reference if needed.
}
\examples{
#test(Fig6, Years, RLI, min, max)

}
