% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificName.r
\docType{data}
\name{ScientificName}
\alias{ScientificName}
\title{ScientificName Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ScientificName
}
\description{
ScientificName Class
}
\section{Fields}{

\describe{
\item{\code{experts}}{}
}}

\keyword{datasets}
