% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServiceAccessPoint.r
\docType{class}
\name{ServiceAccessPoint}
\alias{ServiceAccessPoint}
\title{ServiceAccessPoint Class}
\format{R6 class}
\usage{
# ServiceAccessPoint$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for ServiceAccessPoint objects.
}
\section{Fields}{

\describe{
\item{\code{accessUri}}{character}

\item{\code{format}}{character}

\item{\code{variant}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor ServiceAccessPoint object.

}
\item{\code{$fromList(ServiceAccessPointList)}}{

  Create ServiceAccessPoint object from list.

}

\item{\code{$toList()}}{

  Get list representation of ServiceAccessPoint.

}
\item{\code{fromJSONString(ServiceAccessPointJson)}}{

  Create ServiceAccessPoint object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of ServiceAccessPoint.

}
}
}

\keyword{datasets}
