% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServiceAccessPoint.r
\docType{data}
\name{ServiceAccessPoint}
\alias{ServiceAccessPoint}
\title{ServiceAccessPoint Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ServiceAccessPoint
}
\description{
ServiceAccessPoint Class
}
\section{Fields}{

\describe{
\item{\code{accessUri}}{character}

\item{\code{format}}{character}

\item{\code{variant}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor ServiceAccessPoint object.

}
\item{\code{$fromList(ServiceAccessPointList)}}{

  Create ServiceAccessPoint object from list.

}

\item{\code{$toList()}}{

  Get list representation of ServiceAccessPoint.

}
\item{\code{fromJSONString(ServiceAccessPointJson)}}{

  Create ServiceAccessPoint object from list.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of ServiceAccessPoint.

}
}
}

\keyword{datasets}
