% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SortField.r
\docType{data}
\name{SortField}
\alias{SortField}
\title{SortField Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SortField
}
\description{
SortField Class
}
\section{Fields}{

\describe{
\item{\code{path}}{character}

\item{\code{sortOrder}}{character}

\item{\code{ascending}}{logical}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor SortField object.

}
\item{\code{$fromList(SortFieldList)}}{

  Create SortField object from list.

}

\item{\code{$toList()}}{

  Get list representation of SortField.

}
\item{\code{fromJSONString(SortFieldJson)}}{

  Create SortField object from list.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of SortField.

}
}
}

\keyword{datasets}
