% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SortField.r
\docType{data}
\name{SortField}
\alias{SortField}
\title{SortField Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SortField
}
\description{
SortField Class
}
\section{Fields}{

\describe{
\item{\code{path}}{}

\item{\code{sortOrder}}{}

\item{\code{ascending}}{}
}}

\keyword{datasets}
