% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceSystem.r
\docType{data}
\name{SourceSystem}
\alias{SourceSystem}
\title{SourceSystem Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SourceSystem
}
\description{
SourceSystem Class
}
\section{Fields}{

\describe{
\item{\code{code}}{character}

\item{\code{name}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor SourceSystem object.

}
\item{\code{$fromList(SourceSystemList)}}{

  Create SourceSystem object from list.

}

\item{\code{$toList()}}{

  Get list representation of SourceSystem.

}
\item{\code{fromJSONString(SourceSystemJson)}}{

  Create SourceSystem object from list.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of SourceSystem.

}
}
}

\keyword{datasets}
