% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceSystem.r
\docType{data}
\name{SourceSystem}
\alias{SourceSystem}
\title{SourceSystem Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SourceSystem
}
\description{
SourceSystem Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}
}}

\keyword{datasets}
