% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Specimen.r
\docType{data}
\name{Specimen}
\alias{Specimen}
\title{Specimen Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Specimen
}
\description{
Specimen Class
}
\section{Fields}{

\describe{
\item{\code{theme}}{}
}}

\keyword{datasets}
