% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Specimen.r
\docType{data}
\name{Specimen}
\alias{Specimen}
\title{Specimen Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Specimen
}
\description{
Specimen Class
}
\section{Fields}{

\describe{
\item{\code{sourceSystem}}{}

\item{\code{sourceSystemId}}{}

\item{\code{recordURI}}{}

\item{\code{id}}{}

\item{\code{unitID}}{}

\item{\code{unitGUID}}{}

\item{\code{collectorsFieldNumber}}{}

\item{\code{assemblageID}}{}

\item{\code{sourceInstitutionID}}{}

\item{\code{sourceID}}{}

\item{\code{owner}}{}

\item{\code{licenseType}}{}

\item{\code{license}}{}

\item{\code{recordBasis}}{}

\item{\code{kindOfUnit}}{}

\item{\code{collectionType}}{}

\item{\code{sex}}{}

\item{\code{phaseOrStage}}{}

\item{\code{title}}{}

\item{\code{notes}}{}

\item{\code{preparationType}}{}

\item{\code{numberOfSpecimen}}{}

\item{\code{fromCaptivity}}{}

\item{\code{objectPublic}}{}

\item{\code{multiMediaPublic}}{}

\item{\code{acquiredFrom}}{}

\item{\code{gatheringEvent}}{}

\item{\code{identifications}}{}

\item{\code{associatedMultiMediaUris}}{}

\item{\code{theme}}{}

\item{\code{associatedMultiMediaObjects}}{}
}}

\keyword{datasets}
