% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Specimen.r
\docType{data}
\name{Specimen}
\alias{Specimen}
\title{Specimen Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Specimen
}
\description{
This class models a Specimen record. Specimen records constitute the core of data served by the NBA. They can be unambiguously distinguished by the field unitID. Further fields include (among others) information about Finding place, identification, and multimedia content.
}
\section{Fields}{

\describe{
\item{\code{sourceSystem}}{SourceSystem}

\item{\code{sourceSystemId}}{character}

\item{\code{recordURI}}{character}

\item{\code{id}}{character}

\item{\code{unitID}}{character}

\item{\code{unitGUID}}{character}

\item{\code{collectorsFieldNumber}}{character}

\item{\code{assemblageID}}{character}

\item{\code{sourceInstitutionID}}{character}

\item{\code{sourceID}}{character}

\item{\code{previousSourceID}}{list(character)}

\item{\code{owner}}{character}

\item{\code{licenseType}}{character}

\item{\code{license}}{character}

\item{\code{recordBasis}}{character}

\item{\code{kindOfUnit}}{character}

\item{\code{collectionType}}{character}

\item{\code{sex}}{character}

\item{\code{phaseOrStage}}{character}

\item{\code{title}}{character}

\item{\code{notes}}{character}

\item{\code{preparationType}}{character}

\item{\code{previousUnitsText}}{character}

\item{\code{numberOfSpecimen}}{integer}

\item{\code{fromCaptivity}}{logical}

\item{\code{objectPublic}}{logical}

\item{\code{multiMediaPublic}}{logical}

\item{\code{acquiredFrom}}{Agent}

\item{\code{gatheringEvent}}{GatheringEvent}

\item{\code{identifications}}{list(SpecimenIdentification)}

\item{\code{associatedMultiMediaUris}}{list(ServiceAccessPoint)}

\item{\code{theme}}{list(character)}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Specimen object.

}
\item{\code{$fromList(SpecimenList)}}{

  Create Specimen object from list.

}

\item{\code{$toList()}}{

  Get list representation of Specimen.

}
\item{\code{fromJSONString(SpecimenJson)}}{

  Create Specimen object from list.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Specimen.

}
}
}

\keyword{datasets}
