% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecimenClient.r
\docType{data}
\name{SpecimenClient}
\alias{SpecimenClient}
\title{Specimen operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SpecimenClient
}
\description{
nbaR.Specimen
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

count Get the number of specimens matching a given condition


count_distinct_values Count the distinct number of values that exist for a given field


count_distinct_values_per_group Count the distinct number of field values that exist per the given field to group by


download_query Dynamic download service: Query for specimens and return result as a stream ...


dwca_get_data_set Download dataset as Darwin Core Archive File


dwca_get_data_set_names Retrieve the names of all available datasets


dwca_query Dynamic download service: Query for specimens and return result as Darwin Core Archive File


exists Returns whether or not a unitID for a specimen exists


find Find a specimen by id


find_by_ids Find specimens by ids


find_by_unit_id Find a specimen by unitID


get_distinct_values Get all different values that exist for a field


get_distinct_values_per_group Get all distinct values (and their document count) for the field given divided per distinct value of the field to group by


get_field_info Returns extended information for each field of a specimen document


get_ids_in_collection Retrieve all ids within a &#39;special collection&#39; of specimens


get_named_collections Retrieve the names of all &#39;special collections&#39; of specimens


get_paths Returns the full path of all fields within a document


get_setting Get the value of an NBA setting


get_settings List all publicly available configuration settings for the NBA


group_by_scientific_name Aggregates Taxon and Specimen documents according to their scientific names


is_operator_allowed Checks if a given operator is allowed for a given field


query Query for specimens

}
}

\keyword{datasets}
