% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecimenClient.r
\docType{data}
\name{SpecimenClient}
\alias{SpecimenClient}
\title{Specimen operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SpecimenClient
}
\description{
nbaR.Specimen
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

query Query for specimens

}
}

\keyword{datasets}
