% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecimenClient.r
\docType{data}
\name{SpecimenClient}
\alias{SpecimenClient}
\title{Specimen operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SpecimenClient
}
\description{
Client class for nbaR.Specimen objects
}
\section{Fields}{

\describe{
\item{\code{basePath}}{Stores url path of the request.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\keyword{datasets}
