% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecimenIdentification.r
\docType{data}
\name{SpecimenIdentification}
\alias{SpecimenIdentification}
\title{SpecimenIdentification Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SpecimenIdentification
}
\description{
SpecimenIdentification Class
}
\section{Fields}{

\describe{
\item{\code{taxonRank}}{}

\item{\code{scientificName}}{}

\item{\code{typeStatus}}{}

\item{\code{dateIdentified}}{}

\item{\code{defaultClassification}}{}

\item{\code{systemClassification}}{}

\item{\code{vernacularNames}}{}

\item{\code{identificationQualifiers}}{}

\item{\code{identifiers}}{}

\item{\code{taxonomicEnrichments}}{}

\item{\code{preferred}}{}

\item{\code{verificationStatus}}{}

\item{\code{rockType}}{}

\item{\code{associatedFossilAssemblage}}{}

\item{\code{rockMineralUsage}}{}

\item{\code{associatedMineralName}}{}

\item{\code{remarks}}{}
}}

\keyword{datasets}
