% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryScientificName.r
\docType{data}
\name{SummaryScientificName}
\alias{SummaryScientificName}
\title{SummaryScientificName Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SummaryScientificName
}
\description{
SummaryScientificName Class
}
\section{Fields}{

\describe{
\item{\code{fullScientificName}}{character}

\item{\code{taxonomicStatus}}{character}

\item{\code{genusOrMonomial}}{character}

\item{\code{subgenus}}{character}

\item{\code{specificEpithet}}{character}

\item{\code{infraspecificEpithet}}{character}

\item{\code{authorshipVerbatim}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor SummaryScientificName object.

}
\item{\code{$fromList(SummaryScientificNameList)}}{

  Create SummaryScientificName object from list.

}

\item{\code{$toList()}}{

  Get list representation of SummaryScientificName.

}
\item{\code{fromJSONString(SummaryScientificNameJson)}}{

  Create SummaryScientificName object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of SummaryScientificName.

}
}
}

\keyword{datasets}
