% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryScientificName.r
\docType{class}
\name{SummaryScientificName}
\alias{SummaryScientificName}
\title{SummaryScientificName Class}
\format{R6 class}
\usage{
# SummaryScientificName$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for SummaryScientificName objects.
}
\section{Fields}{

\describe{
\item{\code{fullScientificName}}{character}

\item{\code{taxonomicStatus}}{character}

\item{\code{genusOrMonomial}}{character}

\item{\code{subgenus}}{character}

\item{\code{specificEpithet}}{character}

\item{\code{infraspecificEpithet}}{character}

\item{\code{authorshipVerbatim}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor SummaryScientificName object.

}
\item{\code{$fromList(SummaryScientificNameList)}}{

  Create SummaryScientificName object from list.

}

\item{\code{$toList()}}{

  Get list representation of SummaryScientificName.

}
\item{\code{fromJSONString(SummaryScientificNameJson)}}{

  Create SummaryScientificName object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of SummaryScientificName.

}
}
}

\keyword{datasets}
