% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryScientificName.r
\docType{data}
\name{SummaryScientificName}
\alias{SummaryScientificName}
\title{SummaryScientificName Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SummaryScientificName
}
\description{
SummaryScientificName Class
}
\section{Fields}{

\describe{
\item{\code{fullScientificName}}{}

\item{\code{taxonomicStatus}}{}

\item{\code{genusOrMonomial}}{}

\item{\code{subgenus}}{}

\item{\code{specificEpithet}}{}

\item{\code{infraspecificEpithet}}{}

\item{\code{authorshipVerbatim}}{}
}}

\keyword{datasets}
