% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarySourceSystem.r
\docType{class}
\name{SummarySourceSystem}
\alias{SummarySourceSystem}
\title{SummarySourceSystem Class}
\format{R6 class}
\usage{
# SummarySourceSystem$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for SummarySourceSystem objects.
}
\section{Fields}{

\describe{
\item{\code{code}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor SummarySourceSystem object.

}
\item{\code{$fromList(SummarySourceSystemList)}}{

  Create SummarySourceSystem object from list.

}

\item{\code{$toList()}}{

  Get list representation of SummarySourceSystem.

}
\item{\code{fromJSONString(SummarySourceSystemJson)}}{

  Create SummarySourceSystem object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of SummarySourceSystem.

}
}
}

\keyword{datasets}
