% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarySourceSystem.r
\docType{data}
\name{SummarySourceSystem}
\alias{SummarySourceSystem}
\title{SummarySourceSystem Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SummarySourceSystem
}
\description{
SummarySourceSystem Class
}
\section{Fields}{

\describe{
\item{\code{code}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor SummarySourceSystem object.

}
\item{\code{$fromList(SummarySourceSystemList)}}{

  Create SummarySourceSystem object from list.

}

\item{\code{$toList()}}{

  Get list representation of SummarySourceSystem.

}
\item{\code{fromJSONString(SummarySourceSystemJson)}}{

  Create SummarySourceSystem object from list.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of SummarySourceSystem.

}
}
}

\keyword{datasets}
