% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryVernacularName.r
\docType{data}
\name{SummaryVernacularName}
\alias{SummaryVernacularName}
\title{SummaryVernacularName Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SummaryVernacularName
}
\description{
SummaryVernacularName Class
}
\section{Fields}{

\describe{
\item{\code{name}}{character}

\item{\code{language}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor SummaryVernacularName object.

}
\item{\code{$fromList(SummaryVernacularNameList)}}{

  Create SummaryVernacularName object from list.

}

\item{\code{$toList()}}{

  Get list representation of SummaryVernacularName.

}
\item{\code{fromJSONString(SummaryVernacularNameJson)}}{

  Create SummaryVernacularName object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of SummaryVernacularName.

}
}
}

\keyword{datasets}
