% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryVernacularName.r
\docType{data}
\name{SummaryVernacularName}
\alias{SummaryVernacularName}
\title{SummaryVernacularName Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SummaryVernacularName
}
\description{
SummaryVernacularName Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{language}}{}
}}

\keyword{datasets}
