% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxon.r
\docType{data}
\name{Taxon}
\alias{Taxon}
\title{Taxon Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Taxon
}
\description{
Taxon Class
}
\section{Fields}{

\describe{
\item{\code{sourceSystem}}{}

\item{\code{sourceSystemId}}{}

\item{\code{recordURI}}{}

\item{\code{id}}{}

\item{\code{sourceSystemParentId}}{}

\item{\code{taxonRank}}{}

\item{\code{taxonRemarks}}{}

\item{\code{occurrenceStatusVerbatim}}{}

\item{\code{acceptedName}}{}

\item{\code{defaultClassification}}{}

\item{\code{systemClassification}}{}

\item{\code{synonyms}}{}

\item{\code{vernacularNames}}{}

\item{\code{descriptions}}{}

\item{\code{references}}{}

\item{\code{experts}}{}

\item{\code{validName}}{}
}}

\keyword{datasets}
