% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxon.r
\docType{data}
\name{Taxon}
\alias{Taxon}
\title{Taxon Class}
\format{}
\usage{

}
\description{
A taxon record stores the hierarchical classification of a taxon, its scientific names and synonyms, and other relevant data retrieved from the respective source system. Taxonomic information is sourced from two species checklists: The Catalogue of Life (COL) and the Dutch Species Register (NSR). For more info on model classes, please refer to https://docs.biodiversitydata.nl
}
\section{Fields}{

\describe{
\item{\code{sourceSystem}}{SourceSystem}

\item{\code{sourceSystemId}}{character}

\item{\code{recordURI}}{character}

\item{\code{id}}{character}

\item{\code{sourceSystemParentId}}{character}

\item{\code{taxonRank}}{character}

\item{\code{taxonRemarks}}{character}

\item{\code{occurrenceStatusVerbatim}}{character}

\item{\code{acceptedName}}{ScientificName}

\item{\code{defaultClassification}}{DefaultClassification}

\item{\code{systemClassification}}{list(Monomial)}

\item{\code{synonyms}}{list(ScientificName)}

\item{\code{vernacularNames}}{list(VernacularName)}

\item{\code{descriptions}}{list(TaxonDescription)}

\item{\code{references}}{list(Reference)}

\item{\code{experts}}{list(Expert)}

\item{\code{validName}}{ScientificName}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Taxon object.

}
\item{\code{$fromList(TaxonList)}}{

  Create Taxon object from list.

}

\item{\code{$toList()}}{

  Get list representation of Taxon.

}
\item{\code{fromJSONString(TaxonJson)}}{

  Create Taxon object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Taxon.

}
}
}

\keyword{datasets}
