% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxon.r
\docType{data}
\name{Taxon}
\alias{Taxon}
\title{Taxon Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Taxon
}
\description{
Taxon Class
}
\section{Fields}{

\describe{
\item{\code{validName}}{}
}}

\keyword{datasets}
