% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxon.r
\docType{data}
\name{Taxon}
\alias{Taxon}
\title{Taxon Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Taxon
}
\description{
Taxon Class
}
\section{Fields}{

\describe{
\item{\code{sourceSystem}}{SourceSystem}

\item{\code{sourceSystemId}}{character}

\item{\code{recordURI}}{character}

\item{\code{id}}{character}

\item{\code{sourceSystemParentId}}{character}

\item{\code{taxonRank}}{character}

\item{\code{taxonRemarks}}{character}

\item{\code{occurrenceStatusVerbatim}}{character}

\item{\code{acceptedName}}{ScientificName}

\item{\code{defaultClassification}}{DefaultClassification}

\item{\code{systemClassification}}{list(Monomial)}

\item{\code{synonyms}}{list(ScientificName)}

\item{\code{vernacularNames}}{list(VernacularName)}

\item{\code{descriptions}}{list(TaxonDescription)}

\item{\code{references}}{list(Reference)}

\item{\code{experts}}{list(Expert)}

\item{\code{validName}}{ScientificName}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor Taxon object.

}
\item{\code{$fromList(TaxonList)}}{

  Create Taxon object from list.

}

\item{\code{$toList()}}{

  Get list representation of Taxon.

}
\item{\code{fromJSONString(TaxonJson)}}{

  Create Taxon object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of Taxon.

}
}
}

\keyword{datasets}
