% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonClient.r
\docType{data}
\name{TaxonClient}
\alias{TaxonClient}
\title{Taxon operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TaxonClient
}
\description{
Client class for nbaR.Taxon objects
}
\section{Fields}{

\describe{
\item{\code{basePath}}{Stores url path of the request.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\keyword{datasets}
