% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonClient.r
\docType{data}
\name{TaxonClient}
\alias{TaxonClient}
\title{Taxon operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TaxonClient
}
\description{
nbaR.Taxon
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

count_http_get1 Get the number of taxa matching a condition


count_http_post_json3 Get the number of taxa matching a condition


dwca_get_data_set Download dataset as Darwin Core Archive File


dwca_get_data_set_names Retrieve the names of all available datasets


dwca_query_http_get Dynamic download service: Query for taxa and return result as Darwin Core Archive File


dwca_query_http_post_json Dynamic download service: Query for taxa and return result as Darwin Core Archive File


find3 Find a taxon by id


find_by_ids3 Find taxa by ids


get_distinct_values Get all different values that can be found for one field


get_distinct_values_http_post_json3 Get all different values that exist for a field


get_field_info Returns extended information for each field of a specimen document


get_paths Returns the full path of all fields within a document


get_settings6 List all publicly available configuration settings for the NBA


get_settings7 Get the value of an NBA setting


group_by_scientific_name_http_get Aggregates Taxon and Specimen documents according to their scientific names


group_by_scientific_name_http_post_json Aggregates Taxon and Specimen documents according to their scientific names


is_operator_allowed3 Checks if a given operator is allowed for a given field


query Query for taxa


query_http_post_json3 Query for taxa

}
}

\keyword{datasets}
