% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonDescription.r
\docType{class}
\name{TaxonDescription}
\alias{TaxonDescription}
\title{TaxonDescription Class}
\format{R6 class}
\usage{
# TaxonDescription$new()
}
\description{
For more information on the NBA object model, please refer to the
official NBA documentation at
\href{https://docs.biodiversitydata.nl}{https://docs.biodiversitydata.nl} and
the NBA model and endpoints reference at
\href{https://docs.biodiversitydata.nl/endpoints-reference}{https://docs.biodiversitydata.nl/endpoints-reference}.
}
\details{
Model class for TaxonDescription objects.
}
\section{Fields}{

\describe{
\item{\code{category}}{character}

\item{\code{description}}{character}

\item{\code{language}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor TaxonDescription object.

}
\item{\code{$fromList(TaxonDescriptionList)}}{

  Create TaxonDescription object from list.

}

\item{\code{$toList()}}{

  Get list representation of TaxonDescription.

}
\item{\code{fromJSONString(TaxonDescriptionJson)}}{

  Create TaxonDescription object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of TaxonDescription.

}
}
}

\keyword{datasets}
