% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonDescription.r
\docType{data}
\name{TaxonDescription}
\alias{TaxonDescription}
\title{TaxonDescription Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TaxonDescription
}
\description{
TaxonDescription Class
}
\section{Fields}{

\describe{
\item{\code{category}}{}

\item{\code{description}}{}

\item{\code{language}}{}
}}

\keyword{datasets}
