% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonDescription.r
\docType{data}
\name{TaxonDescription}
\alias{TaxonDescription}
\title{TaxonDescription Class}
\format{}
\usage{

}
\description{
TaxonDescription Class
}
\section{Fields}{

\describe{
\item{\code{category}}{character}

\item{\code{description}}{character}

\item{\code{language}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor TaxonDescription object.

}
\item{\code{$fromList(TaxonDescriptionList)}}{

  Create TaxonDescription object from list.

}

\item{\code{$toList()}}{

  Get list representation of TaxonDescription.

}
\item{\code{fromJSONString(TaxonDescriptionJson)}}{

  Create TaxonDescription object from JSON.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of TaxonDescription.

}
}
}

\keyword{datasets}
