% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonomicEnrichment.r
\docType{data}
\name{TaxonomicEnrichment}
\alias{TaxonomicEnrichment}
\title{TaxonomicEnrichment Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TaxonomicEnrichment
}
\description{
TaxonomicEnrichment Class
}
\section{Fields}{

\describe{
\item{\code{taxonId}}{}
}}

\keyword{datasets}
