% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonomicEnrichment.r
\docType{data}
\name{TaxonomicEnrichment}
\alias{TaxonomicEnrichment}
\title{TaxonomicEnrichment Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TaxonomicEnrichment
}
\description{
TaxonomicEnrichment Class
}
\section{Fields}{

\describe{
\item{\code{vernacularNames}}{list(SummaryVernacularName)}

\item{\code{synonyms}}{list(SummaryScientificName)}

\item{\code{sourceSystem}}{SummarySourceSystem}

\item{\code{taxonId}}{character}
}}

\section{Methods}{

\describe{

\item{\code{$new()}}{

  Constructor TaxonomicEnrichment object.

}
\item{\code{$fromList(TaxonomicEnrichmentList)}}{

  Create TaxonomicEnrichment object from list.

}

\item{\code{$toList()}}{

  Get list representation of TaxonomicEnrichment.

}
\item{\code{fromJSONString(TaxonomicEnrichmentJson)}}{

  Create TaxonomicEnrichment object from list.

}
\item{\code{toJSONString(pretty=TRUE)}}{

  Get JSON representation of TaxonomicEnrichment.

}
}
}

\keyword{datasets}
