% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{thr_donut_plot}
\alias{thr_donut_plot}
\title{Threat status donut plot}
\usage{
<<<<<<< HEAD
thr_donut_plot(DF, FILL)
=======
thr_donut_plot(DF, YMAX, YMIN, FILL, COLOUR)
>>>>>>> fb05fd85839fa7d3171432d9ed95955c368eae33
}
\arguments{
\item{DF}{The data frame that contains the information on threat status}

<<<<<<< HEAD
\item{FILL}{The groups}
=======
\item{YMAX}{The threat status percentages}

\item{YMIN}{The groups}

\item{FILL}{The threat status categories}

\item{COLOUR}{colour for the threat status}
>>>>>>> fb05fd85839fa7d3171432d9ed95955c368eae33
}
\value{
Returns a bar donut plot of threat status
}
\description{
<<<<<<< HEAD
Donut plots for the threat status.
This function requires a dataframe that contains a columns of the
the ecosystem function groups or taxa, ecosystem types or species, and their threat status.
}
\details{
The function will group the ecosystem types or species by the functional
group or taxa and threat status, and calculate the number of ecosystem types or species within each functional
group or taxa with that threat status.
It will also create the colours and axis titles.
}
\examples{
#test <- thr_donut_plot(mydata, threat_status)
=======
Donut plots for the threat status of the ecosystem types
}
\examples{
#test <- thr_donut_plot(mydata, ecosystem_functional_grps, percentages, threat_status)
>>>>>>> fb05fd85839fa7d3171432d9ed95955c368eae33

}
