% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{thr_donut_plot}
\alias{thr_donut_plot}
\title{Threat status donut plot}
\usage{
thr_donut_plot(DF, YMAX, YMIN, FILL, COLOUR, COUNTS)
}
\arguments{
\item{DF}{The data frame that contains the information on threat status}

\item{YMAX}{The threat status percentages}

\item{YMIN}{The groups}

\item{FILL}{The threat status categories}

\item{COLOUR}{colour for the threat status}

\item{COUNTS}{The frequency counts of the number of ecosystems within each threat status}
}
\value{
Returns a bar donut plot of threat status
}
\description{
Donut plots for the threat status of the ecosystem types
}
\examples{
#test <- thr_donut_plot(mydata, ecosystem_functional_grps, percentages, threat_status)

}
