% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{thr_efg}
\alias{thr_efg}
\title{Threat status bar plot}
\usage{
thr_efg(DAT, X, Y, FILL, LABEL)
}
\arguments{
\item{X}{The groups}

\item{Y}{The threat status percentages}

\item{FILL}{The threat status categories}

\item{LABEL}{The frequency counts of the number of ecosystems within each protection level}

\item{DF}{The data frame that contains the information on threat status}
}
\value{
Returns a bar graph of threat status
}
\description{
Horizontal barplots for the threat status of ecosystem function groups (efgs)
}
\examples{
#test <- thr_efg(mydata, ecosystem_functional_grps, percentages, threat_status)
}
