% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{thr_tbl}
\alias{thr_tbl}
\title{Protection level table}
\usage{
thr_tbl(DF, COL)
}
\arguments{
\item{DF}{The data frame that contains the data}

\item{COL}{the column containing the Protection level categories}
}
\description{
A basic table with a purple heading bar and coloured blocks around the
Protection level categories
This data should be in the same fomat as for the plot functions
}
\examples{
#test <- thr_tbl(bird_data, Protection_level)

}
