% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VernacularName.r
\docType{data}
\name{VernacularName}
\alias{VernacularName}
\title{VernacularName Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VernacularName
}
\description{
VernacularName Class
}
\section{Fields}{

\describe{
\item{\code{experts}}{}
}}

\keyword{datasets}
