% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNKCJE.R
\name{writePG}
\alias{writePG}
\title{Download data from NBA.com/Stats/ and build your own database in local PostgreSQL instance}
\usage{
writePG(data, dbname, schemat = "rd")
}
\description{
Download data from NBA.com/Stats/ and build your own database in local PostgreSQL instance
}

